### 2021-04-08 #3158 @danielskatz
👋 @Garyfallidis - thanks for your submission.  The editors will now discuss if it meets the [research software](https://joss.readthedocs.io/en/latest/submitting.html#what-we-mean-by-research-software) criterion for review by JOSS.  You should hear back in a week or so.

In addition, you could work on the possibly missing DOIs that whedon suggests, but note that some may be incorrect. Please feel free to make changes to your .bib file, then use the command `@whedon check references` to check again, and the command `@whedon generate pdf` when the references are right to make a new PDF. Whedon commands need to the be the first entry in a new comment.
