+++ paper.bib
@@ -41,7 +41,8 @@
   volume={445},
   pages={51--56},
   year={2010},
-  organization={Austin, TX}
+  organization={Austin, TX},
+  doi={10.25080/majora-92bf1922-00a}
 }
 
 @article{pedregosa2011scikit,
@@ -83,7 +84,8 @@
   number={03},
   pages={90--95},
   year={2007},
-  publisher={IEEE Computer Society}
+  publisher={IEEE Computer Society},
+  doi={10.1109/mcse.2007.55}
 }
 
 @misc{BulletRe79:online,
@@ -108,7 +110,8 @@ note = {(Accessed on 02/23/2021)}
   title={VisPy: harnessing the GPU for fast, high-level visualization},
   author={Campagnola, Luke and Klein, Almar and Larson, Eric and Rossant, Cyrille and Rougier, Nicolas P},
   booktitle={Proceedings of the 14th Python in Science Conference},
-  year={2015}
+  year={2015},
+  doi={10.25080/majora-7b98e3ed-00e}
 }
 
 @inproceedings{rougier2015glumpy,
@@ -126,7 +129,8 @@ note = {(Accessed on 02/23/2021)}
   number={2},
   pages={40--51},
   year={2011},
-  publisher={IEEE}
+  publisher={IEEE},
+  doi={10.1109/mcse.2011.35}
 }
 
 @article{garyfallidis2014dipy,
@@ -136,7 +140,8 @@ note = {(Accessed on 02/23/2021)}
   volume={8},
   pages={8},
   year={2014},
-  publisher={Frontiers}
+  publisher={Frontiers},
+  doi={10.3389/fninf.2014.00008}
 }
 
 
@@ -218,7 +223,8 @@ note = {(Accessed on 02/23/2021)}
   booktitle={2011 IEEE Nanotechnology Materials and Devices Conference},
   pages={75--79},
   year={2011},
-  organization={IEEE}
+  organization={IEEE},
+  doi={10.1109/nmdc.2011.6155316}
 }
 
 @article{klimeck2008nanohub,
@@ -247,4 +253,4 @@ title={Cuda fluid simulation in nvidia physx},
 author={Harris, Mark},
 journal={Siggraph Asia},
 year={2009}
-} 
+}
