+++ paper.bib
@@ -214,7 +214,8 @@ note = {(Accessed on 02/23/2021)}
   number={1},
   pages={107--117},
   year={2013},
-  publisher={De Gruyter}
+  publisher={De Gruyter},
+  doi={10.1515/ntrev-2012-0043}
 }
 
 @inproceedings{klimeck2011network,
