### 2021-04-09 955b4d9
Update viz_solar_system.py
Reduced some repetitive code. and some more changes like list name changed.
