+++ docs/tutorials/01_introductory/viz_solar_system.py
@@ -22,14 +22,20 @@ scene = window.Scene()
 ##############################################################################
 # Load in a texture for each of the actors. To do this, we will create
 # a function called ``init_planet``, which will initialize each planet actor
-# given its corresponding filename and actor name. It will also add each
-# actor to the scene that has already been created.
+# given its corresponding filename and actor name.It will also set the
+# position of the planet and rescale it. It will also add each actor to
+# the scene that has already been created.
 
-planet_filenames = ["8k_mercury.jpg", "8k_venus_surface.jpg",
-                    "1_earth_8k.jpg", "8k_mars.jpg", "jupiter.jpg",
-                    "8k_saturn.jpg", "8k_saturn_ring_alpha.png",
-                    "2k_uranus.jpg", "2k_neptune.jpg",
-                    "8k_sun.jpg"]
+planet_filenames = {"8k_mercury.jpg" : [7, (0.4, 0.4, 0.4)],
+                    "8k_venus_surface.jpg" : [9, (0.6, 0.6, 0.6)],
+                    "1_earth_8k.jpg": [11, (0.4, 0.4, 0.4)],
+                    "8k_mars.jpg" : [13, (0.8, 0.8, 0.8)],
+                    "jupiter.jpg" : [16, (2, 2, 2)],
+                    "8k_saturn.jpg" : [19, (2, 2, 2)],
+                    "8k_saturn_ring_alpha.png" : [19, (3, 0.5, 3)],
+                    "2k_uranus.jpg" : [22, (1, 1, 1)],
+                    "2k_neptune.jpg" : [25, (1, 1, 1)],
+                    "8k_sun.jpg" : [0, (5, 5, 5)]}
 fetch_viz_textures()
 
 
@@ -38,8 +44,10 @@ def init_planet(filename):
 
     Parameters
     ----------
-    filename : str
-        The filename for the corresponding planet texture.
+    filename : dict
+        The filename is a dictionary, the key is the name of texture file
+        name and the value is the list of the radius of the orbit of planet
+        and the scale of the planet.
 
     Returns
     -------
@@ -49,6 +57,8 @@ def init_planet(filename):
     planet_file = read_viz_textures(filename)
     planet_image = io.load_image(planet_file)
     planet_actor = actor.texture_on_sphere(planet_image)
+    planet_actor.SetPosition(planet_filenames[filename][0], 0, 0)
+    planet_actor.SetScale(planet_filenames[filename][1])
     scene.add(planet_actor)
     return planet_actor
 
@@ -75,51 +85,17 @@ sun_actor = planet_actor_list[9]
 utils.rotate(jupiter_actor, (90, 1, 0, 0))
 
 ##############################################################################
-# Next, change the positions and scales of the planets according to their
-# position and size within the solar system (relatively). For the purpose
-# of this tutorial, planet sizes and positions will not be completely
-# accurate.
-
-r_mercury = 7
-r_venus = 9
-r_earth = 11
-r_mars = 13
-r_jupiter = 16
-r_saturn = 19
-r_uranus = 22
-r_neptune = 25
-
-sun_actor.SetScale(5, 5, 5)
-mercury_actor.SetScale(0.4, 0.4, 0.4)
-venus_actor.SetScale(0.6, 0.6, 0.6)
-earth_actor.SetScale(0.4, 0.4, 0.4)
-mars_actor.SetScale(0.8, 0.8, 0.8)
-jupiter_actor.SetScale(2, 2, 2)
-saturn_actor.SetScale(2, 2, 2)
-saturn_rings_actor.SetScale(3, 0.5, 3)
[…]
