### 2021-04-13 4478f30
creating tests for resizing background bug
