### 2021-04-13 88e3f3e
Moved SelectorManager class to pick.py
