+++ fury/tests/test_pick.py
@@ -70,88 +70,6 @@ def test_picking_manager():
     assert_greater(np.sum(np.abs(np.diff(np.array(record_indices['xyz']),
                                          axis=0))), 0)
 
-
-
-
-def test_selector_manager_tmp():
-    import vtk 
-    hsel = vtk.vtkHardwareSelector()
-    hsel.SetFieldAssociation(vtk.vtkDataObject.FIELD_ASSOCIATION_CELLS)
-    hsel.SetRenderer(scene)
-
-    event_pos = showm.iren.GetEventPosition()
-    pass
-
-import vtk
-from fury.utils import numpy_support as nps
-
-class SelectorManager(object):
-
-    def __init__(self, select='faces'):
-        self.hsel = vtk.vtkHardwareSelector()
-        self.selection_type(select)
-        # self.hsel.SetActorPassOnly(True)
-
-    def selection_type(self, select):        
-        if select == 'faces' or select == 'edges':
-            self.hsel.SetFieldAssociation(vtk.vtkDataObject.FIELD_ASSOCIATION_CELLS)
-        if select == 'points' or select == 'vertices':
-            self.hsel.SetFieldAssociation(vtk.vtkDataObject.FIELD_ASSOCIATION_POINTS)
-       
-    def pick(self, disp_xy, sc):
-        self.select(disp_xy, sc, area=1)
-
-    def select(self, disp_xy, sc, area=1):
-        self.hsel.SetRenderer(sc)
-        picking_area = area
-        self.hsel.SetArea(disp_xy[0] - picking_area, disp_xy[1] - picking_area,
-                          disp_xy[0] + picking_area, disp_xy[1] + picking_area)
-        res = self.hsel.Select()
-        # print(res)
-        num_nodes = res.GetNumberOfNodes()
-        if (num_nodes < 1):
-            sel_node = None
-        else:
-            print('Number of Nodes ', num_nodes)
-            for i in range(num_nodes):
-                print('Node ', i)
-                sel_node = res.GetNode(i)
-                
-                if(sel_node is not None):
-                    selected_nodes = set(np.floor(nps.vtk_to_numpy(
-                        sel_node.GetSelectionList())).astype(int))
-                    
-                    print('#>>>>', id(sel_node.GetProperties().Get(sel_node.PROP())))
-
-                    # selected_node = list(selected_nodes)[0]
-                    print('Selected Nodes ', selected_nodes)
-                    # print('Prop ', sel_node.GetProperties())
-                    # print('Prop ID', sel_node.PROP_ID())
-                    # print('Prop ', sel_node.PROP())
-        
-        # selected_actor.text.SetText(str(selected_node))
-        # if(selected_node is not None):
-        #     if(labels is not None):
-        #         selected_actor.text.SetText(labels[selected_node])
-        #     else:
-        #         selected_actor.text.SetText("#%d" % selected_node)
-        #     selected_actor.SetPosition(positions[selected_node])
-
-        # else:
-        #     selected_actor.text.SetText("")
-
-    def event_position(self, iren):
-        """ Returns event display position from interactor
-
-        Parameters
-        ----------
-        iren : interactor
-            The interactor object can be retrieved for example
-            using providing ShowManager's iren attribute.
-        """
-        return iren.GetEventPosition()
-
-
 def test_selector_manager():
 
     xyz = 10 * np.random.rand(100, 3)
@@ -193,7 +111,7 @@ def test_selector_manager():
     # use itertools to avoid global variables
     counter = itertools.count()
 
-    selm = SelectorManager(select='points')
+    selm = pick.SelectorManager(select='points')
 
     record_indices = {'vertex_indices': [],
[…]
