### 2021-04-13 8e11f9d
Updated SelectionManager and select returns info
