### 2021-04-14 7e11233
Update viz_solar_system.py
Some more changes according to review.
