+++ docs/tutorials/01_introductory/viz_solar_system.py
@@ -20,45 +20,57 @@ from fury.data import read_viz_textures, fetch_viz_textures
 scene = window.Scene()
 
 ##############################################################################
-# Load in a texture for each of the actors. To do this, we will create
-# a function called ``init_planet``, which will initialize each planet actor
-# given its corresponding filename and actor name.It will also set the
-# position of the planet and rescale it. It will also add each actor to
-# the scene that has already been created.
+# Define information relevant for each planet actor including its
+# texture name, relative position, and scale.
 
-planet_filenames = {"8k_mercury.jpg": [7, (0.4, 0.4, 0.4)],
-                    "8k_venus_surface.jpg": [9, (0.6, 0.6, 0.6)],
-                    "1_earth_8k.jpg": [11, (0.4, 0.4, 0.4)],
-                    "8k_mars.jpg": [13, (0.8, 0.8, 0.8)],
-                    "jupiter.jpg": [16, (2, 2, 2)],
-                    "8k_saturn.jpg": [19, (2, 2, 2)],
-                    "8k_saturn_ring_alpha.png": [19, (3, 0.5, 3)],
-                    "2k_uranus.jpg": [22, (1, 1, 1)],
-                    "2k_neptune.jpg": [25, (1, 1, 1)],
-                    "8k_sun.jpg": [0, (5, 5, 5)]}
+planets_data = [{'filename': '8k_mercury.jpg', 'position': 7,
+                 'scale': (.4, .4, .4)},
+                {'filename': '8k_venus_surface.jpg', 'position': 9,
+                 'scale': (.6, .6, .6)},
+                {'filename': '1_earth_8k.jpg', 'position': 11,
+                 'scale': (.4, .4, .4)},
+                {'filename': '8k_mars.jpg', 'position': 13,
+                 'scale': (.8, .8, .8)},
+                {'filename': 'jupiter.jpg', 'position': 16,
+                 'scale': (2, 2, 2)},
+                {'filename': '8k_saturn.jpg', 'position': 19,
+                 'scale': (2, 2, 2)},
+                {'filename': '8k_saturn_ring_alpha.png', 'position': 19,
+                 'scale': (3, .5, 3)},
+                {'filename': '2k_uranus.jpg', 'position': 22,
+                 'scale': (1, 1, 1)},
+                {'filename': '2k_neptune.jpg', 'position': 25,
+                 'scale': (1, 1, 1)},
+                {'filename': '8k_sun.jpg', 'position': 0,
+                 'scale': (5, 5, 5)}]
 fetch_viz_textures()
 
+##############################################################################
+# To take advantage of the previously defined data structure we are going to
+# create an auxiliary function that will load and apply the respective
+# texture, set its respective properties (relative position and scale),
+# and add the actor to a previously created scene.
 
-def init_planet(filename):
+
+def init_planet(planet_data):
     """Initialize a planet actor.
 
     Parameters
     ----------
-    filename : dict
-        The filename is a dictionary, the key is the name of texture file
-        name and the value is the list of the radius of the orbit of planet
-        and the scale of the planet.
+    planet_data : dict
+        The planet_data is a dictionary, and the keys are filename(texture),
+        position and scale.
 
     Returns
     -------
     planet_actor: actor
         The corresponding sphere actor with texture applied.
     """
-    planet_file = read_viz_textures(filename)
+    planet_file = read_viz_textures(planet_data['filename'])
     planet_image = io.load_image(planet_file)
     planet_actor = actor.texture_on_sphere(planet_image)
-    planet_actor.SetPosition(planet_filenames[filename][0], 0, 0)
-    planet_actor.SetScale(planet_filenames[filename][1])
+    planet_actor.SetPosition(planet_data['position'], 0, 0)
+    planet_actor.SetScale(planet_data['scale'])
[…]
