### 2021-04-14 #3158 @kyleniemeyer
Hi @Garyfallidis, the JOSS editorial board reviewed your submission and found that it is within our scope. However, due to a lack of suitable and available editors, I'm going to put this on our waitlist for now. We will assign this to an editor as soon as one becomes available.
