### 2021-04-15 2fb518f
Update viz_solar_system.py
Final touch to tutorial.
