### 2021-04-16 a8aa66b
updated test for parametric_surface actor
