### 2021-04-16 013f010
added value error message for parametric_surface
