### 2021-04-16 ff5ffdc
reduce code duplication
