+++ fury/primitive.py
@@ -854,14 +854,26 @@ def prim_cylinder(radius=0.5, height=1, sectors=36, capped=True):
     return vertices, triangles
 
 
-def init_parametric(u_lower_bound, u_upper_bound, v_lower_bound,
-                    v_upper_bound, npoints):
+def build_parametric(u_lower_bound, u_upper_bound, v_lower_bound,
+                     v_upper_bound, npoints, surface_equation):
     u = np.linspace(u_lower_bound, u_upper_bound, npoints)
     v = np.linspace(v_lower_bound, v_upper_bound, npoints)
     u, v = np.meshgrid(u, v)
     u = u.reshape(-1)
     v = v.reshape(-1)
-    return u, v, np.sin, np.cos, np.mean
+    points2D = np.vstack([u, v]).T
+    tri = Delaunay(points2D)
+    triangles = tri.simplices
+
+    x, y, z = surface_equation(u, v)
+    # Centering the surface
+    x -= np.mean(x)
+    y -= np.mean(y)
+    z -= np.mean(z)
+    xyz = np.vstack([x, y, z]).T
+    vertices = np.ascontiguousarray(xyz)
+
+    return vertices, triangles
 
 
 def prim_para_mobius_strip(npoints=100):
@@ -875,21 +887,15 @@ def prim_para_mobius_strip(npoints=100):
         triangles that compose our Möbius strip
     """
 
-    u, v, sin, cos, mean = init_parametric(0, 2*np.pi, -1, 1, npoints)
-    points2D = np.vstack([u, v]).T
-    tri = Delaunay(points2D)
-    triangles = tri.simplices
+    def mobius_equation(u, v):
+        sin = np.sin
+        cos = np.cos
+        x = (1 + v/2 * cos(u/2)) * cos(u)
+        y = (1 + v/2 * cos(u/2)) * sin(u)
+        z = v/2 * sin(u/2)
+        return x, y, z
 
-    x = (1 + v/2 * cos(u/2)) * cos(u)
-    y = (1 + v/2 * cos(u/2)) * sin(u)
-    z = v/2 * sin(u/2)
-    # Centering the surface
-    x -= mean(x)
-    y -= mean(y)
-    z -= mean(z)
-    xyz = np.vstack([x, y, z]).T
-    vertices = np.ascontiguousarray(xyz)
-    return vertices, triangles
+    return build_parametric(0, 2*np.pi, -1, 1, npoints, mobius_equation)
 
 
 def prim_para_kleins_bottle(npoints=100):
@@ -904,25 +910,19 @@ def prim_para_kleins_bottle(npoints=100):
 
     """
 
-    u, v, sin, cos, mean = init_parametric(0, np.pi, 0, 2*np.pi, npoints)
-    points2D = np.vstack([u, v]).T
-    tri = Delaunay(points2D)
-    triangles = tri.simplices
+    def klein_equation(u, v):
+        sin = np.sin
+        cos = np.cos
+        x = -2/15*cos(u)*(3*cos(v) - 30*sin(u) + 90*cos(u)**4*sin(u) -
+                            60*cos(u)**6*sin(u) + 5*cos(u)*cos(v)*sin(u))
+        y = -1/15*sin(u)*(3*cos(v) - 3*cos(u)**2*cos(v) - 48*cos(u)**4*cos(v) +
+                            48*cos(u)**6*cos(v) - 60*sin(u) + 5*cos(u)*cos(v)*sin(u)
+                            - 5*cos(u)**3*cos(v)*sin(u) - 80*cos(u)**5*cos(v)*sin(u)
[…]
