+++ fury/tests/test_ui.py
@@ -298,7 +298,7 @@ def test_ui_textbox(recording=False):
     npt.assert_equal(another_textbox_test.size, (83, 17))
 
     textbox_bg_test = ui.TextBox2D(text="FURY\nfury", width=10, height=10,
-                                    font_size=50)
+                                        font_size=50)
     show_manager.scene.add(textbox_bg_test)
     npt.assert_equal(textbox_bg_test.size, (116, 99))
 
