### 2021-04-19 1117c22
feat: added test for _get_size
