### 2021-04-23 407f71d
added an example about graph-tool  and nested stochastic block model
