+++ docs/examples/viz_nested_block_model.py
@@ -1,0 +1,229 @@
+ 
+"""
+============================================================
+Nested Stochastic Block inference and Graph-tool integration
+============================================================
+
+Graph-tool it's a superb lib to perform a myriad of network inference methods. 
+However, graph-tool uses cairo to render the plots.
+Here, I'll show to you how to use fury  to draw a nested stochastic block model
+"""
+
+
+#####################################################################################
+# First, we need to load the network data and extract the largest component
+
+
+from graph_tool.all import *
+import graph_tool as gt
+import numpy as np
+
+
+g = gt.collection.data["celegansneural"]
+g = gt.GraphView(g, vfilt=gt.topology.label_largest_component(g))
+g.purge_vertices()
+
+
+#####################################################################################
+# Here is a simple function to extract the information of the inferred nested block
+# model of  any graph-tool instance
+
+def getNestedInference(g):
+    '''Given a graph-tool instance this function extract the results obtained from  
+    nested block model inference algorithm
+    
+    Parameters:
+    -----------
+        g: Graph or GraphView instance
+            A graph with |V| vertices and |E| edges
+    Returns:
+    --------
+        cts: list of arrays (float)
+            A lists containing |E| arrays of different sizes. Each array 
+            represents a set of control points for each edge
+        pos: array (float)
+            A 2d array of |V|x2 size. Each row represents a  node positions in a circular layour
+        labels: list (int)
+            a list of  |V| int numbers. Each number represents the comunitiy which the node belongs
+    '''
+    state = gt.inference.minimize_nested_blockmodel_dl(g)
+    
+    # cts represent the control points of the Beizer curves used for graph-tool to draw the edges
+    tg = gt.inference.nested_blockmodel.get_hierarchy_tree(state)[0]
+    tpos = pos = gt.draw.radial_tree_layout(tg, tg.vertex(tg.num_vertices() - 1), weighted=True)
+    cts = gt.draw.get_hierarchy_control_points(g, tg, tpos) 
+    
+    cts = list(cts)
+    pos = np.array(list(g.own_property(tpos)))
+    labels = list(state.levels[0].b)
+    
+    return cts, pos, labels
+
+# Now, we can peform the nested block model inference
+cts, pos, labels = getNestedInference(g)
+
+
+#####################################################################################
+#  Each diferent comunity obtained from the nested block model  algorithm will have a 
+# different color
+from fury import colormap as cmap
+
+comunityColors = cmap.distinguishable_colormap(nb_colors=len(np.unique(labels)))
+
+nodeColors = np.array([comunityColors[c]
+                   for c in labels])
+
+
+
+#####################################################################################
+# To plot the resuts of the inference we will need to draw curved lines in FURY. 
+# To do so, we  will use the BSpline interpolation with the  control's  points  obtained 
+# from the function  getNestedInference.
+
+from scipy import interpolate
+
+def bSplineInterpolation(ctsCoord, points):
+    '''
+    Parameters:
+    -----------
+    
+    Return:
+    -------
+    
+    '''
+    
+    tck, u = interpolate.splprep(ctsCoord.T)
+    
+    coord = interpolate.splev(points, tck)
+    coord = np.array(coord)
+    return coord
+
+#####################################################################################
+# Graph-tool stores the control point coordinates  relative to the nodes positions. 
[…]
