### 2021-04-24 dccce39
modifing implementation and adding tests
