+++ fury/ui.py
@@ -3796,7 +3796,7 @@ class ListBox2D(UI):
         self.scroll_bar.width = int(size[0] * 0.05)
         self.margin = int(size[0] * 0.1)
         self.panel.update_element(self.scroll_bar, (size - self.scroll_bar.size
-                                                        - self.margin))
+                                    - self.margin))
 
         # Slots resize
         self.font_size = int(size[0] * 0.2)
