+++ fury/actor.py
@@ -2543,3 +2543,71 @@ def sdf(centers, directions=(1, 0, 0), colors=(1, 0, 0), primitives='torus',
     shader_to_actor(box_actor, "fragment", impl_code=fs_impl_code,
                     block="light")
     return box_actor
+
+
+def marker_billboard(
+        centers, colors=(0, 1, 0), scales=1, 
+        edgeWidth=.1, edgeColor=(255, 255, 255)):
+    """Create a billboard actor.
+
+    Billboards are 2D elements incrusted in a 3D world. It offers you the
+    possibility to draw differents shapes/elements at the shader level.
+
+    Parameters
+    ----------
+    centers : ndarray, shape (N, 3)
+        Superquadrics positions
+    colors : ndarray (N,3) or (N, 4) or tuple (3,) or tuple (4,)
+        RGB or RGBA (for opacity) R, G, B and A should be at the range [0, 1]
+    scales : ndarray, shape (N) or (N,3) or float or int, optional
+        The height of the cone.
+    vs_dec : str or list of str, optional
+        vertex shaders code that contains all variable/function delarations
+    vs_impl : str or list of str, optional
+        vertex shaders code that contains all variable/function implementation
+    fs_dec : str or list of str, optional
+        Fragment shaders code that contains all variable/function delarations
+    fs_impl : str or list of str, optional
+        Fragment shaders code that contains all variable/function
+        implementation
+    gs_dec : str or list of str, optional
+        Geometry shaders code that contains all variable/function delarations
+    gs_impl : str or list of str, optional
+        Geometry shaders code that contains all variable/function
+        mplementation
+
+    Returns
+    -------
+    vtkActor
+
+    """
+
+    verts, faces = fp.prim_square()
+    res = fp.repeat_primitive(verts, faces, centers=centers, colors=colors,
+                              scales=scales)
+
+    big_verts, big_faces, big_colors, big_centers = res
+    sq_actor = get_actor_from_primitive(big_verts, big_faces, big_colors)
+    sq_actor.GetMapper().SetVBOShiftScaleMethod(False)
+    sq_actor.GetProperty().BackfaceCullingOff()
+
+    attribute_to_actor(sq_actor, big_centers, 'center')
+    attribute_to_actor(
+        sq_actor,
+        np.repeat([edgeColor], big_centers.shape[0], axis=0), 'edgeColor')
+    attribute_to_actor(
+        sq_actor, np.repeat(edgeWidth, big_centers.shape[0]), 'edgeWidth')
+
+    vs_dec_code = load("marker_billboard_dec.vert")
+    vs_impl_code = load("marker_billboard_impl.vert")
+
+    fs_dec_code = load('marker_billboard_dec.frag')
+    fs_impl_code = load('marker_billboard_impl.frag')
+
+    shader_to_actor(sq_actor, "vertex", impl_code=vs_impl_code,
+                    decl_code=vs_dec_code)
+    shader_to_actor(sq_actor, "fragment", decl_code=fs_dec_code)
+    shader_to_actor(sq_actor, "fragment", impl_code=fs_impl_code,
+                    block="light")
+
+    return sq_actor
