### 2021-04-25 de985aa
reduce mem usage through uniform definitions
