### 2021-04-26 2dd2f83
This allows fury to deal with Uniforms in shaders and change them dynamically
An uniform variable should be chosen when we have any kind
of information which act as constants, at least during the  draw call.

This commit implements the following behaviors:
1 - avoids creating an observable-vtk for each uniform variable
2 - allows changing the uniform variables dynamically

Here it's a simple example

```python

markerUniforms = [
        Uniform(name='edgeColor', type='3f', value=edgeColor),
        Uniform(name='edgeWidth', type='f', value=edgeWidth),
        Uniform(name='edgeOpacity', type=
[…]
