+++ fury/shaders/base.py
@@ -170,3 +170,54 @@ def attribute_to_actor(actor, arr, attr_name, deep=True):
     mapper = actor.GetMapper()
     mapper.MapDataArrayToVertexAttribute(
         attr_name, attr_name, vtk.vtkDataObject.FIELD_ASSOCIATION_POINTS, -1)
+
+
+class Uniform:
+    def __init__(self, name, type, value):
+        '''
+        Args:
+        -----
+            name: str
+            type: str
+            value: 
+        '''
+        self.name = name
+        self.value = value
+        self.type = type
+        self.vtk_func_uniform = ''
+        self.setType()
+
+    def setType(self, ):
+        validTypes = [
+            '1fv', '1iv', '2f', '2fv', '2i', '3f', '3fv',
+            '3uc', '4f', '4fv', '4uc', 'GroupUpdateTime', 'Matrix',
+            'Matrix3x3', 'Matrix4x4', 'Matrix4x4v', 'f', 'i']
+        if self.type not in validTypes:
+            raise ValueError(
+                f"""Uniform type {self.type} not valid. 
+                Choose one of this values: {validTypes}""")
+
+        self.vtk_func_uniform = f'SetUniform{self.type}'
+
+    def executeProgram(self, program):
+        '''
+        Args:
+        -----
+            program: vtkmodules.vtkRenderingOpenGL2.vtkShaderProgram
+        '''
+        program.__getattribute__(self.vtk_func_uniform)(
+                self.name, self.value)
+
+
+class Uniforms:
+    def __init__(self, uniforms):
+        self.uniforms = uniforms
+        for obj in self.uniforms:
+            setattr(self, obj.name, obj)
+
+    def __call__(self, _caller, _event, calldata=None,):
+        program = calldata
+        if program is None:
+            return
+        for uniform in self.uniforms:
+            uniform.executeProgram(program)
