### 2021-04-26 2f74f9c
Allows to pass the marker as a symbol
Before this commit

```python
nodes_actor = actor.marker_billboard(...,markers=[0, 1,0],..)
```
Now, it's also possible:
```python
nodes_actor = actor.marker_billboard(...,markers=['o', 's','^'],..)
```
