### 2021-04-26 42525c8
fix: this removes the example that was added by mistake
