### 2021-04-26 e63f407
add docstrings
