### 2021-04-26 60819ce
Allows the marker symbol to be chosen using a char
After this commit you can do the following

```python

nodes_actor = actor.marker_billboard(centers,
        marker='^',
        markers=None,
        )
```

Or if you want to use a differnt marker for each center

```python

nodes_actor = actor.marker_billboard(centers,
        markers=['^' , 'o', 's',...,'*'],
        )
```
