### 2021-04-26 0b34967
change from camel case to underscore
As sugested by @skoudoro. Add docstrings
