### 2021-04-26 36833ef
Add test functions for Uniform and Uniforms obj
As sugested by @skoudoro
