+++ fury/tests/test_shaders.py
@@ -1,4 +1,5 @@
 import os
+import pytest
 
 import numpy.testing as npt
 
@@ -16,3 +17,36 @@ def test_load():
     npt.assert_string_equal(fs.load(dummy_file_name), dummy_file_contents)
 
     os.remove(os.path.join(fs.SHADERS_DIR, dummy_file_name))
+
+
+def test_uniform_tools():
+    # print('\nTest: Uniform tools')
+    # print('\t...creating a Uniform obj')
+    uniform = fs.Uniform('uniform_name', 'f', 1.)
+    assert uniform.vtk_func_uniform == 'SetUniformf'
+    assert uniform.value == 1.
+
+    class Dummy_Program:
+        def SetUniformf(self, name, value):
+            """camel case because we need to simulate the
+            program used in vtk with the same attr name
+            """
+            assert name == 'uniform_name'
+            assert value == 1.
+            # print(f'\t...uniform set for {name} and value {value}')
+
+    dummy_program = Dummy_Program()
+    uniform.execute_program(dummy_program)
+
+    # print('\t...invalid uniform type should create an exception')
+    with pytest.raises(Exception):
+        fs.Uniform('uniform_name', 'invalid_type', 1.)
+
+    # print('\t...creating an Uniforms object')
+    uniforms = fs.Uniforms([uniform])
+    # print('\t...invalid uniform list  should create an exception')
+    with pytest.raises(Exception):
+        fs.Uniforms([1, '--'])
+
+    # print('\t... test __call__ method used in callback')
+    uniforms(None, None)
