+++ fury/actor.py
@@ -787,13 +787,73 @@ def scalar_bar(lookup_table=None, title=" "):
     return scalar_bar
 
 
-def cube_axis_actor(scene,bounds,label_X_color=[1,0,0], label_Y_color=[0,1,0], label_Z_color=[0,0,1],
-                    title_X="X-Axis",title_Y="Y-Axis",title_Z="Z-Axis",
-                    draw_grid_lines=True,minor_tick_visibility=False):
+def cube_axis_actor(scene, bounds, label_X_color=[1, 0, 0], label_Y_color=[0, 1, 0], label_Z_color=[0, 0, 1],
+                    title_X="X-Axis", title_Y="Y-Axis", title_Z="Z-Axis",
+                    draw_grid_lines=True, minor_tick_visibility=False):
+    """ Cube Axis Actor creates an actor with the coordinate's system axes,
+    This is intended for use where you would need a scale for reference, or a grid on reference planes.
+    Recommended with parallel (orthogonal) camera view.
+
+    Parameters
+    ----------
+    scene : vtkOpenGLRenderer Object
+        the Scene you wish to render this actor on.
+    bounds : array[6,] or tuple(6,)
+        the limits of your scales on axes,
+        Higher bound, then lower bound, for X,Y & Z.
+        e.g., [1, 0, 1, 0, 1, 0] or (1, 0, 1, 0, 1, 0)
+    label_X_color : array[3,] or tuple(3,), optional
+        This gives X-axis its color
+        Default is [1, 0, 0].
+    label_X_color : array[3,] or tuple(3,), optional
+        This gives Y-axis its color.
+        Default is [0, 1, 0].
+    label_Z_color : array[3,] or tuple(3,), optional
+        This gives Z-axis its color
+        Default is [0, 0, 1].
+    title_X : string, optional
+        Custom title for X-Axis.
+        Default is "X-Axis".
+    title_Y : string, optional
+        Custom title for Y-Axis.
+        Default is "Y-Axis".
+    title_Z : string, optional
+        Custom title for Z-Axis.
+        Default is "Z-Axis".
+    draw_grid_lines : boolean, optional
+        Draws gridlines on reference XY, YZ, ZX planes farthest from the camera.
+        Default is set to True.
+    minor_tick_visibility : boolean, optional
+        This turns on smaller sub division ticks on the axes.
+        Default is set to False
+
+    Returns
+    -------
+    cube_axis_actor : vtkCubeAxesActor
+
+    See Also
+    --------
+    :func:`fury.actor.axes`
+    :func:`fury.actor.point`
+
+    Examples
+    --------
+        >>> from fury import window, actor
+        >>> scene = window.Scene()
+        >>> bounds = (0, 5, 0, 5, 0, 5)
+        >>> data = np.random.randint(5, size=(25, 3))
+        >>> for point in data:
+        >>>     scene.AddActor(actor.point(data, color))
+        >>> cube_axes_actor = cube_axis_actor(scene=scene, bounds=bounds)
+        >>> scene.add(cube_axes_actor)
+        >>> # window.show(scene)
+
+    """
     cubeAxesActor = vtk.vtkCubeAxesActor()
     cubeAxesActor.SetBounds(bounds)
-    # cubeAxesActor.SetCamera(renderer.GetActiveCamera()) # this is the vtk 
+
     cubeAxesActor.SetCamera(scene.GetActiveCamera())
+
[…]
