### 2021-04-27 94551ec
Pep8fix and make sure fury tests ran sucessfully
