### 2021-04-28 28540d4
Get the Id of the vtkObserver and set the priority
The reason to do that it’s because after creating a shader callback we
can’t remove the observer (and the callback) without knowing the
unsigned int. In addition, it’s impossible to control the order of the
callbacks without the priority parameter.
