### 2021-04-29 497bf77
add extra tests related with shader callbacks
