### 2021-04-29 a222a5c
this implements all suggestions made by @skoudoro
