### 2021-04-29 97fe013
added dots_actor parameter
