### 2021-04-29 b74403a
added tests for dots and points
