### 2021-04-29 028aa47
most important gl_numbers
