### 2021-04-29 909c795
allows to define the gl state for opaque objects
In addtion, this has the benefit of speeding up
the rendering of the image when it's necessary
