### 2021-04-29 0963d32
this add more blending options
It is necessary to check each one of the blending equations.
Obs: Sources for the blending equations used in this commit

http://apoorvaj.io/alpha-compositing-opengl-blending-and-premultiplied-alpha/
https://github.com/mrdoob/three.js/blob/bb1188178cee68b6bb1b2e0afbe459b6dfff6b39/src/renderers/webgl/WebGLState.js#L592
