### 2021-04-29 0b88093
This removes some features about gl-context
The aim it's focus only in the gl-contexts that can be used
to improve the performance and/or help in the scatter plots / network
visualizations algorithms
