+++ docs/examples/viz_fine_tuning_gl_context.py
@@ -24,6 +24,10 @@ import itertools
 from functools import partial
 
 
+###############################################################################
+# We just proceeds as usual: creating the actors and initializing a scene in
+# fury
+
 centers = 1*np.array([
     [0, 0, 0],
     [-1, 0, 0],
@@ -56,29 +60,34 @@ showm = window.ShowManager(scene,
                            order_transparent=True)
 
 ###############################################################################
-# All actors need to be added in a scene
+# All actors must be added  in a scene
 
 scene.add(actorAdd)
 scene.add(actors)
 scene.add(actorNoDeph)
 
 ###############################################################################
-# Now we will enter in the topic of this example. 
-# First we need to create (or use one of the pre-built gl_function of fury) to 
-# change the OpenGL state of a given fury window instance (showm.window). 
-# The function bellow it's a simple example and can be used to change the GL_DEPTH_STATE
-# from the opengl context of vtk. VTK allway's change this state to True before the draw
-# call, therefore we need to set them using inside a shader callback,
+# Now we will enter in the topic of this example. First, we need to create
+# (or use one of the pre-built gl_function of fury) to
+# change the OpenGL state of a given fury window instance (showm.window).
+#
+# The function bellow it's a simple example and can be used to change the
+# GL_DEPTH_STATE from the opengl context of a vtk instance. VTK allway's change
+#  this state to True before the draw call, therefore we need to set them
+# inside of a shader callback, even if you have just one actor.
 
 
 def gl_disable_depth(window):
+    '''this functions it's allways accessible through
+    fury.window.gl_disable_depth
+    '''
     GL_DEPTH_TEST = 2929
     glState = window.GetState()
     glState.vtkglDisable(GL_DEPTH_TEST)
 
 ###############################################################################
-# Next, we write a standard callback function and uses that function 
-# in the add_shader_callback method linking the actors which will be rendered with no 
+# Next, we write a standard callback function and uses that function in the
+# add_shader_callback method linking the actors which will be rendered with no
 # depth_test.
 
 
@@ -97,8 +106,8 @@ id_observer_depth = add_shader_callback(
 
 
 ###############################################################################
-# Here we're using the pre-build fury window functions which add specific 
-# behaviors to the OpenGL context 
+# Here we're using the pre-build fury window functions which add specific
+# behaviors to the OpenGL context
 
 
 id_observer_normal = add_shader_callback(
