### 2021-04-29 e1f357a
gl_get_current state now returns a dict state
