### 2021-04-30 1552da0
this add edge opacity as a parameter
