### 2021-05-01 4461d47
Fix #429: Colormap fix
