### 2021-05-03 9faac1a
allows the composition of opengl effects
After this commit it'll be possible to do the following

```python
from fury.shaders import add_shader_callback, shader_apply_effects
...
id = shader_apply_effects(
    showm, actor,
    effects=[
        window.gl_resset_blend_func, window.gl_disable_blend,
        window.gl_disable_depth])

```

This behavior was implemented in order to:

1- Avoid creating multiple vtkObservers when we are trying
to compose more than one opengl states
2- Provide the user with a method that allows him to
achieve the desired opengl state(effect)
