### 2021-05-03 4b35f20
updates the example to be compatible with the previous commit
