### 2021-05-05 44284fe
fix pep8 removes camelcase
- add refs about the opengl constant numbers
