### 2021-05-11 e69364e
feature: Visualization of a video.
Visualization of a video on a rotating plane.
Fixes Part of #322
