### 2021-05-11 3503e0c
Still updating and reshaping tutorial
