+++ fury/actors/peak.py
@@ -1,0 +1,335 @@
+from fury.colormap import boys2rgb, orient2rgb
+from fury.shaders import attribute_to_actor, load, shader_to_actor
+from fury.utils import apply_affine, numpy_to_vtk_colors, numpy_to_vtk_points
+from vtk.util import numpy_support
+
+
+import numpy as np
+import vtk
+
+
+class PeakActor(vtk.vtkActor):
+    def __init__(self, directions, indices, values=None, affine=None,
+                 colors=None, linewidth=1):
+        if affine is not None:
+            w_pos = apply_affine(affine, np.asarray(indices).T)
+
+        valid_dirs = directions[indices]
+
+        points_array = []
+        centers_array = []
+        diffs_array = []
+        for idx, center in enumerate(zip(indices[0], indices[1], indices[2])):
+            if affine is None:
+                xyz = np.array(center)
+            else:
+                xyz = w_pos[idx, :]
+            valid_peaks = np.nonzero(
+                np.abs(valid_dirs[idx, :, :]).max(axis=-1) > 0.)[0]
+            for pdir in valid_peaks:
+                if values is not None:
+                    pv = values[center][pdir]
+                else:
+                    pv = 1.
+                point_i = directions[center][pdir] * pv + xyz
+                point_e = -directions[center][pdir] * pv + xyz
+                diff = point_e - point_i
+                points_array.append(point_e)
+                points_array.append(point_i)
+                centers_array.append(center)
+                centers_array.append(center)
+                diffs_array.append(diff)
+                diffs_array.append(diff)
+
+        points_array = np.asarray(points_array)
+        centers_array = np.asarray(centers_array)
+        diffs_array = np.asarray(diffs_array)
+
+        vtk_points = numpy_to_vtk_points(points_array)
+
+        vtk_cells = PeakActor._points_to_vtk_cells(points_array)
+
+        peaks_colors_object = PeakActor._peaks_colors_from_points(
+            points_array, colors=colors)
+        vtk_colors, colors_are_scalars, self.__go = peaks_colors_object
+
+        poly_data = vtk.vtkPolyData()
+        poly_data.SetPoints(vtk_points)
+        poly_data.SetLines(vtk_cells)
+        poly_data.GetPointData().SetScalars(vtk_colors)
+
+        self.__mapper = vtk.vtkPolyDataMapper()
+        self.__mapper.SetInputData(poly_data)
+        self.__mapper.ScalarVisibilityOn()
+        self.__mapper.SetScalarModeToUsePointFieldData()
+        self.__mapper.SelectColorArray('colors')
+        self.__mapper.Update()
+
+        self.SetMapper(self.__mapper)
+
+        attribute_to_actor(self, centers_array, 'center')
+        attribute_to_actor(self, diffs_array, 'diff')
+
+        vs_dec_code = load('peak_dec.vert')
+        vs_impl_code = load('peak_impl.vert')
+        fs_dec_code = load('peak_dec.frag')
+        fs_impl_code = load('peak_impl.frag')
+
+        shader_to_actor(self, 'vertex', decl_code=vs_dec_code,
+                        impl_code=vs_impl_code)
+        shader_to_actor(self, 'fragment', decl_code=fs_dec_code)
+        shader_to_actor(self, 'fragment', impl_code=fs_impl_code,
+                        block='light')
+
+        self.__lw = linewidth
+        self.GetProperty().SetLineWidth(self.__lw)
+
+        if self.__go >= 0:
+            self.GetProperty().SetOpacity(self.__go)
+
+        self.__min_centers = np.min(indices, axis=1)
+        self.__max_centers = np.max(indices, axis=1)
+
+        self.__is_range = True
[…]
