+++ fury/actors/tests/test_peak.py
@@ -1,0 +1,182 @@
+from fury.actors.peak import PeakActor
+from vtk.util import numpy_support
+
+
+import numpy as np
+import numpy.testing as npt
+
+
+def generate_peaks():
+    dirs01 = np.array([[-.4, .4, .8], [.7, .6, .1], [.4, -.3, .2],
+                       [0, 0, 0], [0, 0, 0]])
+    dirs10 = np.array([[.6, -.6, -.2], [0, 0, 0], [0, 0, 0], [0, 0, 0],
+                       [0, 0, 0]])
+    dirs11 = np.array([[0., .3, .3], [-.8, .4, -.5], [0, 0, 0], [0, 0, 0],
+                       [0, 0, 0]])
+    dirs12 = np.array([[0, 0, 0], [.7, .6, .1], [0, 0, 0], [0, 0, 0],
+                       [0, 0, 0]])
+
+    peaks_dirs = np.zeros((1, 2, 3, 5, 3))
+
+    peaks_dirs[0, 0, 1, :, :] = dirs01
+    peaks_dirs[0, 1, 0, :, :] = dirs10
+    peaks_dirs[0, 1, 1, :, :] = dirs11
+    peaks_dirs[0, 1, 2, :, :] = dirs12
+
+    peaks_vals = np.zeros((1, 2, 3, 5))
+
+    peaks_vals[0, 0, 1, :] = np.array([.3, .2, .6, 0, 0])
+    peaks_vals[0, 1, 0, :] = np.array([.5, 0, 0, 0, 0])
+    peaks_vals[0, 1, 1, :] = np.array([.2, .5, 0, 0, 0])
+    peaks_vals[0, 1, 2, :] = np.array([0, .7, 0, 0, 0])
+    return peaks_dirs, peaks_vals, np.eye(4)
+
+
+def test__orientation_colors():
+    points = np.array([[1, 0, 0], [-1, 0, 0], [0, 1, 0], [0, -1, 0], [0, 0, 1],
+                       [0, 0, -1]])
+
+    colors = PeakActor._orientation_colors(points, cmap='rgb_standard')
+    expected = [[1, 0, 0], [0, 1, 0], [0, 0, 1]]
+    npt.assert_array_equal(colors, expected)
+    # TODO: Add boys test
+
+
+def test__peaks_colors_from_points():
+    points = np.array([[1, 0, 0], [-1, 0, 0], [0, 1, 0], [0, -1, 0], [0, 0, 1],
+                       [0, 0, -1]])
+
+    peaks_colors_object = PeakActor._peaks_colors_from_points(
+        points, colors=None)
+    vtk_colors, colors_are_scalars, global_opacity = peaks_colors_object
+    desired = [[0, 0, 0], [0, 0, 0], [0, 0, 0], [0, 0, 0], [0, 0, 0],
+               [0, 0, 0]]
+    npt.assert_array_equal(numpy_support.vtk_to_numpy(vtk_colors), desired)
+    npt.assert_equal(colors_are_scalars, False)
+    npt.assert_equal(global_opacity, 1)
+
+    peaks_colors_object = PeakActor._peaks_colors_from_points(
+        points, colors='rgb_standard')
+    vtk_colors, colors_are_scalars, global_opacity = peaks_colors_object
+    desired = [[0, 0, 0], [0, 0, 0], [0, 0, 0], [0, 0, 0], [0, 0, 0],
+               [0, 0, 0]]
+    npt.assert_array_equal(numpy_support.vtk_to_numpy(vtk_colors), desired)
+    npt.assert_equal(colors_are_scalars, False)
+    npt.assert_equal(global_opacity, 1)
+
+    peaks_colors_object = PeakActor._peaks_colors_from_points(
+        points, colors=(0, 1, 0))
[…]
