+++ fury/tests/test_pick.py
@@ -72,32 +72,22 @@ def test_picking_manager():
 
 def test_selector_manager():
 
-    xyz = 10 * np.random.rand(100, 3)
-    colors = np.random.rand(100, 4)
-    radii = np.random.rand(100, 3) + 0.5
-
     centers = 0.5 * np.array([[0, 0, 0], [100, 0, 0], [200, 0, 0.]])
-    colors2 = np.array([[0.8, 0, 0], [0, 0.8, 0], [0, 0, 0.8]])
-    radii2 = 0.1 * np.array([50, 100, 150.])
+    colors = np.array([[0.8, 0, 0], [0, 0.8, 0], [0, 0, 0.8]])
+    radii = 0.1 * np.array([50, 100, 150.])
 
     scene = window.Scene()
 
-    # sphere_actor = actor.sphere(centers=xyz,
-    #                             colors=colors,
-    #                             radii=radii)
-
-    directions = np.array([[np.sqrt(2)/2, 0, np.sqrt(2)/2],
-                          [np.sqrt(2)/2, np.sqrt(2)/2, 0],
-                          [0, np.sqrt(2)/2, np.sqrt(2)/2]])
-    sphere_actor = actor.cube(centers, directions, colors2, scales=radii2)
+    cube_actor = actor.cube(centers, directions=(1, 0, 2), 
+                            colors=colors, scales=radii)
     # sphere_actor.GetProperty().SetRepresentationToWireframe()
-    print('Sphere actor', id(sphere_actor))
+    print('Sphere actor', id(cube_actor))
 
     pts = 100 * (np.random.rand(100, 3) - 0.5) + np.array([20, 0, 0.])
-    pts_actor = actor.dots(pts, dot_size=1)
+    pts_actor = actor.dots(pts, dot_size=10)
     print('Points actor', id(pts_actor))
 
-    scene.add(sphere_actor)
+    scene.add(cube_actor)
     scene.add(pts_actor)
 
     showm = window.ShowManager(scene,
@@ -113,23 +103,22 @@ def test_selector_manager():
 
     selm = pick.SelectionManager(select='faces')
 
-    record_indices = {'vertex_indices': [],
-                      'face_indices': [],
-                      'xyz': [],
-                      'actor': []}
-
     def timer_callback(_obj, _event):
         cnt = next(counter)
         tb.message = "Let's count up to 100 and exit :" + str(cnt)
-        showm.scene.azimuth(0.05 * cnt)
+        # showm.scene.azimuth(0.05 * cnt)
         # sphere_actor.GetProperty().SetOpacity(cnt/100.)
         if cnt % 10 == 0:
-            # pick at position
-            info = selm.select((900//2, 768//2), scene, 3)
-            record_indices['vertex_indices'].append(info['vertex'])
-            record_indices['face_indices'].append(info['face'])
-            record_indices['xyz'].append(info['xyz'])
-            record_indices['actor'].append(info['actor'])
+            # select large area
+            info_plus = selm.select((900//2, 768//2), scene, (30, 30))
+            print(info_plus)
+            for info in info_plus:
+                if info['actor'] in [cube_actor, pts_actor]:
+                    npt.assert_(True)
+                else:
+                    npt.assert_(False)
+            # print(id(info['actor']) == id(cube_actor))
+            # print(id(info['actor']) == id(pts_actor))
 
         showm.render()
         if cnt == 15:
@@ -138,29 +127,20 @@ def test_selector_manager():
 
     def hover_callback(_obj, _event):
         event_pos = selm.event_position(showm.iren)
-        info = selm.select(event_pos, showm.scene, 0)
-        print(info)
+        #pick a single pixel
[…]
