### 2021-05-16 99c3af1
Update selection picker to return multiple selected objects
