### 2021-05-17 314acb8
Addressed PEP8 issues. Added lookup colormap support and test.
