### 2021-05-17 4e44c4d
Simplified Selection test
