### 2021-05-17 febf67a
Added sanity checks tests for peaks directions, values and mask.
