### 2021-05-18 19d8e92
Added metallic left hemisphere example.
