### 2021-05-18 5bdb63c
Addressed fury/actor.py comments and suggestions.
