### 2021-05-18 68b7018
Added opacity and color properties.
