### 2021-05-20 c140228
Addressed fury/actors/peak.py comments and suggestions.
