+++ fury/actors/tests/test_peak.py
@@ -1,5 +1,6 @@
 from fury import actor, window
-from fury.actors.peak import PeakActor
+from fury.actors.peak import (PeakActor, _orientation_colors,
+                              _peaks_colors_from_points, _points_to_vtk_cells)
 from vtk.util import numpy_support
 
 
@@ -37,45 +38,43 @@ def test__orientation_colors():
     points = np.array([[1, 0, 0], [-1, 0, 0], [0, 1, 0], [0, -1, 0], [0, 0, 1],
                        [0, 0, -1]])
 
-    colors = PeakActor._orientation_colors(points, cmap='rgb_standard')
+    colors = _orientation_colors(points, cmap='rgb_standard')
     expected = [[1, 0, 0], [0, 1, 0], [0, 0, 1]]
     npt.assert_array_equal(colors, expected)
+
+    npt.assert_raises(ValueError, _orientation_colors, points, cmap='test')
 
 
 def test__peaks_colors_from_points():
     points = np.array([[1, 0, 0], [-1, 0, 0], [0, 1, 0], [0, -1, 0], [0, 0, 1],
                        [0, 0, -1]])
 
-    peaks_colors_object = PeakActor._peaks_colors_from_points(
-        points, colors=None)
-    vtk_colors, colors_are_scalars, global_opacity = peaks_colors_object
+    colors_tuple = _peaks_colors_from_points(points, colors=None)
+    vtk_colors, colors_are_scalars, global_opacity = colors_tuple
     desired = [[0, 0, 0], [0, 0, 0], [0, 0, 0], [0, 0, 0], [0, 0, 0],
                [0, 0, 0]]
     npt.assert_array_equal(numpy_support.vtk_to_numpy(vtk_colors), desired)
     npt.assert_equal(colors_are_scalars, False)
     npt.assert_equal(global_opacity, 1)
 
-    peaks_colors_object = PeakActor._peaks_colors_from_points(
-        points, colors='rgb_standard')
-    vtk_colors, colors_are_scalars, global_opacity = peaks_colors_object
+    colors_tuple = _peaks_colors_from_points(points, colors='rgb_standard')
+    vtk_colors, colors_are_scalars, global_opacity = colors_tuple
     desired = [[0, 0, 0], [0, 0, 0], [0, 0, 0], [0, 0, 0], [0, 0, 0],
                [0, 0, 0]]
     npt.assert_array_equal(numpy_support.vtk_to_numpy(vtk_colors), desired)
     npt.assert_equal(colors_are_scalars, False)
     npt.assert_equal(global_opacity, 1)
 
-    peaks_colors_object = PeakActor._peaks_colors_from_points(
-        points, colors=(0, 1, 0))
-    vtk_colors, colors_are_scalars, global_opacity = peaks_colors_object
+    colors_tuple = _peaks_colors_from_points(points, colors=(0, 1, 0))
+    vtk_colors, colors_are_scalars, global_opacity = colors_tuple
     desired = [[0, 255, 0], [0, 255, 0], [0, 255, 0], [0, 255, 0],
                [0, 255, 0], [0, 255, 0]]
     npt.assert_array_equal(numpy_support.vtk_to_numpy(vtk_colors), desired)
     npt.assert_equal(colors_are_scalars, False)
     npt.assert_equal(global_opacity, 1)
 
-    peaks_colors_object = PeakActor._peaks_colors_from_points(
-        points, colors=(0, 1, 0, .1))
-    vtk_colors, colors_are_scalars, global_opacity = peaks_colors_object
[…]
