+++ fury/actors/peak.py
@@ -9,6 +9,41 @@ import vtk
 
 
 class PeakActor(vtk.vtkActor):
+    """
+    VTK actor for visualizing slices of ODF field.
+
+    Parameters
+    ----------
+    directions : ndarray
+        Peak directions. The shape of the array should be (X, Y, Z, D, 3).
+    indices : tuple
+        Indices given in tuple(x_indices, y_indices, z_indices)
+        format for mapping 2D ODF array to 3D voxel grid.
+    values : ndarray, optional
+        Peak values. The shape of the array should be (X, Y, Z, D).
+    affine : array, optional
+        4x4 transformation array from native coordinates to world coordinates.
+    colors : None or string ('rgb_standard') or tuple (3D or 4D) or
+             array/ndarray (N, 3 or 4) or array/ndarray (K, 3 or 4) or
+             array/ndarray(N, ) or array/ndarray (K, )
+        If None a standard orientation colormap is used for every line.
+        If one tuple of color is used. Then all streamlines will have the same
+        color.
+        If an array (N, 3 or 4) is given, where N is equal to the number of
+        points. Then every point is colored with a different RGB(A) color.
+        If an array (K, 3 or 4) is given, where K is equal to the number of
+        lines. Then every line is colored with a different RGB(A) color.
+        If an array (N, ) is given, where N is the number of points then these
+        are considered as the values to be used by the colormap.
+        If an array (K,) is given, where K is the number of lines then these
+        are considered as the values to be used by the colormap.
+    lookup_colormap : vtkLookupTable, optional
+        Add a default lookup table to the colormap. Default is None which calls
+        :func:`fury.actor.colormap_lookup_table`.
+    linewidth : float, optional
+        Line thickness. Default is 1.
+
+    """
     def __init__(self, directions, indices, values=None, affine=None,
                  colors=None, lookup_colormap=None, linewidth=1):
         if affine is not None:
@@ -16,42 +51,43 @@ class PeakActor(vtk.vtkActor):
 
         valid_dirs = directions[indices]
 
-        points_array = []
-        centers_array = []
-        diffs_array = []
+        num_dirs = len(np.nonzero(np.abs(valid_dirs).max(axis=-1) > 0)[0])
+
+        pnts_per_line = 2
+
+        points_array = np.empty((num_dirs * pnts_per_line, 3))
+        centers_array = np.empty_like(points_array, dtype=int)
+        diffs_array = np.empty_like(points_array)
+        line_count = 0
         for idx, center in enumerate(zip(indices[0], indices[1], indices[2])):
             if affine is None:
-                xyz = np.array(center)
+                xyz = np.asarray(center)
             else:
                 xyz = w_pos[idx, :]
             valid_peaks = np.nonzero(
                 np.abs(valid_dirs[idx, :, :]).max(axis=-1) > 0.)[0]
-            for pdir in valid_peaks:
+            for direction in valid_peaks:
                 if values is not None:
-                    pv = values[center][pdir]
+                    pv = values[center][direction]
                 else:
                     pv = 1.
-                point_i = directions[center][pdir] * pv + xyz
-                point_e = -directions[center][pdir] * pv + xyz
+                point_i = directions[center][direction] * pv + xyz
+                point_e = -directions[center][direction] * pv + xyz
                 diff = point_e - point_i
-                points_array.append(point_e)
-                points_array.append(point_i)
-                centers_array.append(center)
-                centers_array.append(center)
-                diffs_array.append(diff)
-                diffs_array.append(diff)
-
-        points_array = np.asarray(points_array)
-        centers_array = np.asarray(centers_array)
-        diffs_array = np.asarray(diffs_array)
[…]
