### 2021-05-20 da49fe2
Added missing cmap.lower() comparison.
