### 2021-05-20 #3158 @danielskatz
👋 @Garyfallidis - Can you suggest some potential reviewers by mentioning them here in this thread (without tagging them with an @).

[This list of people](https://bit.ly/joss-reviewers) have already agreed to review for JOSS and may be suitable for this submission (please start at the bottom of the list).
