### 2021-05-20 #3158 @danielskatz
👋 @matthewturk - Would you (or someone in your group/lab) be willing to review this for JOSS?  As I think you know, we carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html
