### 2021-05-20 e24428e
New test added that play hoevering events
