### 2021-05-20 44617c5
Added check for sequence or not
