### 2021-05-20 c4e6c80
Added similar routine in PickingManager
