### 2021-05-20 33bf2a1
Recording seems accurate - testing of vertices and actor options of Selector are next
