### 2021-05-22 e05650a
Added vertices test
