### 2021-05-25 559ba9a
fix: pep8 issues, addressed review comments
