+++ fury/ui.py
@@ -5269,7 +5269,7 @@ class BulletList2D(UI):
         indent: int
             Base indentation for the root node
         draggable: Bool
-            If the card should be draggable
+            If the UI should be draggable
         position : (float, float)
             Absolute coordinates (x, y) of the lower-left corner of the
             UI component
@@ -5281,18 +5281,18 @@ class BulletList2D(UI):
             Background opacity
         """
 
-        self.points = points
+        self.points = list(points)
         self.nodes = []
         self.dict = {}
         self.content_size = size
         self.indent = np.clip(indent, 10, int(self.content_size[0]/2))
         self._last_child = None
         self.draggable = draggable
-        self.bg_color = bg_color
-        self.bg_opacity = bg_opacity
         self.text_color = text_color
         self.bullet_bg_color = bullet_bg_color
         super(BulletList2D, self).__init__()
+        self.panel.color = bg_color
+        self.panel.opacity = bg_opacity
         self.position = position
 
     def _setup(self):
@@ -5303,9 +5303,9 @@ class BulletList2D(UI):
         Add the bullet point to the Panel2D
         """
 
-        self.panel = Panel2D(size=self.content_size,
-                             color=self.bg_color, opacity=self.bg_opacity)
-        self.create_list(self.points, head=None)
+        self.panel = Panel2D(size=self.content_size)
+
+        self.create_nodes_list(self.points, head=None)
         _offsetY = 50
 
         for node in self.dict.items():
@@ -5356,47 +5356,44 @@ class BulletList2D(UI):
     def _get_size(self):
         self.panel.size
 
-    def create_list(self, lst, head=None):
+    def create_nodes_list(self, lst, head=None):
         """Create a flat list with all nodes ordered sequentially
         """
-        for ele in lst:
-            if isinstance(ele, list):
+        for element in lst:
+            if isinstance(element, list):
 
                 if not(len(self.nodes)):
                     raise TypeError(
                         f"First element should be str or str castable")
 
                 head = self.nodes[-1]
-                self.create_list(ele, head)
+                self.create_nodes_list(element, head)
                 head = None
             else:
-                ele = BulletPoint2D(node=ele, size=(self.content_size[0] -
-                                                    self.indent, 20),
-                                    text_color=self.text_color,
-                                    bg_color=self.bullet_bg_color)
+                element = BulletPoint2D(label=element,
+                                        size=(self.content_size[0] -
+                                              self.indent, 20))
                 if head:
-                    head.add_node(ele)
-                    self.nodes = [head if ele.node == head.node else
-                                  ele for ele in self.nodes]
-                self.nodes.append(ele)
-        self.create_dict()
+                    head.add_node(element)
+                    self.nodes = [head if node.label == head.label else
+                                  node for node in self.nodes]
+                self.nodes.append(element)
+        self.create_nodes_dict()
 
-    def create_dict(self):
+    def create_nodes_dict(self):
         """Create a dictionary from the nodes list
         """
 
-        for bullet_node in self.nodes:
-            self.dict[bullet_node.node] = bullet_node
+        for node in self.nodes:
+            self.dict[node.label] = node
 
     def add_child_node(self, root_node, child_node):
         """Add a child node in a specific node
         """
         root_node = self.dict[root_node]
-        child_node = BulletPoint2D(node=child_node,
+        child_node = BulletPoint2D(label=child_node,
                                    size=(self.content_size[0] -
-                                         root_node.indent - 30, 20),
[…]
