### 2021-05-25 d92e70e
fix: updated tests for changed method names
