### 2021-05-25 1b22ac3
Adding new texture actor for 2d overlays
