+++ docs/examples/viz_glassy_brain.py
@@ -3,66 +3,12 @@ from fury import actor, ui, window
 from fury.data import read_viz_textures
 from fury.io import load_polydata
 from fury.utils import get_actor_from_polydata
-from fury.shaders import shader_to_actor
+from fury.shaders import add_shader_callback, load, shader_to_actor
 
 
 import numpy as np
 import os
 import vtk
-
-
-__FRAGMENT_DEC = \
-    """
-    float chiGGX(float v)
-    {
-        return v > 0 ? 1. : .0;
-    }
-    
-    vec3 FresnelSchlick(float HdV, vec3 F0)
-    {
-        return F0 + (1 - F0) * pow(1 - HdV, 5);
-    }
-    
-    float GGXDistribution(float NdH, float alpha)
-    {
-        float alpha2 = alpha * alpha;
-        float NdH2 = NdH * NdH;
-        float den = NdH2 * alpha2 + (1 - NdH2);
-        return (chiGGX(NdH) * alpha2) / (PI * den * den);
-    }
-    
-    float GGXPartialGeometryTerm(float VdH, float VdN, float alpha)
-    {
-        float cVdH = clamp(VdH, .0, 1.);
-        float chi = chiGGX(cVdH / clamp(VdN, .0, 1.));
-        float tan2 = (1 - cVdH) / cVdH;
-        return (chi * 2) / (1 + sqrt(1 + alpha * alpha * tan2));
-    }
-    """
-__FRAGMENT_IMPL = \
-    """
-    vec3 ior = vec3(2.4);
-    vec3 F0_t = abs((1. - ior) / (1. + ior));
-    F0_t *= F0_t;
-    F0_t = mix(F0_t, albedo, metallic);
-    
-    vec3 Lo_t = vec3(.0);
-    
-    vec3 F_t = F_Schlick(1., F0_t);
-    vec3 specular_t = D * Vis * F_t;
-    vec3 diffuse_t = (1. - metallic) * (1. - F_t) * DiffuseLambert(albedo);
-    Lo_t += (diffuse_t + specular_t) * lightColor0 * NdV;
-    
-    vec3 kS_t = F_SchlickRoughness(max(NdV, .0), F0_t, roughness);
-    vec3 kD_t = 1. - kS_t;
-    kD_t *= 1. - metallic;
-    vec3 ambient_t = (kD_t * irradiance * albedo + prefilteredColor * (kS_t * brdf.r + brdf.g));
-    vec3 color_t = ambient_t + Lo_t;
-    color_t = mix(color_t, color_t * ao, aoStrengthUniform);
-    color_t += emissiveColor;
-    color_t = pow(color_t, vec3(1. / 2.2));
-    fragOutput0 = vec4(color_t, opacity);
-    """
 
 
 def build_label(text, font_size=16, color=(1, 1, 1), bold=False, italic=False,
@@ -86,9 +32,29 @@ def change_slice_metallic(slider):
     obj_actor.GetProperty().SetMetallic(slider._value)
 
 
+def change_slice_specular(slider):
+    global obj_actor
+    obj_actor.GetProperty().SetSpecular(slider._value)
+
+
+def change_slice_specular_tint(slider):
+    global obj_actor
+    obj_actor.GetProperty().SetSpecularPower(slider._value)
+
+
 def change_slice_roughness(slider):
     global obj_actor
     obj_actor.GetProperty().SetRoughness(slider._value)
+
+
+def change_slice_sheen(slider):
+    global sheen
+    sheen = slider._value
+
+
+def change_slice_sheen_tint(slider):
+    global sheen_tint
+    sheen_tint = slider._value
 
 
 def get_cubemap(files_names):
[…]
