### 2021-05-25 b3ece39
Starting updating tutorial to combine texture_2d with selector
