### 2021-05-27 634a328
adds a simple FuryStreamClient
related with #435
