### 2021-05-27 fce8e07
allows to stop the stream if requested
