### 2021-05-28 6e925a4
docstring syntax fix
