+++ fury/tests/test_actors.py
@@ -1,9 +1,3 @@
-npoints^2 = number of points which will be used for generating
-        the triangles. The quality of the surface generated will be
-        better if npoints is high but this becomes computationally taxing
-        for large values of n.
-        default: 100 i.e. by default, 100^2 = 10,000 points are used to
-                generate a parametric surface
 import os
 import itertools
 from tempfile import TemporaryDirectory as InTemporaryDirectory
