### 2021-05-28 8626370
a simple example using shared mem. and streaming client
