### 2021-05-28 bd2c6d8
Removed VTK specular. Added anisotropic slider. Refactored BRDF panel.
