+++ docs/examples/viz_bxdf_pbr.py
@@ -36,18 +36,25 @@ def change_slice_metallic(slider):
 
 
 def change_slice_specular(slider):
-    global obj_actor
-    obj_actor.GetProperty().SetSpecular(slider._value)
+    global obj_actor, specular
+    #obj_actor.GetProperty().SetSpecular(slider._value)
+    specular = slider._value
 
 
 def change_slice_specular_tint(slider):
-    global obj_actor
-    obj_actor.GetProperty().SetSpecularPower(slider._value)
+    global obj_actor, specular_tint
+    #obj_actor.GetProperty().SetSpecularPower(slider._value)
+    specular_tint = slider._value
 
 
 def change_slice_roughness(slider):
     global obj_actor
     obj_actor.GetProperty().SetRoughness(slider._value)
+
+
+def change_slice_anisotropic(slider):
+    global anisotropic
+    anisotropic = slider._value
 
 
 def change_slice_sheen(slider):
@@ -130,8 +137,12 @@ def obj_surface():
 
 
 def uniforms_callback(_caller, _event, calldata=None):
-    global clearcoat, clearcoat_gloss, sheen, sheen_tint
+    global anisotropic, clearcoat, clearcoat_gloss, sheen, sheen_tint, \
+        specular, specular_tint
     if calldata is not None:
+        calldata.SetUniformf('specularValue', specular)
+        calldata.SetUniformf('specularTint', specular_tint)
+        calldata.SetUniformf('anisotropic', anisotropic)
         calldata.SetUniformf('sheen', sheen)
         calldata.SetUniformf('sheenTint', sheen_tint)
         calldata.SetUniformf('clearcoat', clearcoat)
@@ -139,16 +150,17 @@ def uniforms_callback(_caller, _event, calldata=None):
 
 
 def win_callback(obj, event):
-    global panel, size
+    global brdf_panel, size
     if size != obj.GetSize():
         size_old = size
         size = obj.GetSize()
         size_change = [size[0] - size_old[0], 0]
-        panel.re_align(size_change)
+        brdf_panel.re_align(size_change)
 
 
 if __name__ == '__main__':
-    global clearcoat, clearcoat_gloss, panel, sheen, sheen_tint, size
+    global anisotropic, clearcoat, clearcoat_gloss, panel, sheen, sheen_tint, \
+        size, specular, specular_tint
 
     #obj_actor = obj_brain()
     #obj_actor = obj_surface()
@@ -158,6 +170,7 @@ if __name__ == '__main__':
     specular = .0
     specular_tint = .0
     roughness = .0
+    anisotropic = .0
     sheen = .0
     sheen_tint = .0
     clearcoat = .0
@@ -166,11 +179,12 @@ if __name__ == '__main__':
     # TODO: Add opacity panel
     obj_actor.GetProperty().SetOpacity(.5)
 
-    #specular_color = vtk.vtkNamedColors().GetColor3d('White')
-
     obj_actor.GetProperty().SetInterpolationToPBR()
     obj_actor.GetProperty().SetMetallic(metallic)
     obj_actor.GetProperty().SetRoughness(roughness)
+
+    # NOTE: Specular parameters don't seem to work
+    #specular_color = vtk.vtkNamedColors().GetColor3d('White')
     #obj_actor.GetProperty().SetSpecular(specular)
     #obj_actor.GetProperty().SetSpecularPower(specular_tint)
     #obj_actor.GetProperty().SetSpecularColor(specular_color)
@@ -221,9 +235,11 @@ if __name__ == '__main__':
                                 order_transparent=True)
     show_m.initialize()
 
[…]
