+++ fury/stream/___init__.py
@@ -1,0 +1,4 @@
+from fury.stream.webrtc.server import webrtc_server
+
+
+__all__ = ['webrtc_server']
