+++ fury/stream/webrtc/www/client.js
@@ -1,0 +1,197 @@
+var pc = null;
+var bandwidth = 5;
+function removeBandwidthRestriction(sdp) {
+  return sdp.replace(/b=AS:.*\r\n/, "").replace(/b=TIAS:.*\r\n/, "");
+}
+
+function setMediaBitrates(sdp) {
+  return setMediaBitrate(setMediaBitrate(sdp, "video", 5), "audio", 50);
+}
+
+function setMediaBitrate(sdp, media, bitrate) {
+  var lines = sdp.split("\n");
+  var line = -1;
+  for (var i = 0; i < lines.length; i++) {
+    if (lines[i].indexOf("m=" + media) === 0) {
+      line = i;
+      break;
+    }
+  }
+  if (line === -1) {
+    console.debug("Could not find the m line for", media);
+    return sdp;
+  }
+  console.debug("Found the m line for", media, "at line", line);
+
+  // Pass the m line
+  line++;
+
+  // Skip i and c lines
+  while (lines[line].indexOf("i=") === 0 || lines[line].indexOf("c=") === 0) {
+    line++;
+  }
+
+  // If we're on a b line, replace it
+  if (lines[line].indexOf("b") === 0) {
+    console.debug("Replaced b line at line", line);
+    lines[line] = "b=AS:" + bitrate;
+    return lines.join("\n");
+  }
+
+  // Add a new b line
+  console.debug("Adding new b line before line", line);
+  var newLines = lines.slice(0, line);
+  newLines.push("b=AS:" + bitrate);
+  newLines = newLines.concat(lines.slice(line, lines.length));
+  return newLines.join("\n");
+}
+// videoBandwidth = 10;
+function updateBandwidthRestriction(sdp, bandwidth) {
+  let modifier = "AS";
+  if (adapter.browserDetails.browser === "firefox") {
+    bandwidth = (bandwidth >>> 0) * 1000;
+    modifier = "TIAS";
+  }
+  if (sdp.indexOf("b=" + modifier + ":") === -1) {
+    // insert b= after c= line.
+    sdp = sdp.replace(
+      /c=IN (.*)\r\n/,
+      "c=IN $1\r\nb=" + modifier + ":" + bandwidth + "\r\n"
+    );
+  } else {
+    sdp = sdp.replace(
+      new RegExp("b=" + modifier + ":.*\r\n"),
+      "b=" + modifier + ":" + bandwidth + "\r\n"
+    );
+  }
+  return sdp;
+}
+
+function negotiate() {
+  pc.addTransceiver("video", { direction: "recvonly" });
+  pc.addTransceiver("audio", { direction: "recvonly" });
+  return pc
+    .createOffer()
+    .then(function (offer) {
+      let offerDescription = pc.setLocalDescription(offer);
+      offer.sdp = setMediaBitrates(offer.sdp);
+      return offerDescription;
+    })
+    .then(function () {
+      // wait for ICE gathering to complete
+      return new Promise(function (resolve) {
+        if (pc.iceGatheringState === "complete") {
+          resolve();
+        } else {
+          function checkState() {
+            if (pc.iceGatheringState === "complete") {
+              pc.removeEventListener("icegatheringstatechange", checkState);
+              resolve();
+            }
+          }
+          pc.addEventListener("icegatheringstatechange", checkState);
+        }
+      });
+    })
+    .then(function () {
+      var offer = pc.localDescription;
+      return fetch("/offer", {
+        body: JSON.stringify({
+          sdp: setMediaBitrates(offer.sdp),
+          type: offer.type,
+        }),
+        headers: {
+          "Content-Type": "application/json",
+        },
+        method: "POST",
+      });
+    })
+    .then(function (response) {
+      return response.json();
+    })
[…]
