### 2021-05-29 e8044bc
reorganize stream files and folders
Doing that way because the FuryStreamClient will have a lot of more behaviors, like user interaction.
