### 2021-05-29 0c919a7
this separates the js script of webrtcstream into a module
