+++ fury/stream/servers/webrtc/www/js/main.js
@@ -1,0 +1,10 @@
+import {startWebRTC} from '/js/webrtc.js'
+
+document.addEventListener("DOMContentLoaded", (event) => {
+  document.getElementById('startBtn').addEventListener('click', (e)=>{startWebRTC()})
+});
+
+window.addEventListener('wheel', (event)=>
+{
+    console.log(event.deltaY)
+})
