### 2021-05-31 af66c6d
implements a circular queue using shared buffers
This can be used for example to store/read user interactions
