### 2021-05-31 72436bc
a porposal for user interaction for furystream
This uses a circular queue as proposed by @filipinascimento
