### 2021-05-31 8b12ff2
simplifies the use of the streaming system
