+++ docs/examples/viz_streaming.py
@@ -17,30 +17,8 @@ if __name__ == '__main__':
     # First we will set the resolution which it'll be used by the streamer
 
     window_size = (400, 400)
-    ##############################################################################
-    # ms define the amount of mileseconds that will be used in the timer event.
-    # Otherwise, if ms it's equal to zero the shared memory it's updated in each 
-    # render event
-    ms = 16
-    ###############################################################################
-    #
-    arr = np.random.randint(
-        0, 255, size=window_size[0]*window_size[1]*3).astype('uint8')
-    #
-    image_buffer = multiprocessing.RawArray('B', arr)
-    #
-    arr = None
-    #
-    arr = np.array([window_size[0], window_size[1], 3])
-    info_buffer = multiprocessing.RawArray('I', arr)
-    arr = None
-    ###########################################################################
-    #
-    p = multiprocessing.Process(
-        target=webrtc_server, args=(image_buffer, info_buffer))
-    p.start()
 
-    #
+  
     files, folder = fetch_viz_wiki_nw()
     categories_file, edges_file, positions_file = sorted(files.keys())
     positions = np.loadtxt(pjoin(folder, positions_file))
@@ -82,17 +60,27 @@ if __name__ == '__main__':
 
     interactive = False
 
-    scene.set_camera(position=(0, 0, 1000), focal_point=(0.0, 0.0, 0.0),
-                    view_up=(0.0, 0.0, 0.0))
-                    
+    scene.set_camera(
+        position=(0, 0, 1000),
+        focal_point=(0.0, 0.0, 0.0), view_up=(0.0, 0.0, 0.0))
+
     showm = window.ShowManager(scene, reset_camera=False, size=(
         window_size[0], window_size[1]), order_transparent=False,
         # multi_samples=8
     )
 
     from fury.stream.client import FuryStreamClient
+    ##############################################################################
+    # ms define the amount of mileseconds that will be used in the timer event.
+    # Otherwise, if ms it's equal to zero the shared memory it's updated in each 
+    # render event
+    ms = 16
     showm.initialize()
     stream = FuryStreamClient(
-        showm, False, image_buffer=image_buffer, info_buffer=info_buffer)
+        showm, window_size=window_size)
+    p = multiprocessing.Process(
+        target=webrtc_server, args=(stream, ))
+    p.start()
+
     stream.init(ms,)
     showm.start()
