+++ docs/tutorials/01_introductory/viz_selection.py
@@ -4,35 +4,40 @@
 Selecting multiple objects
 ==========================
 
-Here we present a tutorial showing how to interact with objects in the
+Here we show how to interact with objects in the
 3D world. All objects to be picked are part of a single actor.
-FURY likes to bundle objects in a few actors to reduce code and
-increase speed.
 
-When the objects will be picked they will change size and color.
+FURY likes to bundle objects in a few actors to reduce code and
+increase speed. Nonetheless the method works for multiple actors too.
+
+The difference with the picking tutorial is that here we will
+be able to select more than one object. In addition we can
+select interactively many vertices or faces.
+
+In summary, we will create an actor with thousands of cubes and
+then interactively we will be moving a rectangular box by
+hovering the mouse and making transparent everything that is
+behind that box.
+
 """
 
 import numpy as np
-from fury import actor, window, ui, utils, pick
+from fury import actor, window, utils, pick
+
+###############################################################################
+# Adding many cubes of different sizes and colors
 
 num_cubes = 50000
 
 centers = 10000 * (np.random.rand(num_cubes, 3) - 0.5)
 colors = np.random.rand(num_cubes, 4)
+colors[:, 3] = 1.0
 radii = 100 * np.random.rand(num_cubes) + 0.1
 
-num_faces = num_cubes * 6 * 2  # every quad of each cubes has 2 triangles
-
-selected = np.zeros(num_cubes, dtype=bool)
-
 ###############################################################################
-# Let's create a panel to show what is picked
+# Keep track of total number of faces
 
-panel = ui.Panel2D(size=(400, 200), color=(1, .5, .0), align="right")
-panel.center = (150, 200)
-
-text_block = ui.TextBlock2D(text="Left click on object \n")
-panel.add_element(text_block, (0.3, 0.3))
+num_faces = num_cubes * 6 * 2  # every quad of each cubes has 2 triangles
 
 ###############################################################################
 # Build scene and add an actor with many objects.
@@ -40,9 +45,10 @@ panel.add_element(text_block, (0.3, 0.3))
 scene = window.Scene()
 
 ###############################################################################
-# This actor is using many cubes
+# Build the actor containing all the cubes
 
-cube_actor = actor.cube(centers, directions=(1, 0, 0), colors=colors, scales=radii)
+cube_actor = actor.cube(centers, directions=(1, 0, 0),
+                        colors=colors, scales=radii)
 
 ###############################################################################
 # Access the memory of the vertices of all the cubes
@@ -57,44 +63,51 @@ num_objects = centers.shape[0]
 vcolors = utils.colors_from_actor(cube_actor, 'colors')
 
 ###############################################################################
-# Adding an actor showing the axes of the world coordinates
-ax = actor.axes(scale=(10, 10, 10))
+# Create a rectangular 2d box as a texture
 
 rgba = 255 * np.ones((100, 200, 4))
 rgba[1:-1, 1:-1] = np.zeros((98, 198, 4))
-# rgba = np.round(255 * np.random.rand(100, 200, 4), 0)
-
 texa = actor.texture_2d(rgba.astype(np.uint8))
 
 scene.add(cube_actor)
-scene.add(ax)
 scene.add(texa)
 scene.reset_camera()
+scene.zoom(3.)
 
 ###############################################################################
-# Create the Picking manager
+# Create the Selection Manager
 
 selm = pick.SelectionManager(select='faces')
 
-
-selm.selectable_off([texa, ax])
 ###############################################################################
-# Time to make the callback which will be called when we pick an object
+# Tell Selection Manager to avoid selecting specific actors
+
+selm.selectable_off(texa)
+
+###############################################################################
+# Time to make the callback which will be called
+# when we hover the mouse
 
 
 def hover_callback(_obj, _event):
     event_pos = selm.event_position(showm.iren)
[…]
