### 2021-06-01 4cd1fb9
Updated docstring for texture_2d
