+++ docs/tutorials/01_introductory/viz_selection.py
@@ -127,7 +127,7 @@ showm.add_iren_callback(hover_callback)
 ###############################################################################
 # Change interactive to True to start interacting with the scene
 
-interactive = True
+interactive = False
 
 if interactive:
 
