+++ fury/stream/servers/webrtc/www/js/constants.js
@@ -1,0 +1,1 @@
+export const millisecMouseMove = 80

[…]
