### 2021-06-02 7eab5a7
fixes mouse movement in Y-axis
It seems that vtk uses left-bottom coordinates
