### 2021-06-02 be0b65f
this avoid pickle issues in OSX
macOS dosen't use fork to achieve multiprocessing.
Therefore, this create some issues for specific python
versions. The solution it's to pass just the buffers to
the childreen process
