+++ fury/stream/servers/webrtc/server.py
@@ -5,22 +5,30 @@ from aiortc import VideoStreamTrack
 import numpy as np
 
 from fury.stream.servers.webrtc.async_app import get_app
+from fury.stream.tools import CircularQueue
 
 
 def webrtc_server(
-        fury_stream_client,
-        # image_buffer, info_buffer, 
-        circular_quequeue=None,
+        stream_client=None,
+        image_buffer=None, info_buffer=None,
+        circular_queue=None,
+        queue_head_tail_buffer=None,
+        queue_buffers_list=None,
         port=8000, host='localhost', flip_img=True,
         www_folder=None, use_vidgear=False):
+
+    if stream_client is not None:
+        image_buffer = stream_client.image_buffer
+        info_buffer = stream_client.info_buffer
 
     class RTCServer(VideoStreamTrack):
         def __init__(self,):
             super().__init__()
 
             # starts with a random image
+
             image_info = np.frombuffer(
-                fury_stream_client.info_buffer, 'uint32')
+                info_buffer, 'uint32')
             # self.image = np.random.randint(
             #      0, 255, (image_info[1], image_info[0], 3),
             #      dtype='uint8')
@@ -31,10 +39,10 @@ def webrtc_server(
             pts, time_base = await self.next_timestamp()
 
             image_info = np.frombuffer(
-                fury_stream_client.info_buffer, 'uint32')
+                info_buffer, 'uint32')
 
             self.image = np.frombuffer(
-                fury_stream_client.image_buffer, 'uint8').reshape(
+                image_buffer, 'uint8').reshape(
                 (image_info[1], image_info[0], 3))
 
             if flip_img:
@@ -51,7 +59,11 @@ def webrtc_server(
                     self.stream = None
             except AttributeError:
                 pass
+    if circular_queue is None:
+        circular_queue = CircularQueue(
+            head_tail_buffer=queue_head_tail_buffer,
+            buffers_list=queue_buffers_list)
 
-    app_fury = get_app(RTCServer(), circular_quequeue=circular_quequeue)
+    app_fury = get_app(RTCServer(), circular_quequeue=circular_queue)
     web.run_app(
         app_fury, host=host, port=port, ssl_context=None)
