### 2021-06-02 c56d3f0
simplifies the use of the StreamInteraction Class
