### 2021-06-02 38f1f9b
allows to run the stream inside of a online server
For example, using ngrok
./ngrok http --subdomain=something 8000
