+++ fury/stream/servers/webrtc/www/js/main.js
@@ -1,24 +1,28 @@
-import { startWebRTC } from "/js/webrtc.js"
-import { millisecMouseMove } from "/js/constants.js"
+import { startWebRTC } from "/js/webrtc.js";
+import { millisecMouseMove } from "/js/constants.js";
 
-let mouseLeftReleased = true
-let mouseOutVideo = false
-let mouseX = 0
-let mouseY = 0
-let ctrlKey = 0
-let shiftKey = 0
-const videoEl = document.getElementById("video")
+let mouseLeftReleased = true;
+let mouseOutVideo = false;
+let mouseX = 0;
+let mouseY = 0;
+let ctrlKey = 0;
+let shiftKey = 0;
+let urlServer = "";
+const videoEl = document.getElementById("video");
 
 document.addEventListener("DOMContentLoaded", (event) => {
   document.getElementById("startBtn").addEventListener("click", (e) => {
-    startWebRTC()
-  })
+    startWebRTC();
+  });
+  let urlServer = `${location.protocol}//${location.hostname}${
+    location.port ? ":" + location.port : ""
+  }/`
 })
 
 videoEl.addEventListener("wheel", (event) => {
-  const data = { deltaY: event.deltaY }
+  const data = { deltaY: event.deltaY };
 
-  fetch("http://localhost:8000/mouse_weel", {
+  fetch(`${urlServer}mouse_weel`, {
     method: "POST", // or 'PUT'
     headers: {
       "Content-Type": "application/json",
@@ -30,25 +34,23 @@ videoEl.addEventListener("wheel", (event) => {
       // console.log("Success:", data)
     })
     .catch((error) => {
-      console.error("Error:", error)
-    })
-})
-
+      console.error("Error:", error);
+    });
+});
 
 videoEl.addEventListener("mousemove", (event) => {
-  mouseOutVideo = false
-  ctrlKey = event.ctrlKey ? 1:0
-  shiftKey = event.shiftKey ? 1:0
-  const width = videoEl.offsetWidth
-  const height = videoEl.offsetHeight
-  mouseX = event.offsetX/width
-  mouseY = event.offsetY/height
-  
-})
+  mouseOutVideo = false;
+  ctrlKey = event.ctrlKey ? 1 : 0;
+  shiftKey = event.shiftKey ? 1 : 0;
+  const width = videoEl.offsetWidth;
+  const height = videoEl.offsetHeight;
+  mouseX = event.offsetX / width;
+  mouseY = event.offsetY / height;
+});
 
 videoEl.addEventListener("mouseleave", (e) => {
-  mouseOutVideo = false
-})
+  mouseOutVideo = false;
+});
 // videoEl.addEventListener("mousemove", (e) => {
 //   if (mouseLeftReleased) return
 //   mouseX = e.offsetX
@@ -75,10 +77,10 @@ videoEl.addEventListener("mouseleave", (e) => {
 //     })
 // })
 
-const mouseMoveCallback = ()=>{
-    if (mouseLeftReleased || mouseOutVideo) return
-    const data = {x:mouseX, y:mouseY, ctrlKey:ctrlKey, shiftKey:shiftKey}
-    fetch("http://localhost:8000/mouse_move", {
+const mouseMoveCallback = () => {
+  if (mouseLeftReleased || mouseOutVideo) return;
+  const data = { x: mouseX, y: mouseY, ctrlKey: ctrlKey, shiftKey: shiftKey };
+  fetch(`${urlServer}mouse_move`, {
     method: "POST", // or 'PUT'
     headers: {
       "Content-Type": "application/json",
@@ -90,76 +92,75 @@ const mouseMoveCallback = ()=>{
       // console.log("Success:", data)
     })
[…]
