### 2021-06-03 c743f63
Added mouse wheel throtling Fix zoom factor for throttling
