### 2021-06-03 504e17f
improves the organization, add iframe mode
iframe mode, background color and interaction can
be defined through url parameters

Example:
http://localhost:8000/?interaction=1&iframe=0&background=black
