+++ fury/stream/servers/webrtc/www/js/main.js
@@ -1,40 +1,51 @@
 import { startWebRTC } from "/js/webrtc.js";
-import { millisecMouseMove } from "/js/constants.js";
-
-let interactionInterval = 200; // for throttling
-let mouseLeftReleased = true;
-let mouseOutVideo = false;
-let mouseX = 0;
-let mouseY = 0;
-let ctrlKey = 0;
-let shiftKey = 0;
-let urlServer = "";
-const videoEl = document.getElementById("video");
+import {
+  millisecMouseMove,
+  urlServer,
+  interactionInterval,
+  videoEl,
+  urlParams,
+} from "/js/constants.js";
 
 document.addEventListener("DOMContentLoaded", (event) => {
   document.getElementById("startBtn").addEventListener("click", (e) => {
     startWebRTC();
   });
-  let urlServer = `${location.protocol}//${location.hostname}${
-    location.port ? ":" + location.port : ""
-  }/`
-})
+  const interaction = urlParams.get("interaction");
+  const runningOnIframe = urlParams.get("iframe");
+  const backgroundColor = urlParams.get('background')
+  if (interaction === null || interaction == 1) addInteraction(videoEl);
+  const videoClass =
+    runningOnIframe == null || runningOnIframe == 1
+      ? "videoIframeMode"
+      : "videoNormalMode";
+  videoEl.className = videoClass
+  if (backgroundColor != null)
+    document.body.style.backgroundColor = backgroundColor;
+});
 
-let enableCallWheel = true;
-let currentWheelEventTotalDeltaY = 0;
-videoEl.addEventListener("wheel", (event) => {
-  currentWheelEventTotalDeltaY+=event.deltaY;
-  if (!enableCallWheel) return;
-  const data = { deltaY: currentWheelEventTotalDeltaY };
-  currentWheelEventTotalDeltaY=0;
+function addInteraction(videoEl) {
+  let mouseLeftReleased = true;
+  let mouseOutVideo = false;
+  let mouseX = 0;
+  let mouseY = 0;
+  let ctrlKey = 0;
+  let shiftKey = 0;
+  let enableCallWheel = true;
+  let currentWheelEventTotalDeltaY = 0;
+  videoEl.addEventListener("wheel", (event) => {
+    currentWheelEventTotalDeltaY += event.deltaY;
+    if (!enableCallWheel) return;
+    const data = { deltaY: currentWheelEventTotalDeltaY };
+    currentWheelEventTotalDeltaY = 0;
 
-  fetch(`${urlServer}mouse_weel`, {
-    method: "POST", // or 'PUT'
-    headers: {
-      "Content-Type": "application/json",
-    },
-    body: JSON.stringify(data),
-  })
+    fetch(`${urlServer}mouse_weel`, {
+      method: "POST", // or 'PUT'
+      headers: {
+        "Content-Type": "application/json",
+      },
+      body: JSON.stringify(data),
+    });
     // .then((response) => response.json())
     // .then((data) => {
     //   // console.log("Success:", data)
@@ -42,136 +53,137 @@ videoEl.addEventListener("wheel", (event) => {
     // .catch((error) => {
     //   console.error("Error:", error);
     // });
-  enableCallWheel = false;
-  setTimeout(() => enableCallWheel = true, interactionInterval);
-});
+    enableCallWheel = false;
+    setTimeout(() => (enableCallWheel = true), interactionInterval);
+  });
 
-videoEl.addEventListener("mousemove", (event) => {
-  mouseOutVideo = false;
-  ctrlKey = event.ctrlKey ? 1 : 0;
-  shiftKey = event.shiftKey ? 1 : 0;
-  const width = videoEl.offsetWidth;
-  const height = videoEl.offsetHeight;
-  mouseX = event.offsetX / width;
[…]
