### 2021-06-03 4548565
allows to create a FuryStreamClient and Interaction with predefined buffers
This allows you to start a webrtc server in the main process
