### 2021-06-03 7145a0c
a simple example using the marker actor
