### 2021-06-04 #3158 @chrishavlin
@danielskatz should the review checklist be automatically added or do I add that? Or does that show up when this is switched from Pre Review to Review?
