+++ fury/stream/servers/webrtc/async_app.py
@@ -3,7 +3,7 @@ import json
 import os
 import numpy as np
 from functools import partial
-
+import aiohttp
 from aiohttp import web
 
 from aiortc import RTCPeerConnection, RTCSessionDescription
@@ -29,10 +29,11 @@ async def offer(request, **kwargs):
     video = kwargs['video']
     if("broadcast" in kwargs and kwargs["broadcast"]):
         video = MediaRelay().subscribe(video)
-    
+
     params = await request.json()
-    #print(params["sdp"])
-    offer = RTCSessionDescription(sdp=params["sdp"], type=params["type"])
+
+    offer = RTCSessionDescription(
+        sdp=params["sdp"], type=params["type"])
 
     pc = RTCPeerConnection()
     pcs.add(pc)
@@ -60,18 +61,26 @@ async def offer(request, **kwargs):
     return web.Response(
         content_type="application/json",
         text=json.dumps(
-            {"sdp": pc.localDescription.sdp, "type": pc.localDescription.type}
+            {
+                "sdp": pc.localDescription.sdp,
+                "type": pc.localDescription.type}
         ),
     )
+
+
+def set_weel(data, circular_queue):
+    deltaY = float(data['deltaY'])
+
+    ok = circular_queue.enqueue(
+        np.array([1, deltaY, 0, 0, 0, 0], dtype='float64'))
+
+    return ok
 
 
 async def mouse_weel(request, **kwargs):
 
     params = await request.json()
-    deltaY = float(params['deltaY'])
-    circular_queue = kwargs['circular_queue']
-    ok = circular_queue.enqueue(
-        np.array([1, deltaY, 0, 0, 0, 0], dtype='float64'))
+    ok = set_weel(params, kwargs['circular_queue'])
     return web.Response(
         content_type="application/json",
         text=json.dumps(
@@ -80,17 +89,23 @@ async def mouse_weel(request, **kwargs):
     )
 
 
+def set_mouse(data, circular_queue):
+    x = float(data['x'])
+    y = float(data['y'])
+    ctrl_key = int(data['ctrlKey'])
+    shift_key = int(data['shiftKey'])
+
+    circular_queue = circular_queue
+    ok = circular_queue.enqueue(
+        np.array([2, 0, x, y,  ctrl_key, shift_key], dtype='float64'))
+
+    return ok
+
+
 async def mouse_move(request, **kwargs):
 
     params = await request.json()
-    x = float(params['x'])
-    y = float(params['y'])
-    ctrl_key = int(params['ctrlKey'])
-    shift_key = int(params['shiftKey'])
-
-    circular_queue = kwargs['circular_queue']
-    ok = circular_queue.enqueue(
-        np.array([2, 0, x, y,  ctrl_key, shift_key], dtype='float64'))
+    ok = set_mouse(params, kwargs['circular_queue'])
     return web.Response(
         content_type="application/json",
         text=json.dumps(
@@ -137,8 +152,36 @@ async def on_shutdown(app):
     pcs.clear()
 
 
-def get_app(RTCServer, folder=None, circular_queue=None,broadcast=True):
-    
+async def websocket_handler(request, **kwargs):
+
+    circular_queue = kwargs['circular_queue']
+    ws = web.WebSocketResponse()
+    await ws.prepare(request)
+
+    async for msg in ws:
+        if msg.type == aiohttp.WSMsgType.TEXT:
+            if msg.data == 'close':
+                await ws.close()
+            else:
+                data = json.loads(msg.data)
+                if data['type'] == 'weel':
+                    set_weel(data, circular_queue)
+                elif data['type'] == 'mouseMove':
+                    set_mouse(data, circular_queue)
+                #print(msg)
+                #await ws.send_str(msg.data + '/answer')
[…]
