+++ fury/stream/servers/webrtc/www/js/constants.js
@@ -2,7 +2,7 @@ export const millisecMouseMove = 80;
 export const urlServer = `${location.protocol}//${location.hostname}${
   location.port ? ":" + location.port : ""
 }/`;
-export const interactionInterval = 200;
+export const interactionInterval = 80;
 export const videoEl = document.getElementById("video");
 
 export const urlParams = new URLSearchParams(window.location.search);
