+++ fury/stream/servers/webrtc/www/js/constants.js
@@ -1,8 +1,9 @@
-export const millisecMouseMove = 80;
 export const urlServer = `${location.protocol}//${location.hostname}${
   location.port ? ":" + location.port : ""
 }/`;
-export const interactionInterval = 80;
 export const videoEl = document.getElementById("video");
 
 export const urlParams = new URLSearchParams(window.location.search);
+
+export const millisecMouseMove = 100;
+export const interactionInterval = 100;

[…]
