### 2021-06-04 4bafc90
this will call render after starting the streaming
This will avoid the first rendering to be just a blank screen inside
the browser
