### 2021-06-04 bfa4881
this allows to pass the time intervals as a parameter
For example,

http://localhost:8000/?interaction_addr=localhost:8080&weel_interval=10
or
http://localhost:8000/?mouse_interval=80
