### 2021-06-04 472a256
avoid calling unecessary render method
