### 2021-06-04 444076b
feat: resize panel when window resizes
