### 2021-06-04 677483b
avoids calling mouse movement with a mouse click
