### 2021-06-04 815afc7
Updated triangles options in SelectionManager
