### 2021-06-04 53389a5
Removed simple_quad_2d
