### 2021-06-05 9337788
fixes the mouse movement inside the browser
Before this commit, the first mouse movement always rotated the image
