+++ docs/examples/viz_streaming_interaction.py
@@ -16,7 +16,7 @@ if __name__ == '__main__':
     # 0 ms_stream means that the frame will be sent to the server
     # right after the rendering
     ms_interaction = 1
-    ms_stream = 0
+    ms_stream = 16
     # max number of interactions to be stored inside the queue
     max_queue_size = 1000
     ######################################################################
@@ -62,12 +62,12 @@ if __name__ == '__main__':
     # Otherwise, if ms it's equal to zero the shared memory it's updated in each 
     # render event
 
-    stream_interaction = FuryStreamInteraction(
-        showm, max_queue_size=max_queue_size)
     showm.initialize()
 
     stream = FuryStreamClient(
         showm, window_size,)
+    stream_interaction = FuryStreamInteraction(
+        showm, max_queue_size=max_queue_size, fury_client=stream)
     # linux
     # p = multiprocessing.Process(
     #     target=webrtc_server,
