### 2021-06-05 c7f2e54
fix: removed panel from tutorial
