### 2021-06-05 b220dfd
fix: fixed docstring w.r.t numpy docs
