### 2021-06-05 46a08e9
fix: fixed docstring w.r.t numpy docs
