### 2021-06-05 6e160e7
fix: fixed card resizing
