### 2021-06-05 135b2e3
allows resizing events and avoid memory issues.
With this commit you can change the resolution after the stream be initiated.
In addition, It will be possible to control the max window size through  max_window_size parameter

```python
stream = FuryStreamClient(
        showm, window_size, max_window_size=max_window_size)
```
