### 2021-06-05 989f085
creates an exception if the max_window_size parameter is invalid
