+++ docs/examples/viz_streaming_interaction_separeted.py
@@ -13,10 +13,11 @@ from fury.stream.client import FuryStreamClient, FuryStreamInteraction
 if __name__ == '__main__':
 
     window_size = (400, 400)
+    max_window_size = (700, 700)
     # 0 ms_stream means that the frame will be sent to the server
     # right after the rendering
     ms_interaction = 1
-    ms_stream = 16
+    ms_stream = 0 
     # max number of interactions to be stored inside the queue
     max_queue_size = 1000
     ##############################################################################
@@ -66,7 +67,7 @@ if __name__ == '__main__':
     showm.initialize()
 
     stream = FuryStreamClient(
-        showm, window_size,)
+        showm, window_size,max_window_size=max_window_size)
     # linux
     # p = multiprocessing.Process(
     #     target=webrtc_server,
