### 2021-06-06 53585f7
Updated default mouse wheel to 30ms (fixes infinite scrolling on safari)
