+++ docs/examples/viz_streaming_interaction_separeted.py
@@ -17,7 +17,7 @@ if __name__ == '__main__':
     # 0 ms_stream means that the frame will be sent to the server
     # right after the rendering
     ms_interaction = 1
-    ms_stream = 0 
+    ms_stream = 16
     # max number of interactions to be stored inside the queue
     max_queue_size = 1000
     ##############################################################################
@@ -92,5 +92,9 @@ if __name__ == '__main__':
     stream_interaction.start(ms=ms_interaction)
     stream.init(ms_stream,)
     showm.start()
+    p.kill()
+    p2.kill()
+    stream.cleanup()
+
     # open a browser using the following the url
     # http://localhost:8000?interaction_addr=localhost:8080
