### 2021-06-06 0d043a6
fix: use partial to provide the obj, iren to window propagate event
