### 2021-06-06 be59f1f
temporary print statement for window resize callback
