### 2021-06-06 864a859
the shared memory approach becomes optional
Grant's compatibility with python <3.8 versions
