+++ fury/stream/client.py
@@ -1,9 +1,14 @@
-from fury import window
-import os
 from vtk.util.numpy_support import vtk_to_numpy, numpy_to_vtk
 import vtk
 import multiprocessing
-from multiprocessing import shared_memory
+import sys
+if sys.version_info.minor >= 8:
+    from multiprocessing import shared_memory
+    PY_VERSION_8 = False
+else:
+    shared_memory = None
+    PY_VERSION_8 = False
+
 import numpy as np
 
 from fury.stream.tools import CircularQueue
@@ -17,7 +22,8 @@ class FuryStreamClient:
             max_window_size=None,
             buffer_count=2,
             image_buffers=None,
-            info_buffer=None):
+            info_buffer=None,
+            use_raw_array=None):
 
         '''
 
@@ -40,7 +46,16 @@ class FuryStreamClient:
         if self.max_size < window_size[0]*window_size[1]:
             raise ValueError('max_window_size must be greater than window_size')
 
-        if info_buffer is None or image_buffers is None:
+        # se use_raw_array não é setado escolhe a abordagem
+        # automaticamente
+        if use_raw_array is None:
+            use_raw_array = not PY_VERSION_8
+
+        if not PY_VERSION_8 and not use_raw_array:
+            raise ValueError("""
+                In order to use the SharedMemory approach
+                you should have to use python 3.8 or higher""")
+        if info_buffer is None:
             # 0 number of components
             # 1 id buffer
             # 2, 3, width first buffer, height first buffer
@@ -52,23 +67,36 @@ class FuryStreamClient:
             self.info_buffer = multiprocessing.RawArray(
                     'I', info_list
             )
+        if use_raw_array:
+            self.image_buffer_names = None
+            if image_buffers is None:
+                for _ in range(self.buffer_count):
+                    self.image_buffers.append(multiprocessing.RawArray(
+                        'B', np.random.randint(
+                            0, 255,
+                            size=max_window_size[0]*max_window_size[1]*3)
+                        .astype('uint8')))
+            else:
+                self.info_buffer = info_buffer
+                self.image_buffers = image_buffers
+                # do not work anymore with shared memory
+        else:
             for _ in range(self.buffer_count):
                 bufferSize = max_window_size[0]*max_window_size[1]*3
-                buffer = shared_memory.SharedMemory(create=True, size=bufferSize)
-                # buffer = multiprocessing.RawArray('B', max_window_size[0]*max_window_size[1]*3)
+                buffer = shared_memory.SharedMemory(
+                    create=True, size=bufferSize)
                 self.image_buffers.append(buffer)
-                self.image_reprs.append(np.ndarray(bufferSize, dtype=np.uint8, buffer=buffer.buf))
+                self.image_reprs.append(
+                    np.ndarray(
+                        bufferSize, dtype=np.uint8, buffer=buffer.buf))
                 self.image_buffer_names.append(buffer.name)
-        else:
-            self.info_buffer = info_buffer
-            self.image_buffers = image_buffers
-            #do not work anymore with shared memory
 
         self._id_timer = None
         self._id_observer = None
         self.sender = None
         self._in_request = False
         self.update = True
+        self.use_raw_array = use_raw_array
         # if broker_url is not None:
         #    pass
         # self.sender = imagezmq.ImageSender(connect_to=broker_url)
@@ -88,30 +116,47 @@ class FuryStreamClient:
                 self.window2image_filter.Modified()
                 vtk_image = window2image_filter.GetOutput()
                 vtk_array = vtk_image.GetPointData().GetScalars()
-                w, h, _ = vtk_image.GetDimensions()
                 num_components = vtk_array.GetNumberOfComponents()
[…]
