+++ fury/stream/servers/webrtc/server.py
@@ -3,13 +3,20 @@
 # import logging
 # logging.basicConfig(level=logging.ERROR)
 import time
+import sys
+if sys.version_info.minor >= 8:
+    from multiprocessing import shared_memory
+    PY_VERSION_8 = False
+else:
+    shared_memory = None
+    PY_VERSION_8 = False
 
 from aiohttp import web
 from av import VideoFrame
 from aiortc import VideoStreamTrack
 import pyximport; pyximport.install()
 from .FuryVideoFrame import FuryVideoFrame
-from multiprocessing import shared_memory
+
 
 import numpy as np
 
@@ -18,21 +25,27 @@ from fury.stream.tools import CircularQueue
 
 
 def webrtc_server(
-        image_buffer_names=None, info_buffer=None,
+        stream_client=None,
+        image_buffers=None,
+        image_buffer_names=None,
+        info_buffer=None,
         circular_queue=None,
         queue_head_tail_buffer=None,
         queue_buffers_list=None,
-        port=8000, host='localhost', flip_img=True,
+        port=8000, host='localhost',
+        flip_img=True,
         www_folder=None):
 
-    # if stream_client is not None:
-    #     image_buffers = stream_client.image_buffers
-    #     info_buffer = stream_client.info_buffer
-    
+    if stream_client is not None:
+        image_buffers = stream_client.image_buffers
+        info_buffer = stream_client.info_buffer
+
+    use_raw_array = image_buffer_names is None
+
     class RTCServer(VideoStreamTrack):
         def __init__(self,):
             super().__init__()
-        
+
             # starts with a random image
 
             image_info = np.frombuffer(
@@ -40,15 +53,17 @@ def webrtc_server(
             self.image = np.random.randint(
                  0, 255, (image_info[1], image_info[0], 3),
                  dtype='uint8')
-            self.frame = None
-            self.image_buffers = []
-            self.image_reprs = []
-            self.image_buffer_names = image_buffer_names
-            for buffer_name in self.image_buffer_names:
-                buffer = shared_memory.SharedMemory(buffer_name)
-                self.image_buffers.append(buffer)
-                self.image_reprs.append(np.ndarray(len(buffer.buf), dtype=np.uint8, buffer=buffer.buf))
-        
+
+            if not use_raw_array:
+                self.frame = None
+                self.image_buffers = []
+                self.image_reprs = []
+                self.image_buffer_names = image_buffer_names
+                for buffer_name in self.image_buffer_names:
+                    buffer = shared_memory.SharedMemory(buffer_name)
+                    self.image_buffers.append(buffer)
+                    self.image_reprs.append(np.ndarray(len(buffer.buf), dtype=np.uint8, buffer=buffer.buf))
+
         async def recv(self):
             pts, time_base = await self.next_timestamp()
             image_info = np.frombuffer(
@@ -57,22 +72,34 @@ def webrtc_server(
             width = image_info[2+buffer_index*2]
             height = image_info[2+buffer_index*2+1]
 
-            self.image = self.image_reprs[buffer_index]
+            if use_raw_array:
+                self.image = np.frombuffer(
+                        image_buffers[buffer_index],
+                        'uint8'
+                    )[0:width*height*3].reshape((width, height, 3))
 
-            # if flip_img:
-            #     self.image = np.flipud(self.image)
-            if(self.frame is None 
-                or self.frame.planes[0].width!=width
-                or self.frame.planes[0].height!=height):
-                print("creating frame with size: %d x %d"%(width,height))
-                self.frame = FuryVideoFrame(width, height, "rgb24")
-            # print(self.image[0:3])
-            self.frame.update_from_ndarray(self.image)
-            self.frame.pts = pts
-            self.frame.time_base = time_base
+                if flip_img:
+                    self.image = np.flipud(self.image)
[…]
