### 2021-06-06 aa86c8e
memory view optimization
