+++ docs/examples/viz_streaming_interaction.py
@@ -10,12 +10,16 @@ from fury.stream.client import FuryStreamClient, FuryStreamInteraction
 # note, if python version is equal or higher than 3.8
 # uses shared memory approach
 if __name__ == '__main__':
+    # se estiver usando python 3.8 e quiser testar o 
+    # raw_array sete use_raw_array=False. 
+    # se estiver usando pytho n3.8 por default escolhe shared memory
+    use_raw_array = None
     use_high_res = False
     if use_high_res:
         window_size = (1280, 720)
         max_window_size = (1920, 1080)
     else:
-        window_size = (500, 400)
+        window_size = (720, 500)
         max_window_size = (600, 600)
     # 0 ms_stream means that the frame will be sent to the server
     # right after the rendering
@@ -70,7 +74,8 @@ if __name__ == '__main__':
     showm.initialize()
 
     stream = FuryStreamClient(
-        showm, window_size, max_window_size=max_window_size)
+        showm, window_size, max_window_size=max_window_size,
+        use_raw_array=use_raw_array)
     stream_interaction = FuryStreamInteraction(
         showm, max_queue_size=max_queue_size, fury_client=stream)
     # linux
