### 2021-06-06 0a757bd
this will avoid unnecessary (m)alloc
