### 2021-06-06 8f2ea81
implements a multidimensional circular queue with just one buffer
