### 2021-06-07 941a56f
feat: resize panel according to precalculated ratio
