### 2021-06-07 f4046c8
Added gamma correction and isotropic specular.
