### 2021-06-08 1d1bf38
fix: cap the panel size to max size
