### 2021-06-08 98acdd8
Investigating if using a global variable can resolve conda 8.2.0 issue
