### 2021-06-09 aa1f214
fixes queue max_size and lock the head_tail buffer
This improves the behavior and performance of user interaction. Before
this commit some interactions on MacOs took a lot of computational time.
