+++ fury/stream/tools.py
@@ -1,5 +1,9 @@
 import numpy as np
 import multiprocessing
+import time
+import logging
+
+logging.basicConfig(level=logging.INFO)
 
 
 class MultiDimensionalBufferList:
@@ -47,36 +51,49 @@ class MultiDimensionalBuffer:
         if dimension is not None and max_size is None:
             max_size = np.frombuffer(
                 buffers_list, 'float64').shape[0]//dimension
+
         if buffers_list is None:
             buffers_list = multiprocessing.RawArray(
-                    'd', np.zeros(max_size*dimension, dtype='float64'))
+                    'd', np.zeros(dimension*(max_size+1), dtype='float64'))
 
+            #max_size = 10
         self._buffers = buffers_list
         self._memory_views = np.ctypeslib.as_array(self._buffers)
+        # self._memory_views = self._buffers
         self.dimension = dimension
         self.max_size = max_size
-
+        self.full_size = int(max_size*dimension)
+    
     @property
     def buffers(self):
         return self._buffers
 
     @buffers.setter
     def buffers(self, data):
-        self._memory_views[:] = data
+        if isinstance(data, (np.ndarray, np.generic) ):
+            if data.dtype == 'float64':
+                self._memory_views[:] = data
+
+    def get_start_end(self, idx):
+        dim = self.dimension
+        start = idx*dim #% self.full_size
+        end = dim*(idx+1) #% self.full_size
+        return start, end
 
     def __getitem__(self, idx):
-        dim = self.dimension
-        start = idx*dim
-        end = dim*(idx+1)
+        start, end = self.get_start_end(idx)
+        logging.info(f'dequeue start {int(time.time()*1000)}')
+        ts = time.time()*1000
         itens = np.frombuffer(self._buffers, 'float64')[start:end]
+        te = time.time()*1000
+        logging.info(f'dequeue frombuffer cost {te-ts:.2f}')
         return itens
 
     def __setitem__(self, idx, data):
-        dim = self.dimension
-        start = idx*dim
-        end = dim*(idx+1)
-        self._memory_views[start:end] = data
-
+        start, end = self.get_start_end(idx)
+        if isinstance(data, (np.ndarray, np.generic) ):
+            if data.dtype == 'float64':
+                self._memory_views[start:end] = data
 
 class CircularQueue:
 
@@ -86,20 +103,21 @@ class CircularQueue:
 
         buffers = MultiDimensionalBuffer(max_size, dimension, buffers_list)
         if head_tail_buffer is None or buffers_list is None:
-            head_tail_buffer = multiprocessing.RawArray(
-                'i', np.array([-1, -1], dtype='int32'),
+            head_tail_buffer = multiprocessing.Array(
+                'l', np.array([-1, -1], dtype='int64'),
             )
 
         self.dimension = buffers.dimension
         self.head_tail_buffer = head_tail_buffer
-        self.head_tail_memview = np.ctypeslib.as_array(self.head_tail_buffer)
+        # self.head_tail_memview = np.ctypeslib.as_array(self.head_tail_buffer)
+        self.head_tail_memview = self.head_tail_buffer
 
         self.max_size = buffers.max_size
         self.buffers = buffers
 
     @property
     def head(self):
-        return np.frombuffer(self.head_tail_buffer, 'int32')[0]
+        return np.frombuffer(self.head_tail_buffer.get_obj(), 'int64')[0]
 
     @head.setter
     def head(self, value):
@@ -107,45 +125,57 @@ class CircularQueue:
 
     @property
     def tail(self):
-        return np.frombuffer(self.head_tail_buffer, 'int32')[1]
[…]
