### 2021-06-09 1adf4ed
this obj it's not necessary anymore
This obj performs badly when compared with MultiDimensionalBuffer
