### 2021-06-09 e534c06
removed redundant code
