### 2021-06-10 5c02e87
added test for space_filling_model actor
