### 2021-06-10 a52d49a
this remove list_of_markers arg
