### 2021-06-10 991eaec
adds a simple test for actor.markers
