### 2021-06-10 e08150e
improves the actor.markers example
