### 2021-06-10 1c00d08
fix: removed flag from the UI constructor
