### 2021-06-10 d441eee
fix: determine is_url from filename
