### 2021-06-10 c795877
fix: fixed docstring accordingly
