### 2021-06-11 e1a1a80
tests: updated tests for Panel2D
