### 2021-06-11 554c496
This simplifies the example.
Suggestion made by @Garyfallidis
