### 2021-06-11 c76fe7d
This fixes some issues related with the docstrings
