### 2021-06-11 38dafe9
this puts all streaming examples into a tutorial folder.
