+++ docs/tutorials/04_stream/README.rst
@@ -1,0 +1,9 @@
+STREAM
+------
+
+These tutorials show:
+
+- How to create a simple streaming server and
+turn that available on internet using ngrok
+- How to create a straeming server with user interaction
+- How to create a server with off-screen rendering
