### 2021-06-12 e8905a2
after this cython becomes optional
