### 2021-06-12 cd86491
this removes the http endpoints for mouse interaction
