### 2021-06-12 7036d6a
buffer_info can now use the shared_memory from py3.8
After this commit  it will be possible to communicate Server and
FuryCleint running in different process. This feature is   useful for
jupyter extensions and for more agnostic solution.
