+++ fury/stream/client.py
@@ -1,10 +1,9 @@
-from vtk.util.numpy_support import vtk_to_numpy, numpy_to_vtk
 import vtk
 import multiprocessing
 import sys
 if sys.version_info.minor >= 8:
     from multiprocessing import shared_memory
-    PY_VERSION_8 = False
+    PY_VERSION_8 = True
 else:
     shared_memory = None
     PY_VERSION_8 = False
@@ -16,15 +15,13 @@ from fury.stream.tools import CircularQueue
 import logging
 import time
 
-logging.basicConfig(level=logging.ERROR)
-
 
 class FuryStreamClient:
     def __init__(
             self, showm,
             window_size=(200, 200),
             max_window_size=None,
-            use_raw_array=None,
+            use_raw_array=True,
             buffer_count=2,
             image_buffers=None,
             info_buffer=None):
@@ -40,6 +37,7 @@ class FuryStreamClient:
         self.window2image_filter.SetInput(self.showm.window)
         self.image_buffers = []
         self.image_buffer_names = []
+        self.info_buffer_name = None
         self.image_reprs = []
         self.buffer_count = buffer_count
         if max_window_size is None:
@@ -48,11 +46,6 @@ class FuryStreamClient:
         self.max_window_size = max_window_size
         if self.max_size < window_size[0]*window_size[1]:
             raise ValueError('max_window_size must be greater than window_size')
-
-        # se use_raw_array não é setado escolhe a abordagem
-        # automaticamente
-        if use_raw_array is None:
-            use_raw_array = not PY_VERSION_8
 
         if not PY_VERSION_8 and not use_raw_array:
             raise ValueError("""
@@ -67,13 +60,26 @@ class FuryStreamClient:
             for _ in range(self.buffer_count):
                 info_list += [self.max_window_size[0]]
                 info_list += [self.max_window_size[1]]
-            self.info_buffer = multiprocessing.RawArray(
-                    'I', info_list
+            info_list = np.array(
+               info_list, dtype='uint64'
             )
+            if use_raw_array:
+                self.info_buffer = multiprocessing.RawArray(
+                        'I', info_list
+                )
+                self.info_buffer_repr = np.ctypeslib.as_array(self.info_buffer)
+            else:
+                self.info_buffer = shared_memory.SharedMemory(
+                    create=True, size=info_list.nbytes)
+                self.info_buffer_repr = np.ndarray(
+                        info_list.shape[0],
+                        dtype='uint64', buffer=self.info_buffer.buf)
+                self.info_buffer_name = self.info_buffer.name
+
         if use_raw_array:
             self.image_buffer_names = None
             if image_buffers is None:
-                self.image_memory_views = []
+                self.image_reprs = []
                 for _ in range(self.buffer_count):
                     buffer = multiprocessing.RawArray(
                         'B', np.random.randint(
@@ -81,13 +87,14 @@ class FuryStreamClient:
                             size=max_window_size[0]*max_window_size[1]*3)
                         .astype('uint8'))
                     self.image_buffers.append(buffer)
-                    self.image_memory_views.append(
+                    self.image_reprs.append(
                         np.ctypeslib.as_array(buffer))
             else:
                 self.info_buffer = info_buffer
                 self.image_buffers = image_buffers
                 # do not work anymore with shared memory
         else:
+            logging.info('Using the shared memory approach')
             for _ in range(self.buffer_count):
                 bufferSize = max_window_size[0]*max_window_size[1]*3
                 buffer = shared_memory.SharedMemory(
@@ -97,7 +104,6 @@ class FuryStreamClient:
                     np.ndarray(
                         bufferSize, dtype=np.uint8, buffer=buffer.buf))
[…]
