### 2021-06-12 c3394b4
user interaction compatibility with shared_memory from py3.8
