+++ fury/stream/tools.py
@@ -2,26 +2,65 @@ import numpy as np
 import multiprocessing
 import time
 import logging
-
-logging.basicConfig(level=logging.ERROR)
+import sys
+if sys.version_info.minor >= 8:
+    from multiprocessing import shared_memory
+    PY_VERSION_8 = True
+else:
+    shared_memory = None
+    PY_VERSION_8 = False
 
 
 class MultiDimensionalBuffer:
-    def __init__(self, max_size=None, dimension=7, buffers_list=None):
+    def __init__(
+            self, max_size=None, dimension=8, buffers_list=None,
+            buffer_name=None,
+            use_raw_array=True):
+        
+        use_raw_array = use_raw_array and buffer_name is None
+        if not PY_VERSION_8 and not use_raw_array:
+            raise ValueError("""
+                In order to use the SharedMemory approach
+                you should have to use python 3.8 or higher""")
 
-        if dimension is not None and max_size is None:
-            max_size = np.frombuffer(
-                buffers_list, 'float64').shape[0]//dimension
+        if buffers_list is None and buffer_name is None:
+            buffer_arr = np.zeros(dimension*(max_size+1), dtype='float64')
+            if use_raw_array:
+                buffers_list = multiprocessing.RawArray(
+                        'd', buffer_arr)
+                self._buffers = buffers_list
+                self._buffer_repr = np.ctypeslib.as_array(self._buffers)
+            else:
+                buffers_list = shared_memory.SharedMemory(
+                    create=True, size=buffer_arr.nbytes)
+                self._buffer_repr = np.ndarray(
+                        buffer_arr.shape[0],
+                        dtype='float64', buffer=buffers_list.buf)
+                self._buffers = buffers_list
+                buffer_name = buffers_list.name
+        else:
+            if buffer_name is None:
+                max_size = int(len(buffers_list)//dimension)
+                max_size -= 1
+                
+                self._buffers = buffers_list
+                self._buffer_repr = np.ctypeslib.as_array(self._buffers)
+            else:
+                buffer = shared_memory.SharedMemory(buffer_name)
+                # 8 represents 8 bytes of float64
+                max_size = int(len(buffer.buf)//dimension/8)
+                max_size -= 1
 
-        if buffers_list is None:
-            buffers_list = multiprocessing.RawArray(
-                    'd', np.zeros(dimension*(max_size+1), dtype='float64'))
+                self._buffers = buffer
+                self._buffer_repr = np.ndarray(
+                        len(buffer.buf)//8,
+                        dtype='float64', buffer=buffer.buf)
+                self._buffers = buffer
 
-        self._buffers = buffers_list
-        self._memory_views = np.ctypeslib.as_array(self._buffers)
+        self.buffer_name = buffer_name
         self.dimension = dimension
         self.max_size = max_size
-        self.full_size = int(max_size*dimension)
+        self.use_raw_array = use_raw_array
 
     @property
     def buffers(self):
@@ -31,7 +70,7 @@ class MultiDimensionalBuffer:
     def buffers(self, data):
         if isinstance(data, (np.ndarray, np.generic)):
             if data.dtype == 'float64':
-                self._memory_views[:] = data
+                self._buffer_repr[:] = data
 
     def get_start_end(self, idx):
         dim = self.dimension
@@ -43,7 +82,10 @@ class MultiDimensionalBuffer:
         start, end = self.get_start_end(idx)
         logging.info(f'dequeue start {int(time.time()*1000)}')
         ts = time.time()*1000
-        itens = np.frombuffer(self._buffers, 'float64')[start:end]
+        if self.use_raw_array:
+            itens = np.frombuffer(self._buffers, 'float64')[start:end]
+        else:
[…]
