### 2021-06-12 95e0f3c
fix: exception handeling for incorrect URL
