### 2021-06-12 a049435
tests: updated tests for io module
