### 2021-06-13 3b3fc3c
fixes error in documentaion
With the interactive mode the documentation can not be generated
