### 2021-06-13 c9180b7
this will close the websocket conections before the shutdown
