### 2021-06-14 214df8f
this implements a simple setTimeInterval-like obj
It's not possible to use the vtk CreateRepeatingTimer if the Interactor Obj
has not started (self.showm.iren). This new obj can be used in order
to call generic events from vtkRenderWindowInteractor. For example
SetEventInformation, MouseMoveEvent etc
