### 2021-06-14 fc241ab
allows a StreamClient and StreamInteraction instances with a  unitialized iren obj
We need this for jupyter/vscode extensions
