+++ fury/stream/client.py
@@ -10,7 +10,7 @@ else:
 
 import numpy as np
 
-from fury.stream.tools import CircularQueue
+from fury.stream.tools import CircularQueue, IntervalTimer
 
 import logging
 import time
@@ -24,14 +24,17 @@ class FuryStreamClient:
             use_raw_array=True,
             buffer_count=2,
             image_buffers=None,
-            info_buffer=None):
-
+            info_buffer=None,
+            whithout_iren_start=False,
+    ):
         '''
 
         Parameters
         ----------
             showm: fury showm manager
         '''
+
+        self._whithout_iren_start = whithout_iren_start
         self.showm = showm
         self.window2image_filter = vtk.vtkWindowToImageFilter()
         self.window2image_filter.SetInput(self.showm.window)
@@ -168,10 +171,21 @@ class FuryStreamClient:
                 self._in_request = False
 
         if ms > 0:
-            id_timer = self.showm.add_timer_callback(
-                True, ms, callback)
-            self._id_timer = id_timer
+            if self._whithout_iren_start:
+                self._interval_timer = IntervalTimer(
+                    ms/1000,
+                    callback,
+                    None,
+                    None)
+            else:
+                id_timer = self.showm.add_timer_callback(
+                    True, ms, callback)
+                # self.showm.window.AddObserver("TimerEvent", callback)
+                # id_timer = self.showm.window.CreateRepeatingTimer(ms)
+                self._id_timer = id_timer
+          
         else:
+            # id_observer = self.showm.iren.AddObserver(
             id_observer = self.showm.iren.AddObserver(
                 'RenderEvent', callback)
             self._id_observer = id_observer
@@ -181,8 +195,8 @@ class FuryStreamClient:
     def stop(self):
         logging.info('stop timers')
         if self._id_timer is not None:
-            self.showm.destroy_timer(self._id_timer)
-
+            # self.showm.destroy_timer(self._id_timer)
+            self.showm.iren.DestroyTimer(self._id_timer)
         if self._id_observer is not None:
             self.showm.iren.RemoveObserver(self._id_observer)
 
@@ -197,12 +211,80 @@ class FuryStreamClient:
                 buffer.unlink()
 
 
+def interaction_callback(
+        circular_queue, showm, iren, render_after=False):
+    ts = time.time()*1000
+    data = circular_queue.dequeue()
+    if data is not None:
+        event_id = data[0]
+        user_timestamp = data[6]
+        logging.info(
+            'Interaction: time to dequeue ' +
+            f'{ts-user_timestamp:.2f} ms')
+
+        ts = time.time()*1000
+        newX = int(showm.size[0]*data[2])
+        newY = int(showm.size[1]*data[3])
+        ctrl_key = int(data[4])
+        shift_key = int(data[5])
+        newY = showm.size[1] - newY
+        if event_id == 1:
+            zoomFactor = 1.0 - data[1] / 1000.0
+            camera = showm.scene.GetActiveCamera()
+            fp = camera.GetFocalPoint()
+            pos = camera.GetPosition()
+            delta = [fp[i] - pos[i] for i in range(3)]
+            camera.Zoom(zoomFactor)
+
+            pos2 = camera.GetPosition()
+            camera.SetFocalPoint(
+                [pos2[i] + delta[i] for i in range(3)])
+            if data[1] < 0:
+                iren.MouseWheelForwardEvent()
+            else:
+                iren.MouseWheelBackwardEvent()
+
+            showm.window.Modified()
+
+        elif event_id == 2:
+            iren.SetEventInformation(
+                newX, newY, ctrl_key, shift_key, chr(0), 0, None)
+
[…]
