### 2021-06-14 6cf7059
implements a simple widget to be used in jupyter
Note: we need to solve the issues related with release the memory.
Calling cleanup inside a jupyter session restarts the current session
