### 2021-06-14 ac4da92
cpkr scheme for Manganese
