+++ fury/actor.py
@@ -2612,6 +2612,7 @@ def space_filling_model(atom_coords, elem_sym_list,
             'CL': [31/255, 240/255, 31/255, 1.75],
             'K': [143/255, 64/255, 212/255, 2.75],
             'CA': [61/255, 255/255, 0, 2.31],
+            'MN': [156/255, 122/255, 199/255, 1.19],
             'FE': [224/255, 102/255, 51/255, 1.16],
             'CO': [240/255, 144/255, 160/255, 2.4],
             'NI': [80/255, 208/255, 80/255, 1.63],
