### 2021-06-14 1776ce2
added example for space filling model using biopython for parsing
