+++ fury/stream/README.md
@@ -1,0 +1,12 @@
+
+## TODO
+- [] Create a buffer manager object
+  - For using DRY principle when multiple encondings became availbale
+- [] Widget object compatibility with python versions lower than 3.8
+  - memory socket communication  
+- [] MJPEG econding
+- [] Fix memory release issues with the shared_manager in python >=3.8 inside the Widget object
+- [] Create a tutorial about shared_memory and widget object
+- [] update the examples with multiprocessing to be compatible with MacOs
+  - setting multiprocessing.set_start_method('spawn')
+- [] Implement all the interactions: double-click, touch etc
