### 2021-06-14 6e0c1c6
refactoring - DRY - a image buffer manager object
 - For using DRY principle when multiple encondings became availbale
