+++ fury/stream/servers/webrtc/server.py
@@ -26,6 +26,124 @@ except ImportError:
     CYTHON_AVAILABLE = False
 
 
+class ImageBufferManager:
+    def __init__(
+            self, use_raw_array=True,
+            info_buffer=None, image_buffers=None,
+            info_buffer_name=None, image_buffer_names=None):
+        super().__init__()
+
+        self.use_raw_array = use_raw_array
+        self.image = None
+        if not use_raw_array:
+            self.info_buffer = shared_memory.SharedMemory(info_buffer_name)
+            self.info_buffer_repr = np.ndarray(
+                    6,
+                    dtype='uint64',
+                    buffer=self.info_buffer.buf)
+            self.image_buffers = []
+            self.image_reprs = []
+            self.image_buffer_names = image_buffer_names
+            for buffer_name in self.image_buffer_names:
+                buffer = shared_memory.SharedMemory(buffer_name)
+                self.image_buffers.append(buffer)
+                self.image_reprs.append(np.ndarray(
+                    len(buffer.buf),
+                    dtype=np.uint8,
+                    buffer=buffer.buf))
+        else:
+            self.info_buffer = np.frombuffer(
+                info_buffer, 'uint64')
+            self.info_buffer_repr = np.ctypeslib.as_array(
+                self.info_buffer)
+            self.image_buffers = image_buffers
+
+    def get_infos(self):
+       
+        if self.use_raw_array:
+            self.image_info = np.frombuffer(
+                self.info_buffer, 'uint32')
+        else:
+            self.image_info = self.info_buffer_repr
+
+        buffer_index = int(self.image_info[1])
+
+        self.width = int(self.image_info[2+buffer_index*2])
+        self.height = int(self.image_info[2+buffer_index*2+1])
+
+        if self.use_raw_array:
+            self.image = self.image_buffers[buffer_index]
+        else:
+            self.image = self.image_reprs[buffer_index]
+
+        return self.width, self.height, self.image
+
+    def cleanup(self):
+        logging.info("Freeing buffer")
+        if not self.use_raw_array:
+            for buffer in self.image_buffers:
+                #buffer.close()
+                ...
+            #self.info_buffer.close()
+
+class RTCServer(VideoStreamTrack):
+    def __init__(
+            self, image_buffer_manager, use_raw_array=True,
+            info_buffer=None, image_buffers=None,
+            info_buffer_name=None, image_buffer_names=None):
+        super().__init__()
+
+        self.frame = None
+        self.use_raw_array = use_raw_array
+        self.buffer_manager = image_buffer_manager
+
+    async def recv(self):
+        pts, time_base = await self.next_timestamp()
+
+        width, height, image = self.buffer_manager.get_infos()
+        
+        if self.frame is None \
+            or self.frame.planes[0].width != width \
+                or self.frame.planes[0].height != height:
+            if CYTHON_AVAILABLE:
+                self.frame = FuryVideoFrame(width, height, "rgb24")
+            # else:
+            #    self.frame = VideoFrame(width, height, "rgb24")
+        self.image = image
+
+        if not CYTHON_AVAILABLE:
+            # if the buffer it's already flipped
+            # self.frame.planes[0].update(self.image)
+            self.image = np.frombuffer(
+                        self.image,
+                        'uint8'
+                    )[0:width*height*3].reshape((height, width, 3))
+            self.image = np.flipud(self.image)
+    
+            self.frame = VideoFrame.from_ndarray(self.image)
+        else:
+            self.frame.update_from_buffer(self.image)
+
+        self.frame.pts = pts
+        self.frame.time_base = time_base
+
+        return self.frame
+
+    def release(self):
+        logging.info("Release Server")
+        try:
+            if not (self.stream is None):
+                self.stream.release()
+                self.stream = None
[…]
