### 2021-06-14 b562e3b
this allows mjpeg encoding
To use or disable that you need to control the webrtc parameter
provides_mjpeg=True(False).

In the URL you should pass the encoding parameter

http://localhost:8000?encoding=mjpeg
