### 2021-06-14 9490eb2
refactor!: due the new behavior which allows to deal with multiple encodings
- Widget Object has mjpeg as a default encoding
