+++ fury/stream/server/__init__.py
@@ -1,0 +1,3 @@
+from fury.stream.server.server import web_server
+
+__all__ = ['web_server']
