+++ fury/stream/README.md
@@ -5,6 +5,7 @@
 - [] Widget object compatibility with python versions lower than 3.8
   - memory socket communication  
 - [x] MJPEG econding
+- [x] Allows the user to set the encoding inside the widget object
 - [] Fix memory release issues with the shared_manager in python >=3.8 inside the Widget object
 - [] Create a tutorial about shared_memory and widget object
 - [] update the examples with multiprocessing to be compatible with MacOs
