### 2021-06-14 67084f7
feat: allows to disable the webrtc and use just one type of encoding
