### 2021-06-14 9ebd67b
doc: improvements, adds a widget example (whithout multiprocessing)
