### 2021-06-14 ee53f3e
fix: pep8 issues and doc img generation
