### 2021-06-14 f1dcb78
fix stop the timer interval event inside the FuryStreamClient
