### 2021-06-15 72b0dca
try to avoid the flaws inside of multiprocessing lib from python
related with https://bugs.python.org/issue38119#msg388287
, https://bugs.python.org/issue39959 and
https://github.com/python/cpython/pull/15989/files
