### 2021-06-15 0ece7bb
fixes some typos and shader incompatibility with MacOs
Fix suggested by @filipinascimento
