### 2021-06-15 5aade11
avoids the pickle error in MacOs
