### 2021-06-16 3885e6b
module to store atomic info
