+++ fury/molecular.py
@@ -1,0 +1,21 @@
+# cpk coloring scheme for coloring the atoms
+# kindly note that the fourth entry in each element's list corresponds to
+# to the its Van der Waals radius (in angstroms)
+cpkr = {'H': [255/255, 255/255, 255/255, 1.2],
+        'C': [144/255, 144/255, 144/255, 1.7],
+        'N': [48/255, 80/255, 248/255, 1.55],
+        'O': [255/255, 13/255, 13/255, 1.52],
+        'NA': [171/255, 92/255, 242/255, 2.27],
+        'MG': [138/255, 255/255, 0, 1.73],
+        'P': [255/255, 128/255, 0, 1.8],
+        'S': [255/255, 255/255, 48/255, 1.8],
+        'CL': [31/255, 240/255, 31/255, 1.75],
+        'K': [143/255, 64/255, 212/255, 2.75],
+        'CA': [61/255, 255/255, 0, 2.31],
+        'MN': [156/255, 122/255, 199/255, 1.19],
+        'FE': [224/255, 102/255, 51/255, 1.16],
+        'CO': [240/255, 144/255, 160/255, 2.4],
+        'NI': [80/255, 208/255, 80/255, 1.63],
+        'ZN': [125/255, 128/255, 176/255, 1.39],
+        'CD': [255/255, 217/255, 143/255, 1.58]
+        }
