### 2021-06-16 24a5434
removed the atomic info (stored it in molecular)
