+++ fury/actor.py
@@ -19,6 +19,7 @@ from fury.utils import (lines_to_vtk_polydata, set_input, apply_affine,
 from fury.io import load_image
 from fury.actors.odf_slicer import OdfSlicerActor
 import fury.primitive as fp
+from fury import molecular
 
 
 def slicer(data, affine=None, value_range=None, opacity=1.,
@@ -2571,7 +2572,7 @@ def sdf(centers, directions=(1, 0, 0), colors=(1, 0, 0), primitives='torus',
     return box_actor
 
 
-def space_filling_model(atom_coords, elem_sym_list,
+def molecular_sf(atom_coords, elem_sym_list,
                         return_unique_elements=False):
     """Generate a space-filling molecular model from an array of atomic
     coordinates and an array of corresponding element names
@@ -2601,24 +2602,7 @@ def space_filling_model(atom_coords, elem_sym_list,
     # cpk coloring scheme for coloring the atoms
     # kindly note that the fourth entry in each element's list corresponds to
     # to the its Van der Waals radius (in angstroms)
-    cpkr = {'H': [255/255, 255/255, 255/255, 1.2],
-            'C': [144/255, 144/255, 144/255, 1.7],
-            'N': [48/255, 80/255, 248/255, 1.55],
-            'O': [255/255, 13/255, 13/255, 1.52],
-            'NA': [171/255, 92/255, 242/255, 2.27],
-            'MG': [138/255, 255/255, 0, 1.73],
-            'P': [255/255, 128/255, 0, 1.8],
-            'S': [255/255, 255/255, 48/255, 1.8],
-            'CL': [31/255, 240/255, 31/255, 1.75],
-            'K': [143/255, 64/255, 212/255, 2.75],
-            'CA': [61/255, 255/255, 0, 2.31],
-            'MN': [156/255, 122/255, 199/255, 1.19],
-            'FE': [224/255, 102/255, 51/255, 1.16],
-            'CO': [240/255, 144/255, 160/255, 2.4],
-            'NI': [80/255, 208/255, 80/255, 1.63],
-            'ZN': [125/255, 128/255, 176/255, 1.39],
-            'CD': [255/255, 217/255, 143/255, 1.58]
-            }
+    cpkr = molecular.cpkr
 
     # store names of unique elements rendered in the model and their colors
     if return_unique_elements:
