+++ docs/examples/viz_molecular_sf_2.py
@@ -70,7 +70,7 @@ scene.set_camera(position=(20, 20, 0), focal_point=(0, 0, 0),
 axes_actor = actor.axes()
 scene.add(axes_actor)
 sf_model, elements = actor.molecular_sf(atom_coords, elem_sym_list,
-                                               return_unique_elements=True)
+                                        return_unique_elements=True)
 scene.add(sf_model)
 
 ###############################################################################
