### 2021-06-16 1f440f5
tests: added tests for wrap_overflow
