### 2021-06-16 33d2648
fix: fixed docstring, exception handling
