+++ fury/tests/test_ui.py
@@ -1630,6 +1630,6 @@ def test_wrap_overflow():
     ui.wrap_overflow(text, rectangle.size[0])
     npt.assert_equal("A very very\n long mess\nage to cli\np text ove\nrflow",
                      text.message)
-    
+
     with npt.assert_raises(ValueError):
         ui.wrap_overflow(text, 0)
