### 2021-06-16 88ecfb0
raises an exception in Widget class if python version is lower than 3.8
Due the shared_memory requirement
