### 2021-06-16 41b3c0c
fix: test it the time_invertval is running before removing it
This will avoid errors due to misuse of StreamInteracion class
