### 2021-06-16 4154785
feature: a simple lock/unlock system for python shared_memory
Maybe this can solve the MacOs issues
