### 2021-06-16 27a3ccb
fix:monkey-patch solution for the unlink  python-core issue
Related with
 - "resource tracker destroys shared memory segments when other processes should still have valid access"
   https://bugs.python.org/issue38119
 - python pr Fix shmem resource tracking
   https://github.com/python/cpython/pull/21516

The monkey-patch solution it's due Álvaro Justen (@turicas)
(https://bugs.python.org/msg388287)
