+++ fury/stream/tools.py
@@ -7,10 +7,34 @@ from threading import Timer
 import sys
 if sys.version_info.minor >= 8:
     from multiprocessing import shared_memory
+    from multiprocessing import resource_tracker
+
     PY_VERSION_8 = True
 else:
     shared_memory = None
     PY_VERSION_8 = False
+
+
+def remove_shm_from_resource_tracker():
+    """Monkey-patch multiprocessing.resource_tracker so SharedMemory won't be tracked
+
+    More details at: https://bugs.python.org/issue38119
+    """
+
+    def fix_register(name, rtype):
+        if rtype == "shared_memory":
+            return
+        return resource_tracker._resource_tracker.register(self, name, rtype)
+    resource_tracker.register = fix_register
+
+    def fix_unregister(name, rtype):
+        if rtype == "shared_memory":
+            return
+        return resource_tracker._resource_tracker.unregister(self, name, rtype)
+    resource_tracker.unregister = fix_unregister
+
+    if "shared_memory" in resource_tracker._CLEANUP_FUNCS:
+        del resource_tracker._CLEANUP_FUNCS["shared_memory"]
 
 
 class MultiDimensionalBuffer:
