+++ docs/examples/viz_fine_tuning_gl_context.py
@@ -31,7 +31,7 @@ centers = 1*np.array([
     [1, 0, 0]
 ])
 centers_no_depth_test = centers - np.array([[0, -1, 0]])
-centers_additive = centers_no_depth_test - np.array([[0, -1, 0]])
+centers_normal_blending = centers_no_depth_test - np.array([[0, -1, 0]])
 colors = np.array([
     [1, 0, 0],
     [0, 1, 0],
@@ -39,24 +39,24 @@ colors = np.array([
 ])
 
 actors = actor.sphere(
-    centers, opacity=.5, radii=.4, colors=colors)
+    centers, opacity=.8, radii=.4, colors=colors)
 actors_no_depth_test = actor.sphere(
-    centers_no_depth_test, opacity=.5, radii=.4, colors=colors)
-actor_add = actor.sphere(
-    centers_additive, opacity=1, radii=.4, colors=colors)
+    centers_no_depth_test, opacity=.8, radii=.4, colors=colors)
+actor_normal_blending = actor.sphere(
+    centers_normal_blending, opacity=.8, radii=.4, colors=colors)
 
 renderer = window.Scene()
 scene = window.Scene()
 
 showm = window.ShowManager(scene,
                            size=(900, 768), reset_camera=False,
-                           order_transparent=True)
+                           order_transparent=False)
 
 ###############################################################################
 # All actors must be added  in the scene
 
 scene.add(actors)
-scene.add(actor_add)
+scene.add(actor_normal_blending)
 scene.add(actors_no_depth_test)
 ###############################################################################
 # Now, we will enter in the topic of this example. First, we need to create
@@ -68,11 +68,11 @@ scene.add(actors_no_depth_test)
 
 shader_apply_effects(
     showm, actors,
-    effect=window.gl_set_normal_blending)
+    effects=[window.gl_enable_blend, window.gl_enable_depth])
 
-id_add = shader_apply_effects(
-    showm, actor_add,
-    effect=window.gl_set_additive_blending)
+id_observer = shader_apply_effects(
+    showm, actor_normal_blending,
+    effect=window.gl_set_normal_blending)
 
 ###############################################################################
 # It's also possible to pass a list of effects. The final opengl state it'll
@@ -81,7 +81,7 @@ shader_apply_effects(
     showm, actors_no_depth_test,
     effects=[
         window.gl_reset_blend, window.gl_disable_blend,
-        window.gl_disable_depth])
+        window.gl_disable_depth, window.gl_set_additive_blending])
 
 
 ###############################################################################
@@ -92,14 +92,15 @@ showm.initialize()
 counter = itertools.count()
 
 # After one hundred of steps we will remove the additive blending effect
-# from actor_add object
+# from actor_normal_blending object
 
 
 def timer_callback(obj, event):
     cnt = next(counter)
     showm.render()
+    showm.scene.GetActiveCamera().Azimuth(1)
     if cnt == 100:
-        actor_add.GetMapper().RemoveObserver(id_add)
+        actor_normal_blending.GetMapper().RemoveObserver(id_observer)
     if cnt == 1000:
         showm.exit()
 
