+++ docs/examples/viz_multiple_bonds.py
@@ -1,0 +1,183 @@
+import numpy as np
+from fury import window, actor
+
+# # ethane (single bonds)
+# atom_coords = np.array([[0.5723949486E+01, 0.5974463617E+01, 0.5898320525E+01],
+#                         [0.6840181327E+01, 0.6678078649E+01, 0.5159998484E+01],
+#                         [0.4774278044E+01, 0.6499436628E+01, 0.5782310182E+01],
+#                         [0.5576295333E+01, 0.4957554302E+01, 0.5530844713E+01],
+#                         [0.5926818174E+01, 0.5907771848E+01, 0.6968386044E+01],
+#                         [0.6985130929E+01, 0.7695511362E+01, 0.5526416671E+01],
+#                         [0.7788135127E+01, 0.6150201159E+01, 0.5277430519E+01],
+#                         [0.6632858893E+01, 0.6740709254E+01, 0.4090898288E+01]], dtype=float)
+# elem_names = np.array(['C', 'C', 'H', 'H', 'H', 'H', 'H', 'H'])
+# bond_type = np.array([1 ,1 ,1 ,1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1]) # ethane
+
+# ethene (double bonds)
+atom_coords = np.array([[0.5449769880E+01, 0.5680940296E+01, 0.5519555369E+01],
+                        [0.6346918574E+01, 0.6272796762E+01, 0.6280955432E+01],
+                        [0.4603843481E+01, 0.6218164558E+01, 0.5109232482E+01],
+                        [0.5522586159E+01, 0.4630394155E+01, 0.5265590478E+01],
+                        [0.6275632424E+01, 0.7323831630E+01, 0.6534741329E+01],
+                        [0.7193811558E+01, 0.5736436394E+01, 0.6691459388E+01]], dtype=float)
+elem_names = np.array(['C', 'C', 'H', 'H', 'H', 'H'])
+bond_type = np.array([2, 1, 1, 2, 1, 1, 1, 1, 1, 1]) # ethene
+
+# # ethene (triple bonds)
+# atom_coords = np.array([[0.5899518696E+01, 0.5868718390E+01, 0.5737443048E+01],
+#                         [0.6573681090E+01, 0.6576391430E+01, 0.6424519094E+01],
+#                         [0.5300877605E+01, 0.5237561906E+01, 0.5127884913E+01],
+#                         [0.7173348068E+01, 0.7204870233E+01, 0.7035206023E+01]], dtype=float)
+# elem_names = np.array(['C', 'C', 'H', 'H'])
+# bond_type = np.array([3, 1, 3, 1, 1, 1]) # ethyne
+
+
+atomic_radius = {'H':0.38, 'C':0.77}
+error_factor = 0.05
+atomic_radius = {k:v + error_factor for k,v in atomic_radius.items()}
+
+atom_index = np.arange(0, len(elem_names))
+atoms = elem_names
+
+atoms_rad = [atomic_radius[x] for x in atoms]
+# print(atoms_rad, atom_index)
+
+i_atom = np.copy(atom_index)
+p = np.copy(atom_coords)
+p_compare = p
+r = atoms_rad
+r_compare = r
+
+source_row = np.arange(len(atoms))
+max_atoms = len(atoms)
+
+bonds = np.zeros((len(atoms)+1, max_atoms+1), dtype=np.int8)
+bond_dists = np.zeros((len(atoms)+1, max_atoms+1), dtype=np.float32)
+
+# Algorithm to generate bonding
+for i in range(max_atoms-1):
+    p_compare = np.roll(p_compare, -1, axis=0)
[…]
