+++ docs/examples/viz_multiple_bonds.py
@@ -1,17 +1,167 @@
 import numpy as np
 from fury import window, actor
 
-# # ethane (single bonds)
-# atom_coords = np.array([[0.5723949486E+01, 0.5974463617E+01, 0.5898320525E+01],
-#                         [0.6840181327E+01, 0.6678078649E+01, 0.5159998484E+01],
-#                         [0.4774278044E+01, 0.6499436628E+01, 0.5782310182E+01],
-#                         [0.5576295333E+01, 0.4957554302E+01, 0.5530844713E+01],
-#                         [0.5926818174E+01, 0.5907771848E+01, 0.6968386044E+01],
-#                         [0.6985130929E+01, 0.7695511362E+01, 0.5526416671E+01],
-#                         [0.7788135127E+01, 0.6150201159E+01, 0.5277430519E+01],
-#                         [0.6632858893E+01, 0.6740709254E+01, 0.4090898288E+01]], dtype=float)
-# elem_names = np.array(['C', 'C', 'H', 'H', 'H', 'H', 'H', 'H'])
-# bond_type = np.array([1 ,1 ,1 ,1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1]) # ethane
+def ball_and_stick(atom_coords, elem_names, bond_type):
+    atomic_radius = {'H':0.38, 'C':0.77}
+    error_factor = 0.05
+    atomic_radius = {k:v + error_factor for k,v in atomic_radius.items()}
+
+    atom_index = np.arange(0, len(elem_names))
+    atoms = elem_names
+
+    atoms_rad = [atomic_radius[x] for x in atoms]
+    # print(atoms_rad, atom_index)
+
+    i_atom = np.copy(atom_index)
+    p = np.copy(atom_coords)
+    p_compare = p
+    r = atoms_rad
+    r_compare = r
+
+    source_row = np.arange(len(atoms))
+    max_atoms = len(atoms)
+
+    bonds = np.zeros((len(atoms)+1, max_atoms+1), dtype=np.int8)
+    bond_dists = np.zeros((len(atoms)+1, max_atoms+1), dtype=np.float32)
+
+    # Algorithm to generate bonding
+    for i in range(max_atoms-1):
+        p_compare = np.roll(p_compare, -1, axis=0)
+        #m_compare = np.roll(m_compare, -1, axis=0)
+        r_compare = np.roll(r_compare, -1, axis=0)
+        mask = 0
+        dists = np.linalg.norm(p - p_compare, axis=1) #* mask
+        r_bond = r + r_compare
+        bond = np.where(np.logical_and(dists > 0.0001, dists < r_bond), 1, 0)
+
+        source_row = source_row
+        target_row = source_row + i + 1 # Will be out of bounds of bonds array for some values of i
+        target_row = np.where(np.logical_or(target_row > len(atoms), mask==1), len(atoms), target_row) #If invalid target, write to dummy row
+
+        source_atom = i_atom
+        target_atom = i_atom + i + 1 # Will be out of bounds of bonds array for some values of i
+        target_atom = np.where(np.logical_or(target_atom > max_atoms, mask==1), max_atoms, target_atom) #If invalid target, write to dummy col
+
+        bonds[(source_row, target_atom)] = bond
+        bonds[(target_row, source_atom)] = bond
+        bond_dists[(source_row, target_atom)] = dists
+        bond_dists[(target_row, source_atom)] = dists
+
+    bonds = np.delete(bonds, axis=0, obj=-1) #Delete dummy row
+    bonds = np.delete(bonds, axis=1, obj=-1) #Delete dummy col
+    bond_dists = np.delete(bond_dists, axis=0, obj=-1) #Delete dummy row
[…]
