+++ docs/examples/viz_multiple_bonds.py
@@ -1,16 +1,69 @@
+"""
+====================================
+Multiple bonds visualization
+====================================
+
+This example demonstrate how to create ball and stick model for different
+molecules from their atomic coordinates. It also shows how to visualize
+double and triple bonds
+"""
+
+###############################################################################
+# First, we import some useful modules and methods.
 import numpy as np
 from fury import window, actor
 
+###############################################################################
+
+# Offsets for double and triple bonds
+def offsets(atom_coords, c1_coord, c2_coord, bond_type):
+    # Initial assumption that the molecule is planar in nature
+    planar = 1
+    # finding equation of plane on which the atoms reside
+    p1, p2, p3 = atom_coords[:3]
+    v1 = p3 - p1
+    v2 = p2 - p1
+    eq = np.cross(v1, v2)
+    # if the vector normal to plane is null vector, molecule is not planar
+    if not np.any(eq):
+        planar = 0
+
+    c1 = c1_coord
+    c2 = c2_coord
+    dir_vector = c2_coord - c1_coord
+    offset = bond_type * 0.045
+
+    # generating direction vectors of the bonds to be constructed
+    if planar:
+        sol = np.cross(dir_vector, eq)
+        sol = sol / np.linalg.norm(sol)     # coverting into unit vector
+    else:
+        l = dir_vector[1] - dir_vector[2]
+        m = dir_vector[2] - dir_vector[0]
+        n = dir_vector[0] - dir_vector[1]
+        sol = np.array([l, m, n])
+
+    c1_l = c1 - sol*offset
+    c1_u = c1 + sol*offset
+    c2_l = c2 - sol*offset
+    c2_u = c2 + sol*offset
+    return c1_l, c1_u, c2_l, c2_u
+
+###############################################################################
+# Here, we define the function which will create the ball and stick model
+# It takes in two numpy arrays as input-
+# atom_coords: array of atomic coordinates
+# elem_names: corresponding array of names of elements
 def ball_and_stick(atom_coords, elem_names):
+
+    # atomic radius of elements (will be used to generate bonding)
     atomic_radius = {'H':0.38, 'C':0.77}
+    # error_factor to factor in small discrepancies in radii of atoms
     error_factor = 0.05
     atomic_radius = {k:v + error_factor for k,v in atomic_radius.items()}
 
     atom_index = np.arange(0, len(elem_names))
-    atoms = elem_names
-
-    atoms_rad = [atomic_radius[x] for x in atoms]
-    # print(atoms_rad, atom_index)
+    atoms_rad = [atomic_radius[x] for x in elem_names]
 
     i_atom = np.copy(atom_index)
     p = np.copy(atom_coords)
@@ -18,13 +71,14 @@ def ball_and_stick(atom_coords, elem_names):
     r = atoms_rad
     r_compare = r
 
-    source_row = np.arange(len(atoms))
-    max_atoms = len(atoms)
+    source_row = np.arange(len(elem_names))
+    max_atoms = len(elem_names)
 
-    bonds = np.zeros((len(atoms)+1, max_atoms+1), dtype=np.int8)
-    bond_dists = np.zeros((len(atoms)+1, max_atoms+1), dtype=np.float32)
+    bonds = np.zeros((len(elem_names)+1, max_atoms+1), dtype=np.int8)
+    bond_dists = np.zeros((len(elem_names)+1, max_atoms+1), dtype=np.float32)
 
-    # Algorithm to generate bonding
[…]
