+++ docs/examples/viz_multiple_bonds.py
@@ -5,7 +5,7 @@ Multiple bonds visualization
 
 This example demonstrate how to create ball and stick model for different
 molecules from their atomic coordinates. It also shows how to visualize
-double and triple bonds
+double and triple bonds.
 """
 
 ###############################################################################
@@ -14,8 +14,8 @@ import numpy as np
 from fury import window, actor
 
 ###############################################################################
-
-# Offsets for double and triple bonds
+# Offsets for double and triple bonds (cgenerate oordinates of the streamtubes
+# which will represent bonds)
 def offsets(atom_coords, c1_coord, c2_coord, bond_type):
     # Initial assumption that the molecule is planar in nature
     planar = 1
@@ -53,7 +53,7 @@ def offsets(atom_coords, c1_coord, c2_coord, bond_type):
 # Here, we define the function which will create the ball and stick model
 # It takes in two numpy arrays as input-
 # atom_coords: array of atomic coordinates
-# elem_names: corresponding array of names of elements
+# elem_names: array of names of elements corresponding to atom_coords
 def ball_and_stick(atom_coords, elem_names):
 
     # atomic radius of elements (will be used to generate bonding)
@@ -77,8 +77,7 @@ def ball_and_stick(atom_coords, elem_names):
     bonds = np.zeros((len(elem_names)+1, max_atoms+1), dtype=np.int8)
     bond_dists = np.zeros((len(elem_names)+1, max_atoms+1), dtype=np.float32)
 
-    # Algorithm to generate bonding that is, set connections between different
-    # atoms.
+    # Algorithm to generate bonding (obtain connections between various atoms)
     for i in range(max_atoms-1):
         p_compare = np.roll(p_compare, -1, axis=0)
         #m_compare = np.roll(m_compare, -1, axis=0)
@@ -89,48 +88,67 @@ def ball_and_stick(atom_coords, elem_names):
         bond = np.where(np.logical_and(dists > 0.0001, dists < r_bond), 1, 0)
 
         source_row = source_row
-        target_row = source_row + i + 1 # Will be out of bounds of bonds array for some values of i
+        # target_row will be out of bounds of bonds array for some values of i
+        target_row = source_row + i + 1
+        # If invalid target, write to dummy row
         target_row = np.where(np.logical_or(target_row>len(elem_names),
                                             mask==1), len(elem_names),
-                                            target_row) #If invalid target, write to dummy row
+                                            target_row)
 
         source_atom = i_atom
-        target_atom = i_atom + i + 1 # Will be out of bounds of bonds array for some values of i
+        # target_atom will be out of bounds of bonds array for some values of i
+        target_atom = i_atom + i + 1
+        # If invalid target, write to dummy col
         target_atom = np.where(np.logical_or(target_atom>max_atoms, mask==1),
-                               max_atoms, target_atom) #If invalid target, write to dummy col
+                               max_atoms, target_atom)
 
         bonds[(source_row, target_atom)] = bond
         bonds[(target_row, source_atom)] = bond
         bond_dists[(source_row, target_atom)] = dists
         bond_dists[(target_row, source_atom)] = dists
 
-    bonds = np.delete(bonds, axis=0, obj=-1) #Delete dummy row
-    bonds = np.delete(bonds, axis=1, obj=-1) #Delete dummy col
[…]
