### 2021-06-18 a38b78e
small fixes, pep8
