+++ docs/examples/viz_multiple_bonds.py
@@ -14,8 +14,10 @@ import numpy as np
 from fury import window, actor
 
 ###############################################################################
-# Offsets for double and triple bonds (cgenerate oordinates of the streamtubes
+# Offsets for double and triple bonds (generate coordinates of the streamtubes
 # which will represent bonds)
+
+
 def offsets(atom_coords, c1_coord, c2_coord, bond_type):
     # Initial assumption that the molecule is planar in nature
     planar = 1
@@ -38,10 +40,10 @@ def offsets(atom_coords, c1_coord, c2_coord, bond_type):
         sol = np.cross(dir_vector, eq)
         sol = sol / np.linalg.norm(sol)     # coverting into unit vector
     else:
-        l = dir_vector[1] - dir_vector[2]
-        m = dir_vector[2] - dir_vector[0]
-        n = dir_vector[0] - dir_vector[1]
-        sol = np.array([l, m, n])
+        a = dir_vector[1] - dir_vector[2]
+        b = dir_vector[2] - dir_vector[0]
+        c = dir_vector[0] - dir_vector[1]
+        sol = np.array([a, b, c])
 
     c1_l = c1 - sol*offset
     c1_u = c1 + sol*offset
@@ -54,13 +56,15 @@ def offsets(atom_coords, c1_coord, c2_coord, bond_type):
 # It takes in two numpy arrays as input-
 # atom_coords: array of atomic coordinates
 # elem_names: array of names of elements corresponding to atom_coords
+
+
 def ball_and_stick(atom_coords, elem_names):
 
     # atomic radius of elements (will be used to generate bonding)
-    atomic_radius = {'H':0.38, 'C':0.77}
+    atomic_radius = {'H': 0.38, 'C': 0.77}
     # error_factor to factor in small discrepancies in radii of atoms
     error_factor = 0.05
-    atomic_radius = {k:v + error_factor for k,v in atomic_radius.items()}
+    atomic_radius = {k: v + error_factor for k, v in atomic_radius.items()}
 
     atom_index = np.arange(0, len(elem_names))
     atoms_rad = [atomic_radius[x] for x in elem_names]
@@ -80,10 +84,9 @@ def ball_and_stick(atom_coords, elem_names):
     # Algorithm to generate bonding (obtain connections between various atoms)
     for i in range(max_atoms-1):
         p_compare = np.roll(p_compare, -1, axis=0)
-        #m_compare = np.roll(m_compare, -1, axis=0)
         r_compare = np.roll(r_compare, -1, axis=0)
         mask = 0
-        dists = np.linalg.norm(p - p_compare, axis=1) #* mask
+        dists = np.linalg.norm(p - p_compare, axis=1)
         r_bond = r + r_compare
         bond = np.where(np.logical_and(dists > 0.0001, dists < r_bond), 1, 0)
 
@@ -91,36 +94,33 @@ def ball_and_stick(atom_coords, elem_names):
         # target_row will be out of bounds of bonds array for some values of i
         target_row = source_row + i + 1
         # If invalid target, write to dummy row
-        target_row = np.where(np.logical_or(target_row>len(elem_names),
-                                            mask==1), len(elem_names),
-                                            target_row)
+        target_row = np.where(np.logical_or(target_row > len(elem_names),
+                                            mask == 1), len(elem_names),
+                              target_row)
 
         source_atom = i_atom
         # target_atom will be out of bounds of bonds array for some values of i
         target_atom = i_atom + i + 1
         # If invalid target, write to dummy col
-        target_atom = np.where(np.logical_or(target_atom>max_atoms, mask==1),
[…]
