### 2021-06-18 a6bfed2
fix: FuryStreamClient Obj. has no attribute __interval_timer
