### 2021-06-18 0f45eb0
fix: cleanup of image_buffer_manager when mjpeg not requested
Before this commit, after finishing the web_server the cleanup of
image_buffer_manager obj. was not called
