### 2021-06-18 038ef8e
fix typos and docstrings
