### 2021-06-18 #3384 @whedon
[REVIEW]: FURY: advanced scientific visualization
**Submitting author:** @garyfallidis (<a href="http://orcid.org/0000-0002-3991-2774">Eleftherios Garyfallidis</a>)
**Repository:** <a href="https://github.com/fury-gl/fury" target ="_blank">https://github.com/fury-gl/fury</a>
**Version:** v0.7.1

- [x] #3445
