### 2021-06-18 6fc6cf4
fix: shared_mem repr does not to be converted
