### 2021-06-20 86bf413
triple bonds, double bonds thickness adjusted
