+++ docs/examples/viz_multiple_bonds.py
@@ -1,3 +1,4 @@
+
 """
 ====================================
 Multiple bonds visualization
@@ -11,7 +12,9 @@ double and triple bonds.
 ###############################################################################
 # First, we import some useful modules and methods.
 import numpy as np
-from fury import window, actor
+from fury import disable_warnings, window, actor
+import vtk
+disable_warnings()
 
 ###############################################################################
 # Offsets for double and triple bonds (generate coordinates of the streamtubes
@@ -33,7 +36,10 @@ def offsets(atom_coords, c1_coord, c2_coord, bond_type):
     c1 = c1_coord
     c2 = c2_coord
     dir_vector = c2_coord - c1_coord
-    offset = bond_type * 0.045
+    if bond_type == 2:
+        offset = 0.09
+    else:
+        offset = 0.10
 
     # generating direction vectors of the bonds to be constructed
     if planar:
@@ -125,12 +131,15 @@ def ball_and_stick(atom_coords, elem_names):
     # generate coordinates for bonds
     bond_coords = []
     bond_colors = []
+    doublebond_coords = []
+    doublebond_colors = []
+    triplebond_coords = []
+    triplebond_colors = []
     cpkr = {'H': [1, 1, 1, 1.2], 'C': [144/255, 144/255, 144/255, 1.7]}
 
     # indexes_done checks if the bond has been made between two atoms or not
     # and skips the iteration if the bond has been made
     indexes_done = []
-    i = 0
 
     # Here, we generate coordinates and colors to be passed to the streamtube
     # actor which will be representative of bonds
@@ -152,24 +161,31 @@ def ball_and_stick(atom_coords, elem_names):
                 # the streamtubes accordingly
                 else:
                     if bond_data['bond_lengths'][index][i] < 1.22:
-                        bond_type = 3
-                    elif bond_data['bond_lengths'][index][i] < 1.36:
-                        bond_type = 2
-                    else:
-                        bond_type = 1
-                    if bond_type == 1 or bond_type == 3:
-                        bond_coords += [[atom_coords[bond],
-                                         atom_coords[index]]]
-                        bond_colors += [cpkr[ename][:3]]
-                    if bond_type == 2 or bond_type == 3:
                         c1_l, c1_u, c2_l, c2_u = offsets(atom_coords,
                                                          atom_coords[index],
                                                          atom_coords[bond],
-                                                         bond_type)
+                                                         3)
+                        triplebond_colors += [cpkr[ename][:3]]
+                        triplebond_colors += [cpkr[ename][:3]]
+                        triplebond_coords += [[c1_l, c2_l]]
+                        triplebond_coords += [[c1_u, c2_u]]
+                        triplebond_coords += [[atom_coords[bond],
+                                               atom_coords[index]]]
+                        triplebond_colors += [cpkr[ename][:3]]
+                    elif bond_data['bond_lengths'][index][i] < 1.36:
+                        c1_l, c1_u, c2_l, c2_u = offsets(atom_coords,
+                                                         atom_coords[index],
+                                                         atom_coords[bond],
+                                                         2)
+                        doublebond_colors += [cpkr[ename][:3]]
+                        doublebond_colors += [cpkr[ename][:3]]
[…]
