### 2021-06-20 ea559c6
try to use Pillow if opencv it's not available
fix: flip RGB to BGR representation for jpeg encoding
