### 2021-06-20 0753c18
this allows to control the FPS for the MJPEG streaming
