### 2021-06-21 826e5e4
removes effect arg and fix the example
