### 2021-06-21 a18fbec
feat: force render on resie event
