### 2021-06-21 7ad14f6
serveral optimizations
- Fix GL Number
- Now, gl_something_state accepts only OpenGlState as argument
- add new blending effects
