+++ fury/ui/core.py
@@ -1,0 +1,1067 @@
+"""UI core module that describe UI abstract class."""
+
+__all__ = ["Rectangle2D", "Disk2D", "TextBlock2D"]
+
+import abc
+from warnings import warn
+
+import numpy as np
+import vtk
+
+from fury.interactor import CustomInteractorStyle
+from fury.utils import set_input
+
+
+class UI(object, metaclass=abc.ABCMeta):
+    """An umbrella class for all UI elements.
+
+    While adding UI elements to the scene, we go over all the sub-elements
+    that come with it and add those to the scene automatically.
+
+    Attributes
+    ----------
+    position : (float, float)
+        Absolute coordinates (x, y) of the lower-left corner of this
+        UI component.
+    center : (float, float)
+        Absolute coordinates (x, y) of the center of this UI component.
+    size : (int, int)
+        Width and height in pixels of this UI component.
+    on_left_mouse_button_pressed: function
+        Callback function for when the left mouse button is pressed.
+    on_left_mouse_button_released: function
+        Callback function for when the left mouse button is released.
+    on_left_mouse_button_clicked: function
+        Callback function for when clicked using the left mouse button
+        (i.e. pressed -> released).
+    on_left_mouse_double_clicked: function
+        Callback function for when left mouse button is double clicked
+        (i.e pressed -> released -> pressed -> released).
+    on_left_mouse_button_dragged: function
+        Callback function for when dragging using the left mouse button.
+    on_right_mouse_button_pressed: function
+        Callback function for when the right mouse button is pressed.
+    on_right_mouse_button_released: function
+        Callback function for when the right mouse button is released.
+    on_right_mouse_button_clicked: function
+        Callback function for when clicking using the right mouse button
+        (i.e. pressed -> released).
+    on_right_mouse_double_clicked: function
+        Callback function for when right mouse button is double clicked
+        (i.e pressed -> released -> pressed -> released).
+    on_right_mouse_button_dragged: function
+        Callback function for when dragging using the right mouse button.
+    on_middle_mouse_button_pressed: function
+        Callback function for when the middle mouse button is pressed.
+    on_middle_mouse_button_released: function
+        Callback function for when the middle mouse button is released.
+    on_middle_mouse_button_clicked: function
+        Callback function for when clicking using the middle mouse button
+        (i.e. pressed -> released).
+    on_middle_mouse_double_clicked: function
+        Callback function for when middle mouse button is double clicked
+        (i.e pressed -> released -> pressed -> released).
+    on_middle_mouse_button_dragged: function
+        Callback function for when dragging using the middle mouse button.
+    on_key_press: function
+        Callback function for when a keyboard key is pressed.
+
+    """
+
+    def __init__(self, position=(0, 0)):
+        """Init scene.
+
+        Parameters
+        ----------
+        position : (float, float)
+            Absolute coordinates (x, y) of the lower-left corner of this
+            UI component.
+
+        """
+        self._scene = object()
+        self._position = np.array([0, 0])
+        self._callbacks = []
+
+        self._setup()  # Setup needed actors and sub UI components.
+        self.position = position
+
+        self.left_button_state = "released"
+        self.right_button_state = "released"
+        self.middle_button_state = "released"
+
+        self.on_left_mouse_button_pressed = lambda i_ren, obj, element: None
[…]
