+++ fury/tests/test_ui.py
@@ -1,1611 +1,0 @@
-from os.path import join as pjoin
-import os
-import json
-import itertools
-
-import numpy as np
-import vtk
-
-
-import numpy.testing as npt
-import pytest
-import warnings
-
-from fury import window, actor, ui
-from fury.colormap import distinguishable_colormap
-from fury.data import DATA_DIR, read_viz_icons, fetch_viz_icons
-from fury.decorators import skip_win, skip_osx
-from fury.primitive import prim_sphere
-from fury.testing import assert_arrays_equal, assert_greater
-from fury.utils import shallow_copy
-
-
-class EventCounter(object):
-    def __init__(self, events_names=["CharEvent",
-                                     "MouseMoveEvent",
-                                     "KeyPressEvent",
-                                     "KeyReleaseEvent",
-                                     "LeftButtonPressEvent",
-                                     "LeftButtonReleaseEvent",
-                                     "RightButtonPressEvent",
-                                     "RightButtonReleaseEvent",
-                                     "MiddleButtonPressEvent",
-                                     "MiddleButtonReleaseEvent"]):
-        # Events to count
-        self.events_counts = {name: 0 for name in events_names}
-
-    def count(self, i_ren, _obj, _element):
-        """ Simple callback that counts events occurences. """
-        self.events_counts[i_ren.event.name] += 1
-
-    def monitor(self, ui_component):
-        for event in self.events_counts:
-            for obj_actor in ui_component.actors:
-                ui_component.add_callback(obj_actor, event, self.count)
-
-    def save(self, filename):
-        with open(filename, 'w') as f:
-            json.dump(self.events_counts, f)
-
-    @classmethod
-    def load(cls, filename):
-        event_counter = cls()
-        with open(filename, 'r') as f:
-            event_counter.events_counts = json.load(f)
-
-        return event_counter
-
-    def check_counts(self, expected):
-        npt.assert_equal(len(self.events_counts),
-                         len(expected.events_counts))
-
-        # Useful loop for debugging.
-        msg = "{}: {} vs. {} (expected)"
-        for event, count in expected.events_counts.items():
-            if self.events_counts[event] != count:
-                print(msg.format(event, self.events_counts[event], count))
-
-        msg = "Wrong count for '{}'."
-        for event, count in expected.events_counts.items():
-            npt.assert_equal(self.events_counts[event], count,
-                             err_msg=msg.format(event))
-
-
-def test_callback():
-    events_name = ["CharEvent", "MouseMoveEvent", "KeyPressEvent",
-                   "KeyReleaseEvent", "LeftButtonPressEvent",
-                   "LeftButtonReleaseEvent", "RightButtonPressEvent",
-                   "RightButtonReleaseEvent", "MiddleButtonPressEvent",
-                   "MiddleButtonReleaseEvent"]
-
-    class SimplestUI(ui.UI):
-        def __init__(self):
-            super(SimplestUI, self).__init__()
-
-        def _setup(self):
-            self.actor = vtk.vtkActor2D()
-
-        def _set_position(self, coords):
-            self.actor.SetPosition(*coords)
-
-        def _get_size(self):
-            return
-
-        def _get_actors(self):
-            return [self.actor]
-
-        def _add_to_scene(self, _scene):
-            return
-
-    simple_ui = SimplestUI()
-    current_size = (900, 600)
-    scene = window.Scene()
-    show_manager = window.ShowManager(scene,
[…]
