### 2021-06-22 4fb17d9
shows all the blending effects
