### 2021-06-22 ddafd39
a simple test for opengl_set_state
This dosen't check if the result of the effects are correct
