### 2021-06-22 e7d5cbb
refactor: until now, there is no reason for this behavior
Before this commit was possible to create a FuryStreamClient/Interaction whithout
creating a new raw_array. In addition, there is no good reason to have
a function which can be used to create a server which deals only with the user interactions
