+++ docs/tutorials/04_stream/viz_no_interaction.py
@@ -10,15 +10,6 @@ import numpy as np
 # Then let's download some available datasets.
 
 from fury.data.fetcher import fetch_viz_wiki_nw
-
-try:
-    import cv2 
-    OPENCV_AVAILABLE = True
-except ImportError:
-    cv2 = None
-    OPENCV_AVAILABLE = False
-
-print("\n\nOPENCV ", OPENCV_AVAILABLE,"\n\n")
 from fury.stream.server import web_server
 
 if __name__ == '__main__':
@@ -96,7 +87,6 @@ if __name__ == '__main__':
     p = multiprocessing.Process(
         target=web_server,
         args=(
-            None, 
             stream.image_buffers,
             stream.image_buffer_names,
             stream.info_buffer,
