### 2021-06-23 82dbdb3
adds docstrings and tests for IntervalTimer obj
