### 2021-06-23 b5590fd
adds docstrings and tests for CircularQueue obj
