### 2021-06-23 01bf2a3
fix: shared memory head_tail_buffer initial value
Before this commit, when using SharedMemory the CircularQueue
always started the head_tail_buffer as (0, 0, 0) instead of
(-1, -1, 0)
