### 2021-06-23 c7ffebc
test if the communication between shared multidimensional buffer objs
