+++ fury/tests/test_stream.py
@@ -44,11 +44,46 @@ def test_multidimensional_buffer():
         npt.assert_equal(np.array([.2, .3, .4, .5]), m_buffer[1])
         # cleanup
         m_buffer.cleanup()
-    
+
     # rawarray
-    test(False)
+    test(True)
     if PY_VERSION_8:
-        test(True)
+        test(False)
+
+    # test if the communication it's ok
+    def test_comm(use_raw_array=True):
+        # creates a raw_array
+        max_size = 3
+        dimension = 4
+        m_buffer_org = tools.MultiDimensionalBuffer(
+            max_size, dimension, use_raw_array=use_raw_array)
+        if use_raw_array:
+            m_buffer_0 = tools.MultiDimensionalBuffer(
+                max_size, dimension, buffer=m_buffer_org.buffer)
+        else:
+            m_buffer_0 = tools.MultiDimensionalBuffer(
+                max_size, dimension, buffer_name=m_buffer_org.buffer_name)
+
+        m_buffer_0[1] = np.array([.2, .3, .4, .5])
+
+        # check if the dimension it's correct
+        assert len(m_buffer_org[0]) == len(m_buffer_0[0])
+        # check if the max_size it's correct
+        assert len(m_buffer_org[max_size]) == 4 and\
+            len(m_buffer_org[max_size+1]) == 0
+        # check values
+        npt.assert_equal(np.array([.2, .3, .4, .5]), m_buffer_org[1])
+        # check it the correct max_size was recovered
+        assert m_buffer_org.max_size == m_buffer_0.max_size
+        assert m_buffer_org.dimension == m_buffer_0.dimension
+        # cleanup
+        m_buffer_0.cleanup()
+        m_buffer_org.cleanup()
+
+    # rawarray
+    test_comm(True)
+    if PY_VERSION_8:
+        test(False)
 
 
 def test_circular_queue():
@@ -80,7 +115,7 @@ def test_circular_queue():
         assert queue.head == 2 and queue.tail == 2
         arr_recovered = queue.dequeue()
         npt.assert_equal(arr+2, arr_recovered)
-        assert queue.head == -1 and queue.tail == -1 
+        assert queue.head == -1 and queue.tail == -1
         arr_recovered = queue.dequeue()
         assert arr_recovered is None
         queue.cleanup()
