### 2021-06-23 91703e2
there is no reason to pass the window_size
ShowManager already have the size parameter
