### 2021-06-23 6bb8699
fix: wrong arg passed to the ImageBufferManager
This bug was created by e85f9d233e1
