### 2021-06-23 1aa1b3b
docstrings for Widget Obj
