+++ requirements/default.txt
@@ -4,3 +4,5 @@ vtk>=9.1.0
 scipy>=1.2.0
 pillow>=5.4.1
 packaging
+aiohttp>=3.7.4
+aiortc>=1.2.0
