### 2021-06-23 2e53c67
docstrings: describes the event informations stored in the circular queue
