### 2021-06-23 a7c0529
fix: allows secure (ssl|wss) websocket communication
