### 2021-06-24 1b9a3da
remove this method due redundancy
The buffer data it's accesible through
```
circular_queue.buffer.buffer_repr
```
