### 2021-06-24 ef968af
test: CircularQueue communication
