### 2021-06-24 03f9e36
this will SegFaul or a silent bug due typos
The Circular Queue and the user interactions depends on the correct use
of index associated with the multidimensional buffer. A typo in those
index can create a segfault or a silent bug. In addition, any change in
the queue buffer implicates in changing those index across serveral
different files. This commit mitigates those problems
