### 2021-06-25 ad84404
improves code organization and readability using abstract classes
Also, after this commit it will be easier to define new memory management objs.
For example, through socket communication (py3.7 compatibility)
