+++ fury/stream/tools.py
@@ -3,6 +3,7 @@ import multiprocessing
 import time
 import logging
 from threading import Timer
+from abc import ABC, abstractmethod
 
 import sys
 if sys.version_info.minor >= 8:
@@ -24,28 +25,25 @@ def remove_shm_from_resource_tracker():
     def fix_register(name, rtype):
         if rtype == "shared_memory":
             return
-        return resource_tracker._resource_tracker.register(self, name, rtype)
+        return resource_tracker._resource_tracker.register(
+            self, name, rtype)
     resource_tracker.register = fix_register
 
     def fix_unregister(name, rtype):
         if rtype == "shared_memory":
             return
-        return resource_tracker._resource_tracker.unregister(self, name, rtype)
+        return resource_tracker._resource_tracker.unregister(
+            self, name, rtype)
     resource_tracker.unregister = fix_unregister
 
     if "shared_memory" in resource_tracker._CLEANUP_FUNCS:
         del resource_tracker._CLEANUP_FUNCS["shared_memory"]
 
 
-class MultiDimensionalBuffer:
+class GenericMultiDimensionalBuffer(ABC):
     def __init__(
-            self, max_size=None, dimension=8, buffer=None,
-            buffer_name=None,
-            use_raw_array=True):
-        """This implements a generic multidimensional buffer.
-        This buffer can work with RawArrays or SharedMemory.
-        Stream system uses that to implemenet the CircularQueue
-        with shared memory resources.
+            self, max_size=None, dimension=8):
+        """This implements a abstract (generic) multidimensional buffer.
 
         Parameters
         ----------
@@ -53,69 +51,13 @@ class MultiDimensionalBuffer:
             If buffer_name or buffer was not passed then max_size
             it's mandatory
         dimension : int, default 8
-        buffer : buffer, optional
-            If buffer and buffer name is not passed to __init__
-            then the multidimensional buffer obj will create a new
-            RawArray or SharedMemory object to store the data
-            If buffer is passed than this Obj will read a
-            a already created RawArray
-        buffer_name : str, optional
-            if buffer_name is passed than this Obj will read a
-            a already created SharedMemory
-        use_raw_array : bool
-            if use_raw_array is False(True) and both buffer and buffer_name
-            are None, then this Obj will create a float64
-            SharedMemory(RawArray) with dim = dimension*max_size
-
         """
-        use_raw_array = use_raw_array and buffer_name is None
-        if not PY_VERSION_8 and not use_raw_array:
-            raise ValueError("""
-                In order to use the SharedMemory approach
-                you should have to use python 3.8 or higher""")
-
-        if buffer is None and buffer_name is None:
-            if use_raw_array:
-
-                buffer_arr = np.zeros(dimension*(max_size+1), dtype='float64')
-                buffer = multiprocessing.RawArray(
-                        'd', buffer_arr)
-                self._buffer = buffer
-                self._buffer_repr = np.ctypeslib.as_array(self._buffer)
-            else:
-                buffer_arr = np.zeros(dimension*(max_size+1), dtype='float64')
-                buffer = shared_memory.SharedMemory(
-                    create=True, size=buffer_arr.nbytes)
-                self._buffer_repr = np.ndarray(
-                        buffer_arr.shape[0],
-                        dtype='float64', buffer=buffer.buf)
-                self._buffer = buffer
-                buffer_name = buffer.name
-                self._unlink_shared_mem = True
-                # print('created', max_size, dimension,  len(buffer.buf))
-        else:
-            if buffer_name is None:
-                max_size = int(len(buffer)//dimension)
-                max_size -= 1
-                self._buffer = buffer
[…]
