### 2021-06-25 18367a9
improves queue code organization and lock system
If the circular queue uses multiprocessing.Array there
it's no reason to use a diferent lock implementation instead of
get_lock

Also, after this commit it will be easier to define new memory management objs.
