+++ fury/stream/tools.py
@@ -209,16 +209,12 @@ class SharedMemMultiDimensionalBuffer(GenericMultiDimensionalBuffer):
                     File not found')
 
 
-class CircularQueue:
+class GenericCircularQueue(ABC):
     def __init__(
-        self, max_size=10, dimension=6,
-            head_tail_buffer=None,
-            head_tail_buffer_name=None,  buffer=None,
-            buffer_name=None, use_raw_array=True):
-        """This implements a MultiDimensional Queue which works with
+            self, max_size=None, dimension=8,
+            use_shared_mem=False, buffer=None, buffer_name=None):
+        """This implements a generic circular queue which works with
         shared memory resources.
-        This can work with RawArrays or SharedMemory.
-        Stream system uses that to implemenet user interactions
 
         Parameters
         ----------
@@ -228,85 +224,25 @@ class CircularQueue:
             multidimensional buffer
         dimension : int, default 8
             This will be used to construct the multidimensional buffer
-        head_tail_buffer : buffer, optional
-            If buffer and buffer name is not passed to __init__
-            then this obj will create a new
-            RawArray or SharedMemory object to store head and tail position.
-            If buffer is passed than this Obj will read a
-            a already created RawArray
-        head_tail_buffer_name : str, optional
-            if buffer_name is passed than this Obj will read a
-            a already created SharedMemory with the head and tail
-            informations
-        buffer : buffer, optional
-            If buffer and buffer name is not passed to __init__
-            then the multidimensional buffer obj will create a new
-            RawArray or SharedMemory object to store the data
-            If buffer is passed than this Obj will read a
-            a already created RawArray
-        buffer_name : str, optional
-            if buffer_name is passed than this Obj will read a
-            a already created SharedMemory
-        use_raw_array : bool
-            if use_raw_array is False(True) and both buffer and buffer_name
-            are None, then this Obj will create a int64
-            SharedMemory(RawArray) with dim = dimension*max_size
+        use_shared_mem : bool, default False
+            If the multidimensional memory resource should create or read
+            using SharedMemory or RawArrays
+        buffer : RawArray, optional
+        buffer_name: str, optional
 
         """
-
-        use_raw_array = use_raw_array and buffer_name is None
-        if not PY_VERSION_8 and not use_raw_array:
-            raise ValueError("""
-                In order to use the SharedMemory approach
-                you should have to use python 3.8 or higher""")
-        if use_raw_array:
-            buffer = RawArrayMultiDimensionalBuffer(
+        self._created = False
+        self.head_tail_buffer_name = None
+        self.head_tail_buffer_repr = None
+        self.head_tail_buffer = None
+        if use_shared_mem:
+            self.buffer = SharedMemMultiDimensionalBuffer(
+                max_size=max_size, dimension=dimension, buffer_name=buffer_name
+            )
+        else:
+            self.buffer = RawArrayMultiDimensionalBuffer(
                 max_size=max_size, dimension=dimension, buffer=buffer
             )
-        else:
-            buffer = SharedMemMultiDimensionalBuffer(
-                max_size=max_size, dimension=dimension,
-                buffer_name=buffer_name
-            )
-        # head_tail_arr[0] int; head position
-        # head_tail_arr[1] int; tail position
-        # head_tail_arr[2] 0 or 1; if this memory resource it's busy or not
-        head_tail_arr = np.array([-1, -1, 0], dtype='int64')
-        if head_tail_buffer is None and head_tail_buffer_name is None:
-            self._created_shared_mem = True
-            if use_raw_array:
[…]
