+++ fury/stream/server/__init__.py
@@ -1,3 +1,3 @@
-from fury.stream.server.server import web_server, ImageBufferManager
+from fury.stream.server.server import web_server
 
-__all__ = ['web_server', 'ImageBufferManager']
+__all__ = ['web_server', ]
