+++ fury/stream/client.py
@@ -1,8 +1,6 @@
 import vtk
-import multiprocessing
 import sys
 if sys.version_info.minor >= 8:
-    from multiprocessing import shared_memory
     PY_VERSION_8 = True
 else:
     shared_memory = None
@@ -10,8 +8,10 @@ else:
 
 import numpy as np
 
-from fury.stream.tools import ArrayCircularQueue, SharedMemCircularQueue,\
-    IntervalTimer
+from fury.stream.tools import ArrayCircularQueue, SharedMemCircularQueue
+from fury.stream.tools import (
+    RawArrayImageBufferManager, SharedMemImageBufferManager)
+from fury.stream.tools import IntervalTimer
 from fury.stream.constants import _CQUEUE
 
 import logging
@@ -38,7 +38,7 @@ class FuryStreamClient:
                 Should be greater than the window size.
             use_raw_array : bool, optional
                 If False then FuryStreamClient will use SharedMemory
-                instead of RawArrays. Notice that Python >=3.8 it's a 
+                instead of RawArrays. Notice that Python >=3.8 it's a
                 requirement
                 to use SharedMemory)
             whithout_iren_start : bool, optional
@@ -71,57 +71,12 @@ class FuryStreamClient:
             raise ValueError("""
                SharedMemory works only in python 3.8 or higher""")
 
-        # info_list stores the information about the n frame buffers
-        # as well the respectives sizes.
-        # 0 number of components
-        # 1 id buffer
-        # 2, 3, width first buffer, height first buffer
-        # 4, 5, width second buffer , height second buffer
-        info_list = [3, 0]
-        for _ in range(self.num_buffers):
-            info_list += [self.max_window_size[0]]
-            info_list += [self.max_window_size[1]]
-        info_list = np.array(
-            info_list, dtype='uint64'
-        )
         if use_raw_array:
-            self.info_buffer = multiprocessing.RawArray(
-                    'I', info_list
-            )
-            self.info_buffer_repr = np.ctypeslib.as_array(self.info_buffer)
+            self.img_manager = RawArrayImageBufferManager(
+                max_window_size=max_window_size, num_buffers=num_buffers)
         else:
-            self.info_buffer = shared_memory.SharedMemory(
-                create=True, size=info_list.nbytes)
-            self.info_buffer_repr = np.ndarray(
-                    info_list.shape[0],
-                    dtype='uint64', buffer=self.info_buffer.buf)
-            self.info_buffer_name = self.info_buffer.name
-
-        # creates the shared memory resources to store the
-        # image buffers
-        if use_raw_array:
-            self.image_buffer_names = None
-
-            self.image_reprs = []
-            for _ in range(self.num_buffers):
-                buffer = multiprocessing.RawArray(
-                    'B', np.random.randint(
-                        0, 255,
-                        size=max_window_size[0]*max_window_size[1]*3)
-                    .astype('uint8'))
-                self.image_buffers.append(buffer)
-                self.image_reprs.append(
-                    np.ctypeslib.as_array(buffer))
-        else:
-            for _ in range(self.num_buffers):
-                bufferSize = max_window_size[0]*max_window_size[1]*3
-                buffer = shared_memory.SharedMemory(
-                    create=True, size=bufferSize)
-                self.image_buffers.append(buffer)
-                self.image_reprs.append(
-                    np.ndarray(
-                        bufferSize, dtype=np.uint8, buffer=buffer.buf))
-                self.image_buffer_names.append(buffer.name)
+            self.img_manager = SharedMemImageBufferManager(
+                max_window_size=max_window_size, num_buffers=num_buffers)
 
         self._id_timer = None
[…]
