+++ fury/stream/tools.py
@@ -5,6 +5,7 @@ import logging
 from threading import Timer
 from abc import ABC, abstractmethod
 
+import asyncio
 import sys
 if sys.version_info.minor >= 8:
     from multiprocessing import shared_memory
@@ -14,10 +15,17 @@ if sys.version_info.minor >= 8:
 else:
     shared_memory = None
     PY_VERSION_8 = False
+try:
+    import cv2
+    OPENCV_AVAILABLE = True
+except ImportError:
+    from PIL import Image
+    cv2 = None
+    OPENCV_AVAILABLE = False
 
 
 def remove_shm_from_resource_tracker():
-    """Monkey-patch multiprocessing.resource_tracker so SharedMemory won't 
+    """Monkey-patch multiprocessing.resource_tracker so SharedMemory won't
     be tracked
     More details at: https://bugs.python.org/issue38119
     """
@@ -235,6 +243,7 @@ class GenericCircularQueue(ABC):
         self.head_tail_buffer_name = None
         self.head_tail_buffer_repr = None
         self.head_tail_buffer = None
+        self._use_shared_mem = use_shared_mem
         if use_shared_mem:
             self.buffer = SharedMemMultiDimensionalBuffer(
                 max_size=max_size, dimension=dimension, buffer_name=buffer_name
@@ -246,10 +255,10 @@ class GenericCircularQueue(ABC):
 
     @property
     def head(self):
-        if self.use_raw_array:
-            return np.frombuffer(self.head_tail_buffer.get_obj(), 'int64')[0]
-        else:
+        if self._use_shared_mem:
             return self.head_tail_buffer_repr[0]
+        else:
+            return np.frombuffer(self.head_tail_buffer.get_obj(), 'int64')[0]
 
     @head.setter
     def head(self, value):
@@ -257,10 +266,10 @@ class GenericCircularQueue(ABC):
 
     @property
     def tail(self):
-        if self.use_raw_array:
-            return np.frombuffer(self.head_tail_buffer.get_obj(), 'int64')[1]
-        else:
+        if self._use_shared_mem:
             return self.head_tail_buffer_repr[1]
+        else:
+            return np.frombuffer(self.head_tail_buffer.get_obj(), 'int64')[1]
 
     @tail.setter
     def tail(self, value):
@@ -356,7 +365,6 @@ class ArrayCircularQueue(GenericCircularQueue):
 
         self.head_tail_buffer_name = None
         self.head_tail_buffer_repr = self.head_tail_buffer
-        self.use_raw_array = True
         if self._created:
             self.set_head_tail(-1, -1, 0)
 
@@ -424,7 +432,6 @@ class SharedMemCircularQueue(GenericCircularQueue):
         self.head_tail_buffer_repr = np.ndarray(
                 3,
                 dtype='int64', buffer=self.head_tail_buffer.buf)
-        self.use_raw_array = False
         if self._created:
             self.set_head_tail(-1, -1, 0)
 
@@ -458,6 +465,7 @@ class SharedMemCircularQueue(GenericCircularQueue):
         return ok
 
     def dequeue(self):
+        interactions = None
         if self.is_unlocked():
             self.lock()
             interactions = self._dequeue()
@@ -468,7 +476,6 @@ class SharedMemCircularQueue(GenericCircularQueue):
         self.buffer.cleanup()
         self.head_tail_buffer.close()
         if self._created:
-            print('unlink')
             # this it's due the python core issues
             # https://bugs.python.org/issue38119
             # https://bugs.python.org/issue39959
@@ -481,7 +488,292 @@ class SharedMemCircularQueue(GenericCircularQueue):
                      File not found')
 
 
-class IntervalTimer(object):
+class GenericImageBufferManager(ABC):
[…]

[…]
