### 2021-06-25 a309771
test: Widget, StreamInteraction and RTCStream
