### 2021-06-28 bb45d1c
feat: added TreeNode2D to UI
