+++ azure-pipelines.yml
@@ -24,6 +24,8 @@ jobs:
         python.version: '3.6'
       Python38-64bit:
         python.version: '3.8'
+      Python39-64bit:
+        python.version: '3.9'
       CONDA Python37-64bit + OPTIONAL_DEPS:
         python.version: '3.7'
         OPTIONAL_DEPS: true
