+++ .github/workflows/test.yml
@@ -21,7 +21,7 @@ jobs:
     strategy:
       fail-fast: false
       matrix:
-        python-version: [3.6, 3.7, 3.8, 3.9]
+        python-version: [3.6, 3.7, 3.8]
         os: [ubuntu-latest, macos-latest]
         platform: [x64]
         install_type: [pip, conda]
@@ -44,6 +44,18 @@ jobs:
             experimental: false
           - os: ubuntu-20.04
             python-version: 3.8
+            install_type: pip
+            depends: OPTIONAL_DEPS
+            option-flags: 'USE_PRE'
+            experimental: true
+          - os: ubuntu-20.04
+            python-version: 3.9
+            install_type: pip
+            depends: OPTIONAL_DEPS
+            option-flags: 'USE_PRE'
+            experimental: true
+          - os: macos-latest
+            python-version: 3.9
             install_type: pip
             depends: OPTIONAL_DEPS
             option-flags: 'USE_PRE'
