### 2021-06-28 73f9c57
Some stylistic changes
