### 2021-06-28 55681f6
doc: fixed the docstring
