+++ fury/ui.py
@@ -70,85 +70,93 @@ def clip_overflow(textblock, width, side='right'):
 
 
 class Watcher:
-        def __init__(self, object):
-            """Class to monitor a UI element in runtime
+    """Class to monitor a UI element in runtime
 
-            Parameters
-            ----------
-            instance: :class: `UI`
-                UI element instance
-            show_m: :class: `ShowManager`
-                Show Manager
-            is_running: bool
-                Current running state of the watcher
-            i_ren: :class: `CustomInteractorStyle`
-                CustomInteractorStyle
-            """
-            self.instance = object
-            self.show_m = None
-            self.i_ren = None
-            self.attr = None
-            self.original_attr = None
-            self.updated_attr = None
-            self.is_running = False
-            self.callback = None
+    Attributes
+    ----------
+    instance: :class: `UI`
+        UI element instance
+    show_m: :class: `ShowManager`
+        Show Manager
+    is_running: bool
+        Current running state of the watcher
+    i_ren: :class: `CustomInteractorStyle`
+        CustomInteractorStyle
+    """
+    
+    def __init__(self, object):
+        """Initialize the watcher class
 
-        def start(self, delay, show_m, attr):
-            """Start the watcher
+        Parameters
+        ----------
+        object: :class: `UI`
+            Instance of the UI element
+        """
+        self.instance = object
+        self.show_m = None
+        self.i_ren = None
+        self.attr = None
+        self.original_attr = None
+        self.updated_attr = None
+        self.is_running = False
+        self.callback = None
 
-            Parameters
-            ----------
-            delay: int
-                delay between each update call
-            show_m: :class: `window.ShowManager`
-                show manager
-            attr: str
-                attribute to watch
-            """
-            self.show_m = show_m
-            self.attr = attr
-            self.i_ren = self.show_m.scene.GetRenderWindow()\
-                .GetInteractor().GetInteractorStyle()
+    def start(self, delay, show_m, attr):
+        """Start the watcher
 
-            if hasattr(self.instance, self.attr):
-                self.original_attr = getattr(self.instance, attr)
+        Parameters
+        ----------
+        delay: int
+            delay between each update call
+        show_m: :class: `window.ShowManager`
+            show manager
+        attr: str
+            attribute to watch
+        """
+        self.show_m = show_m
+        self.attr = attr
+        self.i_ren = self.show_m.scene.GetRenderWindow()\
+            .GetInteractor().GetInteractorStyle()
 
-                if type(self.original_attr) == np.array:
-                    self.original_attr = self.original_attr.tolist()
-            else:
-                raise(AttributeError(
-                    f'{self.instance} has no attribute {self.attr}')
-                )
+        if hasattr(self.instance, self.attr):
+            self.original_attr = getattr(self.instance, attr)
 
-            if delay > 0:
-                self.id_timer = self.show_m\
-                    .add_timer_callback(True, delay, self.update)
-            else:
-                self.id_observer = self.i_ren.AddObserver('RenderEvent',
-                                                          self.update)
+            if type(self.original_attr) == np.array:
+                self.original_attr = self.original_attr.tolist()
+        else:
+            raise(AttributeError(
+                f'{self.instance} has no attribute {self.attr}')
+            )
 
-            self.is_running = True
+        if delay > 0:
+            self.id_timer = self.show_m\
+                .add_timer_callback(True, delay, self.update)
+        else:
+            self.id_observer = self.i_ren.AddObserver('RenderEvent',
+                                                        self.update)
 
-        def stop(self):
-            """Stop the watcher
-            """
-            if self.id_timer:
[…]
