### 2021-06-29 5078b9e
Removed OpenGL refractive functionality.
