### 2021-06-29 3f29934
fixes OSx shared memory segment minimun size
MacOS has a default minimum size for a shared memory segment. Before this commit,
memory corruption was possible freezing the user interaction

Refs:
https://www.deepanseeralan.com/tech/playing-with-shared-memory/
http://www.spy-hill.com/help/apple/SharedMemory.html
