### 2021-06-29 #3384 @rougier
Some update on my (stalled) review: I have a hard time installing vtk on OSX Catalina. Once solved, I'll be able to resume my review.
