### 2021-06-30 #3384 @phamvanvung
> Some update on my (stalled) review: I have a hard time installing vtk on OSX Catalina. Once solved, I'll be able to resume my review. [EDIT] Solved.

I am running macOS too (Big Sur) and tried installing the software (all three ways mentioned in the documentation: pip, conda, and development). They work well. It also downloads and installs vtk!=9.0.0,>=8.1.2 (from fury) automatically.
