### 2021-06-30 #3384 @rougier
I've a hard time installing VTK on OSX Catalina, just for the record, the procedure using conda was:
```
conda config --set channel_priority strict
conda create -n vtk -c conda-forge python=3.7 numpy scipy pillow vtk
conda activate vtk
python -c "import vtk"
```

I had no trouble installing Fury following the instructions but the test failed with a segmentation fault (when testing picking). The output was rather scarce so it might be difficult to debug. But I managed to run some of the tutorials (I've not tested all of them). Overall it's a nice and much need package. Each example is a bit slow to start but I suspect this is because of VTK.

Concerning the paper, I've some minor comments:
- The paper is rather short which is ok for JOSS but in the meantime it's a bit frustrating for the reader. For example, you shortly compared with mayavi/glumpy/vispy but it's a bit difficult to compare since you do not provide much details. For example, why did you not re-use one of them (lack of support? some ingredient were missing, etc).
- You explain that FURY is interopable with lot of software but we don't have details. For example, what does it means for matplotlib? I think people (including me) woudl be very interested to have either one example or further explanations.
- In the list of authors, it's nice to recognize "FURY Contributors". I just wondered if you could add either a footnote or a dedicated section to list them but this would imply permission and may be too much trouble.
- For the DOI, I think some are missing (for example Numpy), could you fix them?
