+++ fury/ui.py
@@ -13,7 +13,7 @@ from fury.data import read_viz_icons
 from fury.interactor import CustomInteractorStyle
 from fury.io import load_image
 from fury.utils import set_input, rotate
-from fury.actor import grid
+from fury.actor import grid, label
 
 
 TWO_PI = 2 * np.pi
@@ -5246,13 +5246,199 @@ class GridUI(UI):
         # self.container.SetPosition(*coords)
 
 
+class Tree2D(UI):
+    """Render nodes in tree form
+
+    Attributes
+    ----------
+    structure: List(Dict, Dict, .....)
+        [{'node': ['child', 'child']}, .....]
+    tree_name: str
+        Name of the tree
+    node_height: int
+        Space taken by each node vertically
+    indent: int
+        Global indentation for the parent/child nodes
+    nodes: list of :class: `TreeNode2D`
+        List of all parent nodes present in Tree2D
+    nodes_dict: dict
+        Dict with label, nodes as key/value pairs
+    """
+
+    def __init__(self, structure, tree_name="", position=(0, 0),
+                 size=(300, 300), node_height=30, color=(0.3, 0.3, 0.3),
+                 opacity=0.8, indent=25):
+        """Initialize the UI element
+
+        Parameter
+        ---------
+        structure: List(Dict, Dict, .....)
+            [{'node': ['child', 'child']}, .....]
+        tree_name: str, optional
+            Label for the tree
+        position : (float, float), optional
+            Absolute coordinates (x, y) of the lower-left corner of the
+            UI component
+        size : (int, int), optional
+            Width and height of the pixels of this UI component.
+        node_height: int, optional
+            Space taken by each node vertically
+        color : list of 3 floats, opational
+            Background color of the Tree2D
+        opacity: float, optional
+            Background opacity of the Tree2D
+        indent: int, optional
+            Global indentation for the parent/child nodes
+        """
+        self.structure = structure
+        self.tree_name = tree_name
+        self.indent = indent
+        self._nodes = []
+        self._nodes_dict = {}
+        self.node_height = node_height
+
+        super(Tree2D, self).__init__(position)
+        self.resize(size)
+        self.base_node.color = color
+        self.base_node.opacity = opacity
+
+    def _setup(self):
+        """Setup this UI element."""
+        self.generate_tree(self.structure)
+
+        self.base_node = TreeNode2D(label=self.tree_name, children=self._nodes,
+                                    expandable=False, expanded=True,
+                                    indent=self.indent, child_indent=self.indent,
+                                    child_height=self.node_height)
+
+    def _get_actors(self):
+        """ Get the actors composing this UI component.
+        """
+        return self.base_node.actors
+
+    def _add_to_scene(self, _scene):
+        """ Add all subcomponents or VTK props that compose this UI component.
+
+        Parameters
+        ----------
+        scene : scene
+        """
+        self.base_node.add_to_scene(_scene)
+
+    def _set_position(self, _coords):
+        """ Position the lower-left corner of this UI component.
+
+        Parameters
+        ----------
+        coords: (float, float)
+            Absolute pixel coordinates (x, y).
+        """
+        self.base_node.position = _coords
+
+    def _get_size(self):
+        return self.base_node.size
+
+    def generate_tree(self, structure):
+        """Generate the structure of the Tree2D
+
+        Parameters
+        ----------
+        structure: List(Dict, Dict, .....)
+            [{'node': ['child', 'child']}, .....]
+        """
+        for obj in structure:
+            parent_label = list(obj.keys())[0]
+            child_labels = list(obj.values())[0]
+
+            parent_node = TreeNode2D(label=parent_label)
+            child_nodes = [TreeNode2D(label=child_label) for child_label in
[…]
