### 2021-06-30 336b7df
feat: added watcher to base UI class
