### 2021-06-30 70e04d3
feat: added user hook for panel resizing
