### 2021-06-30 f1d665f
example for panel resizing
