### 2021-06-30 #3384 @phamvanvung
I tested the software on:
macOS (Big Sur), conda 4.10.2 environment.
All the installation instructions work well.

There is an automated test, and I tried executing the test. Many cases showed "PASSED" status, but there are still some errors shown (this might be due to the version compatibility from my environment):
E.g.:
fury/tests/test_pick.py::test_picking_manager Fatal Python error: Segmentation fault

The statement of need is clear from the Software paper (PDF). However, though I can figure out the statement of needs from the Documentation page, it is unclear for me (as required from the checklist).

Given the limitation of the number of pages, the Software paper does a good job stating its aims, architectures, and quick pointers to comparable and commonly used software packages. They also provide the main differences between these similar packages and FURY. Interested readers can refer to these pointers for further comparisons if needed.

Regarding the references:
1. It was hard for me to find the source for the "Bullet physics library" reference. Please add "Available at: http://www.bulletphysics.org," if possible.
2. Here is the DOI for the "Array programming with NumPy" reference: https://doi.org/10.1038/s41586-020-2649-2
3. Here is the DOI for the "Advancing Education and Research in Nanotechnology" reference: http://dx.doi.org/10.1109/MCSE.2008.120
4. Similarly, the authors can try and search for the DOIs of the missing references.

All in all, I have positive feelings regarding the usability of this software and highly support its use.
