+++ azure-pipelines.yml
@@ -13,7 +13,7 @@ jobs:
   variables:
     AZURE_CI_WINDOWS: 'true'
   strategy:
-    maxParallel: 4
+    maxParallel: '4'
     matrix:
       Python37-64bit + OPTIONAL_DEPS:
         python.version: '3.7'
@@ -26,17 +26,16 @@ jobs:
         python.version: '3.8'
       Python39-64bit:
         python.version: '3.9'
-        EXTRA_PIP_FLAGS: "--pre"
-      CONDA Python37-64bit + OPTIONAL_DEPS:
-        python.version: '3.7'
-        OPTIONAL_DEPS: true
-        INSTALL_TYPE: "conda"
-      CONDA Python37-64bit:
-        python.version: '3.7'
-        INSTALL_TYPE: "conda"
-      CONDA Python36-64bit:
-        python.version: '3.6'
-        INSTALL_TYPE: "conda"
+      # CONDA Python37-64bit + OPTIONAL_DEPS:
+      #   python.version: '3.7'
+      #   OPTIONAL_DEPS: true
+      #   INSTALL_TYPE: "conda"
+      # CONDA Python37-64bit:
+      #   python.version: '3.7'
+      #   INSTALL_TYPE: "conda"
+      # CONDA Python36-64bit:
+      #   python.version: '3.6'
+      #   INSTALL_TYPE: "conda"
   steps:
   - task: UsePythonVersion@0
     inputs:
