+++ .github/workflows/test.yml
@@ -21,10 +21,10 @@ jobs:
     strategy:
       fail-fast: false
       matrix:
-        python-version: [3.6, 3.7, 3.8]
+        python-version: [3.6, 3.7, 3.8, 3.9]
         os: [ubuntu-latest, macos-latest]
         platform: [x64]
-        install_type: [pip, conda]
+        install_type: [pip, ]  # conda]
         depends: [DEFAULT_DEPS, OPTIONAL_DEPS]
         pip-flags: ['']
         option-flags: ['']
@@ -47,18 +47,16 @@ jobs:
             install_type: pip
             depends: OPTIONAL_DEPS
             option-flags: 'USE_PRE'
-            experimental: true
+            experimental: false
           - os: ubuntu-20.04
             python-version: 3.9
-            install_type: pip
+            install_type: conda
             depends: OPTIONAL_DEPS
-            option-flags: 'USE_PRE'
             experimental: true
           - os: macos-latest
-            python-version: 3.9
-            install_type: pip
+            python-version: 3.8
+            install_type: conda
             depends: OPTIONAL_DEPS
-            option-flags: 'USE_PRE'
             experimental: true
     env:
       TRAVIS_PYTHON_VERSION: ${{ matrix.python-version }} # Keep backward compatibilty with Travis
@@ -130,6 +128,10 @@ jobs:
     - name: Install FURY and Test it!
       run: |
         bash ci/travis/script.sh
+    - name: Check if experimental
+      if: ${{ failure() && matrix.experimental == 'true' }}
+      run: |
+        echo -e "::warning ::Error on this experimental matrix"
     - name: Deploy docs
       if: success() && github.ref == 'refs/heads/master'
       run: |
