### 2021-07-01 3a79ae4
doc: updated docstring of Panel2D
