### 2021-07-01 a7eca37
fix: updated changes w.r.t UI restructuring
