+++ fury/ui/containers.py
@@ -24,7 +24,8 @@ class Panel2D(UI):
     """
 
     def __init__(self, size, position=(0, 0), color=(0.1, 0.1, 0.1),
-                 opacity=0.7, align="left"):
+                 opacity=0.7, align="left", border_color=(1, 1, 1),
+                 border_width=0):
         """Init class instance.
 
         Parameters
@@ -39,7 +40,13 @@ class Panel2D(UI):
             Must take values in [0, 1].
         align : [left, right]
             Alignment of the panel with respect to the overall screen.
+        border_color: (float, float, float), optional
+            Must take values in [0, 1].
+        border_width: float, optional
+            width of the border
         """
+        self.border_colors = border_color
+        self.border_widths = border_width
         super(Panel2D, self).__init__(position)
         self.resize(size)
         self.alignment = align
@@ -52,19 +59,35 @@ class Panel2D(UI):
         """Setup this UI component.
 
         Create the background (Rectangle2D) of the panel.
-
+        Create the borders (Rectangle2D) of the panel.
         """
         self._elements = []
         self.element_offsets = []
         self.background = Rectangle2D()
+
+        self.borders = {'left': Rectangle2D(),
+                        'right': Rectangle2D(),
+                        'top': Rectangle2D(),
+                        'bottom': Rectangle2D()}
+
+        self.border_coords = {'left': (0, 0),
+                              'right': (0, 0),
+                              'top': (0, 0),
+                              'bottom': (0, 0)}
+
         self.add_element(self.background, (0, 0))
+
+        for key in self.borders.keys():
+            self.borders[key].color = self.border_colors
+            self.add_element(self.borders[key], self.border_coords[key])
 
         # Add default events listener for this UI component.
         self.background.on_left_mouse_button_pressed = self.left_button_pressed
         self.background.on_left_mouse_button_dragged = self.left_button_dragged
 
     def _get_actors(self):
-        """Get the actors composing this UI component."""
+        """Get the actors composing this UI component.
+        """
         actors = []
         for element in self._elements:
             actors += element.actors
@@ -77,7 +100,6 @@ class Panel2D(UI):
         Parameters
         ----------
         scene : scene
-
         """
         for element in self._elements:
             element.add_to_scene(scene)
@@ -92,9 +114,22 @@ class Panel2D(UI):
         ----------
         size : (float, float)
             Panel size (width, height) in pixels.
-
         """
         self.background.resize(size)
+
+        self.borders['left'].resize((self.border_widths,
+                                     size[1]+self.border_widths))
+
+        self.borders['right'].resize((self.border_widths,
+                                      size[1]+self.border_widths))
+
+        self.borders['top'].resize((self.size[0]+self.border_widths,
+                                    self.border_widths))
+
+        self.borders['bottom'].resize((self.size[0]+self.border_widths,
+                                       self.border_widths))
+
+        self.update_border_coords()
 
     def _set_position(self, coords):
         """Set the lower-left corner position of this UI component.
@@ -219,6 +254,73 @@ class Panel2D(UI):
         else:
             msg = "You can only left-align or right-align objects in a panel."
             raise ValueError(msg)
+
+    def update_border_coords(self):
+        """Update the coordinates of the borders
+        """
+        self.border_coords = {'left': (0., 0.),
+                              'right': (1., 0.),
+                              'top': (0., 1.),
+                              'bottom': (0., 0.)}
+
+        for key in self.borders.keys():
+            self.update_element(self.borders[key], self.border_coords[key])
+
+    @property
+    def border_color(self):
+        labels = ['left', 'right', 'top', 'bottom']
+        return [self.borders[label].color for label in labels]
+
+    @border_color.setter
+    def border_color(self, label_color):
[…]
