+++ .github/workflows/test.yml
@@ -129,7 +129,7 @@ jobs:
       run: |
         bash ci/travis/script.sh
     - name: Check if experimental
-      if: ${{ failure() && matrix.experimental == 'true' }}
+      if: ${{ matrix.experimental == 'true' }}
       run: |
         echo -e "::warning ::Error on this experimental matrix"
     - name: Deploy docs
