### 2021-07-02 #3384 @phamvanvung
> Thanks @phamvanvung - as much as possible, can you edit your comment above to make it clear where you think something needs to be fixed/changed before publication, and where things are suggested improvements that don't block you from checking off review criteria? Given that I think you've checked off all the review criteria, I assume none of your comments are blockers, correct?
>
> And it any items are blockers, it you want to create issues in the source repository, you could (but don't have to).

Hi,
For me, the automated test issue is minor and might due to version differences in my environment.
I now changed the references to be a blocker and I think it should be easy and quick for the authors to fix.
