### 2021-07-02 6dc4c97
feat: added a compute sizes method in GridLayout class
