### 2021-07-02 8d006b1
fix: removed UI size calculation from utils
