### 2021-07-02 #3384 @danielskatz
@phamvanvung - Can you be specific (in the copy of the issue that you will create in https://github.com/fury-gl/fury) about what you think needs to be changed?
