### 2021-07-02 475528d
fix: fixed card tutorial to have URL image
