### 2021-07-02 fury-gl/fury#450 @danielskatz
Issue with references related to JOSS review
This is from @phamvanvung - copied from https://github.com/openjournals/joss-reviews/issues/3445

For the review of this software in https://github.com/openjournals/joss-reviews/issues/3384,  @phamvanvung has an issue regarding the references. He wrote

> References: Is the list of references complete, and is everything cited appropriately that should be cited (e.g., papers, datasets, software)? Do references in the text use the proper citation syntax?

I hope he can describe the issue in more detail.
