### 2021-07-03 fury-gl/fury#450 @skoudoro
Thank you for letting us know! We will update the references on the website and the paper.
