+++ fury/stream/README.md
@@ -3,34 +3,34 @@
 - [x] Create a buffer manager object
   - For using DRY principle when multiple encondings became availbale
 - [ ] Widget object compatibility with python versions lower than 3.8
-  - memory socket communication  
+  - memory socket communication (maybe not) 
 - [x] MJPEG econding
 - [x] Allows the user to set the encoding inside the widget object
 - [x] Fix memory release issues with the shared_manager in python >=3.8 inside the Widget object
 - [x] Create a tutorial about shared_memory and widget object
-- [x] update the examples with multiprocessing to be compatible with MacOs
+- [ ] update the examples with multiprocessing to be compatible with MacOs
   - setting multiprocessing.set_start_method('spawn')
-- [ ] Implement all the interactions: double-click, touch etc
+- [ ] Implement all the interactions: double-click etc
 - [x] Remove useless behavior
   - [x] FuryStreamClient/Interaction reading shared memory blocks
   - [x] Separated interaction server
   - [x] Pass circularqueue objects instead of buffer or buffer_names
 - [ ] Widget object: OFF-ScreenRendering should be optional 
-- [ ] MJPEG and H264 enconding:  allows to decreasse the image resolution of  streaming
+- [ ] MJPEG and H264 enconding:  allows to decreasse the resolution of  the streaming
 - [x] Widget object: test the if the drawn port is avilable
 - [x] Fix MJPEG  on ngrok
-- [ ] MJPEG Async should be separated from get_image
+- [x] MJPEG Async should be separated from get_image
   - Get image can be used inside of webrtc streaming if the user request a "print"
 - [x] Fix cleanup on stream client
-- [ ] Monkey-patch solution to the shared_memory issue should be optional in the circular_queue
+- [x] Monkey-patch solution to the shared_memory issue should be optional in the circular_queue
 - [ ] Tests
   - [x] StreamClient
-  - [ ] StreamInteraction
+  - [x] StreamInteraction
   - [x] IntervalTimer
   - [x] CircularQueue
   - [x] Multidimensional Buffer
   - [x] ImageBufferManager
-  - [ ] Widget
+  - [x] Widget
   - [ ] web_server
   - [ ] aiohttp endpoints
 - [ ] Improves the js/html code
