+++ docs/tutorials/02_ui/viz_card.py
@@ -44,7 +44,7 @@ show_manager = window.ShowManager(size=current_size,
 show_manager.scene.add(card)
 
 # To interact with the UI, set interactive = True
-interactive = 1
+interactive = False
 
 if interactive:
     show_manager.start()
