### 2021-07-04 566bb09
removed repeated functions
