### 2021-07-05 cb6de7c
added actors to render sphere, stick and ball & stick models
