### 2021-07-05 da691cf
improved docstring
