### 2021-07-05 abc0afa
fix: code cleanup, reduced duplication, methods
