+++ fury/ui.py
@@ -5310,6 +5310,9 @@ class Tree2D(UI):
                                     expandable=False, expanded=True,
                                     indent=self.indent, child_indent=self.indent,
                                     child_height=self.node_height)
+        
+        for node in self.nodes_dict.values():
+            node.set_visibility(False)
 
     def _get_actors(self):
         """ Get the actors composing this UI component.
@@ -5350,11 +5353,16 @@ class Tree2D(UI):
             parent_label = list(obj.keys())[0]
             child_labels = list(obj.values())[0]
 
-            parent_node = TreeNode2D(label=parent_label)
+            if parent_label in self.nodes_dict.keys():
+                parent_node = self.nodes_dict[parent_label]
+            else:
+                parent_node = TreeNode2D(label=parent_label)
+
             child_nodes = [TreeNode2D(label=child_label) for child_label in
                            child_labels]
 
-            for child_node in child_nodes:
+            for child_node, child_label in zip(child_nodes, child_labels):
+                self._nodes_dict[child_label] = child_node
                 parent_node.add_node(child_node)
 
             self._nodes.append(parent_node)
@@ -5370,6 +5378,37 @@ class Tree2D(UI):
         """
         self.base_node.resize(size)
     
+    def add_content(self, node, content, coords=(0., 0.)):
+        """Add content to a sepcific node
+
+        Parameters
+        ----------
+        node: str
+            Label of the node
+        content: :class: `UI` or `TreeNode2D`
+            The content that is to be added in the node
+        coords : (float, float) or (int, int), optional
+            If float, normalized coordinates are assumed and they must be
+            between [0,1].
+            If int, pixels coordinates are assumed and it must fit within the
+            content panel's size.
+        """
+        _node = self.select_node(node)
+        _node.add_node(content, coords)
+
+        _node.set_visibility(False)
+
+    def select_node(self, node):
+        """Get the instance of a specific node
+
+        Parameters
+        ----------
+        node: str
+            Label of the node
+        """
+        _node = self.nodes_dict[node]
+        return _node
+
     @property
     def nodes(self):
         """Get all the nodes present in the Tree2D
@@ -5411,7 +5450,8 @@ class TreeNode2D(UI):
     def __init__(self, label, parent=None, children=None,
                  position=(0, 0), size=(200, 200), indent=5,
                  child_indent=10, child_height=25, color=(0.3, 0.3, 0.3),
-                 opacity=0.8, expandable=True, expanded=False):
+                 opacity=0.8, expandable=True, expanded=False,
+                 selected_color=(0.8, 0.3, 0.3)):
         """Initialize the UI element
 
         Parameters
@@ -5455,11 +5495,18 @@ class TreeNode2D(UI):
         self.content_size = size
         self.expandable = expandable
         self.expanded = expanded
+        self.selected = False
+        self.selected_color = selected_color
+        self.unselected_color = color
 
         super(TreeNode2D, self).__init__(position)
         self.resize(size)
         self.title_panel.color = color
         self.title_panel.opacity = opacity
+
+        #  hooks for selecting/de-selecting node
+        self.on_node_select = lambda ui: None
+        self.on_node_deselect = lambda ui: None
 
     def _setup(self):
         """Setup this UI element.
@@ -5482,8 +5529,6 @@ class TreeNode2D(UI):
                                  read_viz_icons(fname="circle-down.png")))
 
         self.button = Button2D(icon_fnames=self.button_icons)
-        self.button.on_left_mouse_button_clicked = self.toggle_view
-
         self.title_panel.add_element(self.label_text, (self.indent, 0))
 
         self.title_panel.add_element(self.button,
@@ -5492,14 +5537,18 @@ class TreeNode2D(UI):
[…]
