### 2021-07-06 3123c4d
Organized references. Added missing DOIs and URLs.
