+++ paper.bib
@@ -1,15 +1,45 @@
-@book{schroeder2004visualization,
-  title={The visualization toolkit: an object-oriented approach to 3D graphics},
-  author={Schroeder, Will J and Lorensen, Bill and Martin, Ken},
-  year={2004},
-  publisher={Kitware}
+@article{abadi2016tensorflow,
+  title={Tensorflow: Large-scale machine learning on heterogeneous distributed systems},
+  author={Abadi, Mart{\'\i}n and Agarwal, Ashish and Barham, Paul and Brevdo, Eugene and Chen, Zhifeng and Citro, Craig and Corrado, Greg S and Davis, Andy and Dean, Jeffrey and Devin, Matthieu and others},
+  journal={arXiv preprint arXiv:1603.04467},
+  year={2016}
 }
 
-@book{sellers2016vulkan,
-  title={Vulkan programming guide: The official guide to learning vulkan},
-  author={Sellers, Graham and Kessenich, John},
-  year={2016},
-  publisher={Addison-Wesley Professional}
+@inproceedings{campagnola2015vispy,
+  title={VisPy: harnessing the GPU for fast, high-level visualization},
+  author={Campagnola, Luke and Klein, Almar and Larson, Eric and Rossant, Cyrille and Rougier, Nicolas P},
+  booktitle={Proceedings of the 14th Python in Science Conference},
+  year={2015},
+  doi={10.25080/majora-7b98e3ed-00e}
+}
+
+@article{coumans2013bullet,
+  author  = {Coumans, Erwin and others},
+  journal = {Open source: bulletphysics. org},
+  title   = {Bullet physics library},
+  year    = {2013},
+  number  = {49},
+  pages   = {5},
+  volume  = {15},
+  url     = {http://www.bulletphysics.org},
+}
+
+@article{garyfallidis2014dipy,
+  title={Dipy, a library for the analysis of diffusion MRI data},
+  author={Garyfallidis, Eleftherios and Brett, Matthew and Amirbekian, Bagrat and Rokem, Ariel and Van Der Walt, Stefan and Descoteaux, Maxime and Nimmo-Smith, Ian},
+  journal={Frontiers in neuroinformatics},
+  volume={8},
+  pages={8},
+  year={2014},
+  publisher={Frontiers},
+  doi={10.3389/fninf.2014.00008}
+}
+
+@techreport{hagberg2008exploring,
+  title={Exploring network structure, dynamics, and function using NetworkX},
+  author={Hagberg, Aric and Swart, Pieter and S Chult, Daniel},
+  year={2008},
+  institution={Los Alamos National Lab.(LANL), Los Alamos, NM (United States)}
 }
 
 @article{harris2020array,
@@ -21,59 +51,14 @@
   pages={357--362},
   year={2020},
   publisher={Nature Publishing Group}
+  doi={10.1038/s41586-020-2649-2}
 }
 
-@article{virtanen2020scipy,
-  title={SciPy 1.0: fundamental algorithms for scientific computing in Python},
-  author={Virtanen, Pauli and Gommers, Ralf and Oliphant, Travis E and Haberland, Matt and Reddy, Tyler and Cournapeau, David and Burovski, Evgeni and Peterson, Pearu and Weckesser, Warren and Bright, Jonathan and others},
-  journal={Nature methods},
-  volume={17},
-  number={3},
-  pages={261--272},
-  year={2020},
-  publisher={Nature Publishing Group}
-}
-
-@inproceedings{mckinney2010data,
-  title={Data structures for statistical computing in python},
-  author={McKinney, Wes and others},
[…]
