+++ paper.bib
@@ -14,14 +14,14 @@
 }
 
 @article{coumans2013bullet,
-  author  = {Coumans, Erwin and others},
-  journal = {Open source: bulletphysics. org},
-  title   = {Bullet physics library},
-  year    = {2013},
-  number  = {49},
-  pages   = {5},
-  volume  = {15},
-  url     = {http://www.bulletphysics.org},
+  author={Coumans, Erwin and others},
+  journal={Open source: bulletphysics.org},
+  title={Bullet physics library},
+  year={2013},
+  number={49},
+  pages={5},
+  volume={15},
+  url={http://www.bulletphysics.org},
 }
 
 @article{garyfallidis2014dipy,
@@ -155,6 +155,7 @@
   pages={261--272},
   year={2020},
   publisher={Nature Publishing Group}
+  doi={10.1038/s41592-019-0686-2}
 }
 
 @article{madhavan2013nanohub,
