### 2021-07-07 036ff6d
Added links and DOIs. Replaced TF and VTK references.
