+++ paper.bib
@@ -1,8 +1,16 @@
-@article{abadi2016tensorflow,
-  title={Tensorflow: Large-scale machine learning on heterogeneous distributed systems},
-  author={Abadi, Mart{\'\i}n and Agarwal, Ashish and Barham, Paul and Brevdo, Eugene and Chen, Zhifeng and Citro, Craig and Corrado, Greg S and Davis, Andy and Dean, Jeffrey and Devin, Matthieu and others},
-  journal={arXiv preprint arXiv:1603.04467},
-  year={2016}
+@inproceedings{abadi2016tensorflow,
+  title={TensorFlow: A System for Large-Scale Machine Learning},
+  author={Abadi, Mart\'{\i}n and Barham, Paul and Chen, Jianmin and Chen, Zhifeng and Davis, Andy and Dean, Jeffrey and Devin, Matthieu and Ghemawat, Sanjay and Irving, Geoffrey and Isard, Michael and Kudlur, Manjunath and Levenberg, Josh and Monga, Rajat and Moore, Sherry and Murray, Derek G. and Steiner, Benoit and Tucker, Paul and Vasudevan, Vijay and Warden, Pete and Wicke, Martin and Yu, Yuan and Zheng, Xiaoqiang},
+  year={2016},
+  isbn={9781931971331},
+  publisher={USENIX Association},
+  address={USA},
+  booktitle={Proceedings of the 12th USENIX Conference on Operating Systems Design and Implementation},
+  pages={265–283},
+  numpages={19},
+  location={Savannah, GA, USA},
+  series={OSDI'16},
+  url={https://www.usenix.org/system/files/conference/osdi16/osdi16-abadi.pdf}
 }
 
 @inproceedings{campagnola2015vispy,
@@ -35,11 +43,14 @@
   doi={10.3389/fninf.2014.00008}
 }
 
-@techreport{hagberg2008exploring,
-  title={Exploring network structure, dynamics, and function using NetworkX},
+@inproceedings{hagberg2008exploring,
+  title={Exploring Network Structure, Dynamics, and Function using NetworkX},
   author={Hagberg, Aric and Swart, Pieter and S Chult, Daniel},
+  booktitle={Proceedings of the 7th Python in Science Conference},
+  pages={11-15},
+  address={Pasadena, CA USA},
   year={2008},
-  institution={Los Alamos National Lab.(LANL), Los Alamos, NM (United States)}
+  url={http://conference.scipy.org/proceedings/SciPy2008/paper_2/full_text.pdf}
 }
 
 @article{harris2020array,
@@ -50,7 +61,7 @@
   number={7825},
   pages={357--362},
   year={2020},
-  publisher={Nature Publishing Group}
+  publisher={Nature Publishing Group},
   doi={10.1038/s41586-020-2649-2}
 }
 
@@ -81,7 +92,7 @@
   number={5},
   pages={17--23},
   year={2008},
-  publisher={IEEE}
+  publisher={IEEE},
   doi={10.1109/MCSE.2008.120}
 }
 
@@ -96,21 +107,27 @@
   doi={10.25080/majora-92bf1922-00a}
 }
 
-@article{paszke2019pytorch,
-  title={Pytorch: An imperative style, high-performance deep learning library},
-  author={Paszke, Adam and Gross, Sam and Massa, Francisco and Lerer, Adam and Bradbury, James and Chanan, Gregory and Killeen, Trevor and Lin, Zeming and Gimelshein, Natalia and Antiga, Luca and others},
-  journal={arXiv preprint arXiv:1912.01703},
-  year={2019}
+@inproceedings{paszke2019pytorch,
+  title={PyTorch: An Imperative Style, High-Performance Deep Learning Library},
[…]
