### 2021-07-07 3a1765a
fix: fixed corner resize bug
