### 2021-07-08 d7e0b9c
feat: added is_overflowing method
