### 2021-07-08 5e7fdac
feat: icon for TreeNode2D
