+++ paper.bib
@@ -165,7 +165,8 @@
   title={Vulkan programming guide: The official guide to learning vulkan},
   author={Sellers, Graham and Kessenich, John},
   year={2016},
-  publisher={Addison-Wesley Professional}
+  publisher={Addison-Wesley Professional},
+  isbn={978-0134464541}
 }
 
 @article{virtanen2020scipy,
