### 2021-07-08 fury-gl/fury#450 @Garyfallidis
Updates were addressed in #453, #454 and #455 which are currently merged.
