### 2021-07-08 #3384 @Garyfallidis
Dear Prof. Rougier (@rougier), the conda packages for vtk are not officially provided by Kitware and therefore we suggest using pip directly to install vtk.

> I had no trouble installing Fury following the instructions but the test failed with a segmentation fault (when testing picking).

This is probably because of the conda installation. Time permitted, please uninstall conda vtk and upgrade vtk using pip.

> The output was rather scarce so it might be difficult to debug. But I managed to run some of the tutorials (I've not tested all of them). Overall it's a nice and much needed package.

Thank you for your encouraging words.

> Each example is a bit slow to start but I suspect this is because of VTK.

We should be able to partially import functions with the latest versions of VTK and hopefully reduce starting time. On my desktop, the starting time for the viz_selection.py example was 7 seconds (i7 CPU @ 3.40GHz and NVIDIA GeForce GTX 1070).

> Concerning the paper, I've some minor comments:
> The paper is rather short which is ok for JOSS but in the meantime it's a bit frustrating for the reader. For example, you briefly compared with mayavi/glumpy/vispy but it's a bit difficult to compare since you do not provide much details. For example, why did you not reuse one of them (lack of support? some ingredients were missing, etc).

Great question. VTK was providing some utilities that were hard to implement and not available in the other packages. Nonetheless, I started a project with direct access to OpenGL back in 2008 and I found it hard to scale on academic resources. However, we should say here that all 3 projects mayavi/glumpy and vispy have been an inspiration for FURY and the decision to go with VTK was a multi-year process moving back and forth through different libraries and standards. In the end we really liked the idea of updating existing shaders that VTK provides and we had quite a bit of code from DIPY (using VTK) that we could reuse in FURY.

> You explain that FURY is interoperable with a lot of software but we don't have details. For example, what does it mean for matplotlib? I think people (including me) would be very interested to have either one example or further explanations.

FURY was built to be fully compatible with the data structures provided by numpy, this makes the library compatible with all the functions in Scipy and as some processing done in Pandas, TensorFlow, and Pytorch can output numpy arrays, then that information can be visualized in FURY too. This decision has also allowed developers to visualize the results of experiments obtained from NetworkX or PyBullet. Furthermore, some functionalities of FURY inherit/draw power from structures/functions in Matplotlib and Pillow. For instance, we have a function matplotlib_figure_to_numpy that allows us to grab a figure from matplotlib and show it in the 3D world with FURY. We would love to have the option to directly draw on a Matplotlib figure but this is still not implemented.

> In the list of authors, it's nice to recognize "FURY Contributors". I just wondered if you could add either a footnote or a dedicated section to list them but this would imply permission and may be too much trouble.

We do provide the following page in our website that gives credit to all contributors.
https://fury.gl/latest/community.html
This page is using GitHub API and updates the current list of contributors in each website update (for every new release).
Let me know if linking to this page is adequate.

> For the DOI, I think some are missing (for example Numpy), could you fix them?

Yes, this is done.
