+++ docs/source/introduction.rst
@@ -3,20 +3,24 @@ About
 =====
 
 Overview
----------
+--------
 
-Free Unified Rendering for Python
+Free Unified Rendering in pYthon
 
+What is FURY?
+-------------
 
-What is FURY ?
----------------
+It is a community-driven, open-source, and high-performance scientific visualization library that harnesses the graphics processing unit (GPU) for improved speed, precise interactivity, and visual clarity.
 
+Statement of Need
+-----------------
 
+FURY was created to address the growing necessity of high-performance 3D scientific visualization in an easy-to-use API fully compatible with the Pythonic ecosystem. To achieve this FURY takes ideas from CGI (Computer-Generated Imagery) and deploys them for usage in everyday research practices.
 
 Mission Statement
-------------------
+-----------------
 
-The purpose of FURY is to make it easier to perform advanced visualizations and animations. We aim to build software that is:
+The purpose of FURY is to make it easier to perform advanced visualizations and animations without compromising performance. We aim to build software that is:
 
 * clearly written
 * clearly explained
@@ -26,30 +30,24 @@ The purpose of FURY is to make it easier to perform advanced visualizations and 
 We hope that, if we fail to do this, you will let us know and we will try and make it better.
 
 Features
----------
+--------
 
-**Efficient**
-
-
-
-**Robust**
-
-
-**Multiplatform**
+- **Efficient**
+- **Robust**
+- **Multiplatform**
 
 
 License
----------
+-------
 
 FURY is distributed under the BSD 3 License
 
 Credits
---------
+-------
 
 Go to :doc:`Community page <community>` to see who have been involved in the development of FURY.
 
 Bug reports and support
----------------------------
+-----------------------
 
 Please report any issues via https://github.com/fury-gl/fury/issues. All types of issues are welcome including bug reports, documentation typos, feature requests and so on.
-
