### 2021-07-08 fury-gl/fury#457 @Garyfallidis
Improve loading speed using partial imports
We can try to make our scripts load faster by importing only the functions we need from VTK. This is supposed to be supported from VTK 9 and on. Let's try it.

The issue was reported during our JOSS review https://github.com/openjournals/joss-reviews/issues/3384 by N.P. Rougier.
