### 2021-07-08 #3384 @Garyfallidis
> I tested the software on: macOS (Big Sur), conda 4.10.2 environment. All the installation instructions work well.

Thank you for your review @phamvanvung and happy to hear that the instructions work well.

> There is an automated test, and I tried executing the test. Many cases showed "PASSED" status, but there are still some errors shown (this might be due to the version compatibility from my environment): E.g.:
fury/tests/test_pick.py::test_picking_manager Fatal Python error: Segmentation fault

Yes, this looks like a conda issue. Please try to uninstall conda vtk and use the pip installer for vtk. Hopefully, this change will fix the issue.

> The statement of need is clear from the Software paper (PDF). However, though I can figure out the statement of needs from the Documentation page, it is unclear for me (as required from the checklist).

Apologies, we are currently transitioning to an updated (new)  website. We added the statement of need in the about section of the website.  Please see merged PR here https://github.com/fury-gl/fury/pull/456

> Given the limitation of the number of pages, the Software paper does a good job stating its aims, architectures, and quick pointers to comparable and commonly used software packages. They also provide the main differences between these similar packages and FURY. Interested readers can refer to these pointers for further comparisons if needed.

Thank you greatly for your positive feedback.

> Regarding the references:

> It was hard for me to find the source for the "Bullet physics library" reference. Please add "Available at: http://www.bulletphysics.org," if possible.
>Here is the DOI for the "Array programming with NumPy" reference: https://doi.org/10.1038/s41586-020-2649-2
>Here is the DOI for the "Advancing Education and Research in Nanotechnology" reference: http://dx.doi.org/10.1109/MCSE.2008.120
> Similarly, the authors can try and search for the DOIs of the missing references.
> All in all, I have positive feelings regarding the usability of this software and highly support its use.

Thank you again @phamvanvung . We corrected the links and DOIs as suggested in FURY PRs (453-455)
https://github.com/fury-gl/fury/pull/453
https://github.com/fury-gl/fury/pull/454
https://github.com/fury-gl/fury/pull/455
