### 2021-07-08 fury-gl/fury#450 @danielskatz
👋 @phamvanvung - I'm going to close this, but feel free to repoen it or add another issue if you feel there's still a problem
