### 2021-07-09 #3384 @rougier
@Garyfallidis Thanks for all the explanations and he nice package. I hope we'll have some opportunities to interact / talk about 3D visualization.
@danielskatz I'm done with the review and all my comments have been addressed. For me the paper can be published.
