### 2021-07-09 2c0435a
rendering optimizations
