+++ fury/stream/client.py
@@ -1,5 +1,7 @@
 import vtk
 import sys
+import logging
+import time
 if sys.version_info.minor >= 8:
     PY_VERSION_8 = True
 else:
@@ -7,15 +9,33 @@ else:
     PY_VERSION_8 = False
 
 import numpy as np
-
+from functools import partial
 from fury.stream.tools import ArrayCircularQueue, SharedMemCircularQueue
 from fury.stream.tools import (
     RawArrayImageBufferManager, SharedMemImageBufferManager)
 from fury.stream.tools import IntervalTimer
 from fury.stream.constants import _CQUEUE
 
-import logging
-import time
+
+def callback_stream_client(*args, **kwargs):
+    stream_client = kwargs['stream_client']
+    if not stream_client._in_request:
+        stream_client._in_request = True
+        stream_client.window2image_filter.Update()
+        stream_client.window2image_filter.Modified()
+        vtk_image = stream_client.window2image_filter.GetOutput()
+        vtk_array = vtk_image.GetPointData().GetScalars()
+        # num_components = vtk_array.GetNumberOfComponents()
+
+        w, h, _ = vtk_image.GetDimensions()
+        np_arr = np.frombuffer(vtk_array, dtype='uint8')
+        if np_arr is not None:
+            stream_client.img_manager.write_into(w, h, np_arr)
+            # time.sleep(kwargs['ms']/1000)
+            stream_client._in_request = False
+            if kwargs['force_render']:
+                stream_client.showm.window.Render()
+                stream_client.showm.iren.Render()
 
 
 class FuryStreamClient:
@@ -89,45 +109,42 @@ class FuryStreamClient:
     def start(self, ms=16,):
         if self._started:
             self.stop()
-
-        def callback(caller, timerevent):
-            if not self._in_request:
-                if not self.update:
-                    return
-                self._in_request = True
-                self.window2image_filter.Update()
-                self.window2image_filter.Modified()
-                vtk_image = self.window2image_filter.GetOutput()
-                vtk_array = vtk_image.GetPointData().GetScalars()
-                # num_components = vtk_array.GetNumberOfComponents()
-
-                w, h, _ = vtk_image.GetDimensions()
-                np_arr = np.frombuffer(vtk_array, dtype='uint8')
-                if np_arr is not None:
-                    self.img_manager.write_into(w, h, np_arr)
-                    self._in_request = False
-
         if ms > 0:
             if self._whithout_iren_start:
                 self._interval_timer = IntervalTimer(
                     ms/1000,
-                    callback,
-                    None,
-                    None)
+                    partial(
+                        callback_stream_client,
+                        **{'stream_client': self, 'force_render': False}
+                    )
+                )
             else:
+                def callback(caller, event, *args, **kwargs):
+                    callback_stream_client(*args, **kwargs)
                 self._id_observer = self.showm.iren.AddObserver(
-                    "TimerEvent", callback)
+                    "TimerEvent",
+                    partial(
+                        callback,
+                        **{'stream_client': self, 'force_render': False}
+                    )
+                )
                 self._id_timer = self.showm.iren.CreateRepeatingTimer(ms)
                 # self.showm.window.AddObserver("TimerEvent", callback)
                 # id_timer = self.showm.window.CreateRepeatingTimer(ms)
 
         else:
-            # id_observer = self.showm.iren.AddObserver(
             self._id_observer = self.showm.iren.AddObserver(
-                'RenderEvent', callback)
-        self.showm.render()
+                'RenderEvent',
+                partial(
+                    callback_stream_client,
+                    **{'stream_client': self, 'force_render': False}
+                )
+            )
+        self.showm.window.Render()
+        self.showm.iren.Render()
[…]
