### 2021-07-09 #3384 @danielskatz
👋 @phamvanvung - seeing that you have checked off all your items, are you also supportive of publishing the work?
