### 2021-07-09 #3384 @Garyfallidis
It would be a pleasure and an honor to talk to you @rougier. Your books, code and work have been an inspiration for myself and many members of the FURY team.
