+++ fury/ui/elements.py
@@ -1,6 +1,6 @@
 """UI components module."""
 
-__all__ = ["Button2D", "TextBox2D", "LineSlider2D", "LineDoubleSlider2D",
+__all__ = ["TextBox2D", "LineSlider2D", "LineDoubleSlider2D",
            "RingSlider2D", "RangeSlider", "Checkbox", "Option", "RadioButton",
            "ComboBox2D", "ListBox2D", "ListBoxItem2D", "FileMenu2D"]
 
@@ -11,233 +11,12 @@ from string import printable
 
 
 import numpy as np
-import vtk
 
 from fury.data import read_viz_icons
-from fury.io import load_image
 from fury.ui.core import UI, Rectangle2D, TextBlock2D, Disk2D
 from fury.ui.containers import Panel2D
 from fury.ui.helpers import TWO_PI, clip_overflow
-from fury.utils import set_input
-
-
-class Button2D(UI):
-    """A 2D overlay button and is of type vtkTexturedActor2D.
-
-    Currently supports::
-
-        - Multiple icons.
-        - Switching between icons.
-
-    """
-
-    def __init__(self, icon_fnames, position=(0, 0), size=(30, 30)):
-        """Init class instance.
-
-        Parameters
-        ----------
-        icon_fnames : List(string, string)
-            ((iconname, filename), (iconname, filename), ....)
-        position : (float, float), optional
-            Absolute coordinates (x, y) of the lower-left corner of the button.
-        size : (int, int), optional
-            Width and height in pixels of the button.
-
-        """
-        super(Button2D, self).__init__(position)
-
-        self.icon_extents = dict()
-        self.icons = self._build_icons(icon_fnames)
-        self.icon_names = [icon[0] for icon in self.icons]
-        self.current_icon_id = 0
-        self.current_icon_name = self.icon_names[self.current_icon_id]
-        self.set_icon(self.icons[self.current_icon_id][1])
-        self.resize(size)
-
-    def _get_size(self):
-        lower_left_corner = self.texture_points.GetPoint(0)
-        upper_right_corner = self.texture_points.GetPoint(2)
-        size = np.array(upper_right_corner) - np.array(lower_left_corner)
-        return abs(size[:2])
-
-    def _build_icons(self, icon_fnames):
-        """Convert file names to vtkImageDataGeometryFilters.
-
-        A pre-processing step to prevent re-read of file names during every
-        state change.
-
-        Parameters
-        ----------
-        icon_fnames : List(string, string)
-            ((iconname, filename), (iconname, filename), ....)
-
-        Returns
-        -------
-        icons : List
-            A list of corresponding vtkImageDataGeometryFilters.
-
-        """
-        icons = []
-        for icon_name, icon_fname in icon_fnames:
-            icons.append((icon_name, load_image(icon_fname, as_vtktype=True)))
-
-        return icons
-
-    def _setup(self):
-        """Set up this UI component.
-
-        Creating the button actor used internally.
-
-        """
-        # This is highly inspired by
-        # https://github.com/Kitware/VTK/blob/c3ec2495b183e3327820e927af7f8f90d34c3474/Interaction/Widgets/vtkBalloonRepresentation.cxx#L47
-
-        self.texture_polydata = vtk.vtkPolyData()
-        self.texture_points = vtk.vtkPoints()
-        self.texture_points.SetNumberOfPoints(4)
-
-        polys = vtk.vtkCellArray()
-        polys.InsertNextCell(4)
-        polys.InsertCellPoint(0)
-        polys.InsertCellPoint(1)
-        polys.InsertCellPoint(2)
-        polys.InsertCellPoint(3)
-        self.texture_polydata.SetPolys(polys)
-
-        tc = vtk.vtkFloatArray()
[…]
