+++ docs/examples/2Dwavefunc_animated.py
@@ -17,7 +17,6 @@ Some other interesting functions you can try plotting -
 # Importing necessary modules
 
 from fury import window, actor, ui, utils, colormap
-from vtk.util import numpy_support
 import numpy as np
 import itertools
 
@@ -27,15 +26,10 @@ import itertools
 # points which are being used to plot the wave.
 # Kindly note that only the z coordinate is being modified with time as only
 # the z coordinate is a function of time.
-# When 'animating' is set to False, it returns colors and vertices that can be
-# used to generate a wave.
-# When 'animating' is set to True, it returns the updated colors and vertices
-# for the actor being used to generate the wave. Thus, 'animating' is set to
-# True when updating the coordinates inside the timer_callback function.
 
 def update_coordinates(lower_xbound=-1, upper_xbound=1, lower_ybound=-1,
                        upper_ybound=1, npoints=100, time=0, incre=0,
-                       cmap_name='viridis', animating=True):
+                       cmap_name='viridis'):
 
     # creating the points which will be used to generate the surface
     x = np.linspace(lower_xbound, upper_xbound, npoints)
@@ -48,20 +42,15 @@ def update_coordinates(lower_xbound=-1, upper_xbound=1, lower_ybound=-1,
     rate = 1.5
 
     # Z is the function F i.e. F(x, y, t)
-    z = 0.24*np.sin(4*np.pi*x)*np.cos(4*np.pi*y-rate*time)
+    z = 0.48*np.sin(2*np.pi*x)*np.cos(2*np.pi*y-rate*time)
     xyz = np.vstack([x, y, z]).T
 
     # creating the colormap
     v = np.copy(z)
     v /= np.max(np.abs(v), axis=0)
     colors = colormap.create_colormap(v, name=cmap_name)
-    if animating:
-        temp_surface = actor.surface(xyz, colors=colors)
-        return xyz, numpy_support.vtk_to_numpy(temp_surface.GetMapper().
-                                               GetInput().GetPointData().
-                                               GetArray(0))
-    else:
-        return xyz, colors
+
+    return xyz, colors
 
 
 ###############################################################################
@@ -74,31 +63,31 @@ def update_coordinates(lower_xbound=-1, upper_xbound=1, lower_ybound=-1,
 #     of timer_callback function; (default = 0.1)
 # lower_xbound: float
 #               lower bound of the x values in which the function is plotted
-#               (default = -0.5)
+#               (default = -1)
 # upper_xbound: float
 #               Upper bound of the x values in which the function is plotted
-#               (default = 0.5)
+#               (default = 1)
 # lower_ybound: float
 #               lower bound of the y values in which the function is plotted
-#               (default = -0.5)
+#               (default = -1)
 # upper_ybound: float
 #               Upper bound of the y values in which the function is plotted
-#               (default = 0.5)
+#               (default = 1)
 # npoints: int
 #          For high quality rendering, keep the number high but kindly note
 #          that higher values for npoints slows down the animation
-#          (default = 50)
+#          (default = 200)
 # cmap_name: string
 #            name of the colormap being used to color the wave
 #            (default = 'inferno')
 
 time = 0
 dt = 0.1
-lower_xbound = -0.5
-upper_xbound = 0.5
-lower_ybound = -0.5
-upper_ybound = 0.5
-npoints = 50
+lower_xbound = -1
+upper_xbound = 1
+lower_ybound = -1
+upper_ybound = 1
+npoints = 200
 cmap_name = 'inferno'
 
 
@@ -107,7 +96,7 @@ cmap_name = 'inferno'
 # colors refers to the colormap that'll be used to color the wave
 xyz, colors = update_coordinates(lower_xbound, upper_xbound, lower_ybound,
[…]
