### 2021-07-10 885821b
global x, y
