+++ docs/examples/2Dwavefunc_animated.py
@@ -177,7 +177,7 @@ def timer_callback(_obj, _event):
 
 showm.add_timer_callback(True, 30, timer_callback)
 
-interactive = True
+interactive = False
 if interactive:
     showm.start()
 
