### 2021-07-10 0466df0
added example for the molecular module
