### 2021-07-10 5a172b1
changed the style to make it more pythonic
