+++ fury/molecular.py
@@ -1,194 +1,239 @@
+from fury import molecular
 import vtk
 from vtk.util import numpy_support
-
 
 class Molecule(vtk.vtkMolecule):
     """Your molecule class.
 
     An object that is used to create molecules and store molecular data (e.g.
     coordinate and bonding data).
-    This is a more pythonic version of ``vtkMolecule`` providing simple methods
-    for creating molecules. It provides access to essential functionality
-    available in ``vtkMolecule``.
+    This is a more pythonic version of ``vtkMolecule``.
     """
 
-    def addAtom(self, atomic_num, x_coord, y_coord, z_coord):
-        """Add atomic data to our molecule.
+def add_atom(molecule, atomic_num, x_coord, y_coord, z_coord):
+    """Add atomic data to our molecule.
 
-        Parameters
-        ----------
-        atomic_num : int
-            Atomic number of the atom.
-        x_coord : float
-            x-coordinate of the atom.
-        y_coord : float
-            y-coordinate of the atom.
-        z_coord : float
-            z-coordinate of the atom.
-        """
-        self.AppendAtom(atomic_num, x_coord, y_coord, z_coord)
+    Parameters
+    ----------
+    molecule : Molecule() object
+        The molecule to which the atom is to be added.
+    atomic_num : int
+        Atomic number of the atom.
+    x_coord : float
+        x-coordinate of the atom.
+    y_coord : float
+        y-coordinate of the atom.
+    z_coord : float
+        z-coordinate of the atom.
+    """
+    molecule.AppendAtom(atomic_num, x_coord, y_coord, z_coord)
 
-    def addBond(self, atom1_index, atom2_index, bond_type=1):
-        """Add bonding data to our molecule. Establish a bond of type bond_type
-        between the atom at atom1_index and the atom at atom2_index.
+def add_bond(molecule, atom1_index, atom2_index, bond_type=1):
+    """Add bonding data to our molecule. Establish a bond of type bond_type
+    between the atom at atom1_index and the atom at atom2_index.
 
-        Parameters
-        ----------
-        atom1_index : int
-            Index of the first atom.
-        atom2_index : int
-            Index of the second atom.
-        bond_type : int (optional)
-            Type of bond (single/double/triple). Default: 1
-        """
-        self.AppendBond(atom1_index, atom2_index, bond_type)
+    Parameters
+    ----------
+    molecule : Molecule() object
+        The molecule to which the bond is to be added.
+    atom1_index : int
+        Index of the first atom.
+    atom2_index : int
+        Index of the second atom.
+    bond_type : int (optional)
+        Type of bond (single/double/triple). Default: 1
+    """
+    molecule.AppendBond(atom1_index, atom2_index, bond_type)
 
-    def getTotalNumAtoms(self):
-        """Returns the total number of atoms in the molecule.
-        """
-        return self.GetNumberOfAtoms()
+def get_total_num_atoms(molecule):
+    """Returns the total number of atoms in a given molecule.
 
-    def getTotalNumBonds(self):
-        """Returns the total number of bonds in the molecule.
-        """
-        return self.GetNumberOfBonds()
+    Parameters
+    ----------
+    molecule : Molecule() object
+    """
+    return molecule.GetNumberOfAtoms()
 
-    def getAtomicNumber(self, atom_index):
-        """Get the atomic number of an atom for a specified index.
+def get_total_num_bonds(molecule):
+    """Returns the total number of bonds in a given molecule.
 
-        Returns the atomic number of the atom present at index atom_index.
+    Parameters
+    ----------
+    molecule : Molecule() object
+    """
+    return molecule.GetNumberOfBonds()
 
-        Parameters
-        ----------
-        atom_index : int
-            Index of the atom whose atomic number is to be obtained.
-        """
[…]
