+++ fury/molecular.py
@@ -1,6 +1,6 @@
-from fury import molecular
 import vtk
 from vtk.util import numpy_support
+
 
 class Molecule(vtk.vtkMolecule):
     """Your molecule class.
@@ -9,6 +9,7 @@ class Molecule(vtk.vtkMolecule):
     coordinate and bonding data).
     This is a more pythonic version of ``vtkMolecule``.
     """
+
 
 def add_atom(molecule, atomic_num, x_coord, y_coord, z_coord):
     """Add atomic data to our molecule.
@@ -28,6 +29,7 @@ def add_atom(molecule, atomic_num, x_coord, y_coord, z_coord):
     """
     molecule.AppendAtom(atomic_num, x_coord, y_coord, z_coord)
 
+
 def add_bond(molecule, atom1_index, atom2_index, bond_type=1):
     """Add bonding data to our molecule. Establish a bond of type bond_type
     between the atom at atom1_index and the atom at atom2_index.
@@ -45,6 +47,7 @@ def add_bond(molecule, atom1_index, atom2_index, bond_type=1):
     """
     molecule.AppendBond(atom1_index, atom2_index, bond_type)
 
+
 def get_total_num_atoms(molecule):
     """Returns the total number of atoms in a given molecule.
 
@@ -54,6 +57,7 @@ def get_total_num_atoms(molecule):
     """
     return molecule.GetNumberOfAtoms()
 
+
 def get_total_num_bonds(molecule):
     """Returns the total number of bonds in a given molecule.
 
@@ -62,6 +66,7 @@ def get_total_num_bonds(molecule):
     molecule : Molecule() object
     """
     return molecule.GetNumberOfBonds()
+
 
 def get_atomic_number(molecule, atom_index):
     """Get the atomic number of an atom for a specified index.
@@ -76,6 +81,7 @@ def get_atomic_number(molecule, atom_index):
         Index of the atom whose atomic number is to be obtained.
     """
     return molecule.GetAtomAtomicNumber(atom_index)
+
 
 def set_atomic_number(molecule, atom_index, atomic_num):
     """Set the atomic number of an atom for a specified index.
@@ -94,6 +100,7 @@ def set_atomic_number(molecule, atom_index, atomic_num):
     """
     molecule.SetAtomAtomicNumber(atom_index, atomic_num)
 
+
 def get_atomic_position(molecule, atom_index):
     """Get the atomic coordinates of an atom for a specified index.
 
@@ -107,6 +114,7 @@ def get_atomic_position(molecule, atom_index):
         Index of the atom whose atomic coordinates are to be obtained.
     """
     return molecule.GetAtomPosition(atom_index)
+
 
 def set_atomic_position(molecule, atom_index, x_coord, y_coord, z_coord):
     """Set the atomic coordinates of an atom for a specified index.
@@ -129,6 +137,7 @@ def set_atomic_position(molecule, atom_index, x_coord, y_coord, z_coord):
     """
     molecule.SetAtomPosition(atom_index, x_coord, y_coord, z_coord)
 
+
 def get_bond_type(molecule, bond_index):
     """Get the type of bond for a specified index.
 
@@ -143,6 +152,7 @@ def get_bond_type(molecule, bond_index):
         Index of the bond whose type is to be obtained.
     """
     return molecule.GetBondOrder(bond_index)
+
 
 def set_bond_type(molecule, bond_index, bond_type):
     """Set the bond type of a bond for a specified index.
@@ -161,6 +171,7 @@ def set_bond_type(molecule, bond_index, bond_type):
     """
     return molecule.SetBondOrder(bond_index, bond_type)
 
+
 def get_atomic_number_array(molecule):
     """Returns an array of atomic numbers corresponding to the atoms
[…]
