### 2021-07-10 #3384 @phamvanvung
> 👋 @phamvanvung - seeing that you have checked off all your items, are you also supportive of publishing the work?

Sorry for the late reply, yes, as from my review and the checklist. I recommend publishing this work.
