### 2021-07-11 08bd02e
fix: check for UI by using add_to_scene attr
