+++ fury/layout.py
@@ -2,7 +2,6 @@ import numpy as np
 import math
 from fury.utils import get_bounding_box_sizes
 from fury.utils import get_grid_cells_position
-from fury.ui.core import UI
 
 
 class Layout(object):
@@ -14,7 +13,7 @@ class Layout(object):
 
         for a, pos in zip(actors, positions):
             
-            if isinstance(a, UI):
+            if hasattr(a, 'add_to_scene'):
                 a.position = (pos[0], pos[1])
             else:
                 anchor = np.array(getattr(a, 'anchor', (0, 0, 0)))
@@ -55,7 +54,8 @@ class GridLayout(Layout):
         dim : tuple of int (optional)
             Dimension (nb_rows, nb_cols) of the grid. If provided,
             `aspect_ratio` will be ignored.
-
+        position_offset: tuple (optional)
+            Offset the grid by some factor
         """
         self.cell_shape = cell_shape
         self.aspect_ratio = aspect_ratio
@@ -84,12 +84,12 @@ class GridLayout(Layout):
 
         if self.cell_shape == "rect":
             bounding_box_sizes = np.asarray(
-                    list(map(self.compute_sizes, actors)))
+                    list(map(get_bounding_box_sizes, actors)))
             cell_shape = np.max(bounding_box_sizes, axis=0)[:2]
             shapes = [cell_shape] * len(actors)
         elif self.cell_shape == "square":
             bounding_box_sizes = np.asarray(
-                    list(map(self.compute_sizes, actors)))
+                    list(map(get_bounding_box_sizes, actors)))
             cell_shape = np.max(bounding_box_sizes, axis=0)[:2]
             shapes = [(max(cell_shape),)*2] * len(actors)
         elif self.cell_shape == "diagonal":
@@ -97,7 +97,7 @@ class GridLayout(Layout):
             # of the largest bounding box.
             diagonals = []
             for a in actors:
-                if isinstance(a, UI):
+                if hasattr(a, "add_to_scene"):
                     width, height = a.size
                     diagonal = math.sqrt(width**2 + height**2)
                     diagonals.append(diagonal)
@@ -138,24 +138,3 @@ class GridLayout(Layout):
 
         positions += self.position_offset
         return positions
-    
-    def compute_sizes(self, actor):
-        """Compute the size of the bounding boxes of actors/UI
-
-        Parameters
-        ----------
-        actor: `vtkProp3D` or `UI` element
-            Actor/UI element to whose size is to be calculated
-        
-        Returns
-        -------
-        tuple
-            The bounding box size of actor/UI element
-        """
-
-        if isinstance(actor, UI):
-            width, height = actor.size
-            return (width, height, 0)
-        
-        size = get_bounding_box_sizes(actor)
-        return size
