### 2021-07-11 49a20cc
feat: compute bounding box size of UI elements
