### 2021-07-11 3e31720
tests: test layout with UI elements
