+++ fury/tests/test_layout.py
@@ -2,6 +2,7 @@ import numpy as np
 import numpy.testing as npt
 from fury import actor
 from fury.layout import GridLayout, Layout
+from fury.ui.containers import Panel2D
 
 
 def get_default_cubes(centers=np.asarray([[[0, 0, 0]], [[5, 5, 5]]]),
@@ -34,12 +35,34 @@ def get_default_cubes(centers=np.asarray([[[0, 0, 0]], [[5, 5, 5]]]),
     return (cube_first, cube_second)
 
 
+def get_default_panels(sizes=[(100, 100), (200, 200)],
+                       colors=np.random.rand(2, 3)):
+    """Provides Panels with default parameters
+
+    Parameters
+    ----------
+    sizes: list(tuple, tuple)
+        Sizes of the two panels
+    colors: ndarray ndarray (2,3) or (2, 4)
+        RGB or RGBA (for opacity)
+    """
+    panel_first_size, panel_second_size = sizes
+    panel_first_color, panel_second_color = colors
+
+    panel_first = Panel2D(size=panel_first_size, color=panel_first_color)
+    panel_second = Panel2D(size=panel_second_size, color=panel_second_color)
+
+    return (panel_first, panel_second)
+
+
 def test_layout_apply():
 
     cube_first, cube_second = get_default_cubes()
+    panel_first, panel_second = get_default_panels()
 
     layout = Layout()
     layout.apply([cube_first, cube_second])
+    layout.apply([panel_first, panel_second])
 
     cube_first_center = cube_first.GetCenter()
     cube_second_center = cube_second.GetCenter()
@@ -47,20 +70,28 @@ def test_layout_apply():
     npt.assert_array_equal(cube_first_center, [0, 0, 0])
     npt.assert_array_equal(cube_second_center, [5, 5, 5])
 
+    npt.assert_array_equal([panel_first.position, panel_second.position],
+                           [(0, 0), ]*2)
+
 
 def test_layout_compute_postions():
 
     cube_first, cube_second = get_default_cubes()
+    panel_first, panel_second = get_default_panels()
 
     layout = Layout()
 
-    positions = layout.compute_positions([cube_first, cube_second])
-    npt.assert_array_equal(positions, [])
+    cube_positions = layout.compute_positions([cube_first, cube_second])
+    panel_positions = layout.compute_positions([panel_first, panel_second])
+
+    npt.assert_array_equal(cube_positions, [])
+    npt.assert_array_equal(panel_positions, [])
 
 
 def test_grid_layout_get_cell_shape():
 
     cube_first, cube_second = get_default_cubes()
+    panel_first, panel_second = get_default_panels()
 
     grid = GridLayout()
     grid_square = GridLayout(cell_shape="square")
@@ -70,18 +101,32 @@ def test_grid_layout_get_cell_shape():
     shape = grid.get_cells_shape([cube_first, cube_second])
     shape_square = grid_square.get_cells_shape([cube_first, cube_second])
     shape_diagonal = grid_diagonal.get_cells_shape([cube_first, cube_second])
+
+    panel_shape = grid.get_cells_shape([panel_first, panel_second])
+    panel_shape_sqaure = grid_square.get_cells_shape([panel_first,
+                                                      panel_second])
+
+    panel_shape_diagonal = grid_diagonal.get_cells_shape([panel_first,
+                                                          panel_second])
+
     with npt.assert_raises(ValueError):
-        shape_invalid = invalid_gird.get_cells_shape([cube_first, cube_second])
+        _ = invalid_gird.get_cells_shape([cube_first, cube_second])
 
     npt.assert_array_equal(shape, [[1.5, 1.5], [1.5, 1.5]])
     npt.assert_array_equal(shape_square, [[1.5, 1.5], [1.5, 1.5]], 0)
[…]
