### 2021-07-12 17d845a
feat: added a compute_sizes method in GridLayout
