### 2021-07-12 bd507bd
fix: removed UI bounding box size calculation from utils
