### 2021-07-13 7d5b975
test: remove opengl_state and gl_get_current_state
