### 2021-07-13 0578e3a
feat: TreeNode2D parent auto resizing
