### 2021-07-13 ca37ee8
Added properties tests.
