+++ fury/ui/tests/test_core.py
@@ -6,6 +6,7 @@ import warnings
 from fury.data import DATA_DIR, read_viz_icons, fetch_viz_icons
 from fury import window, ui
 from fury.testing import EventCounter
+import numpy as np
 
 
 def test_ui_button_panel(recording=False):
@@ -66,6 +67,21 @@ def test_ui_button_panel(recording=False):
     npt.assert_raises(ValueError, panel.add_element, another_rectangle_test,
                       (-0.5, 0.5))
 
+    npt.assert_equal(panel.border_width, [0.0, ]*4)
+    npt.assert_equal(panel.border_color, [np.asarray([1, 1, 1]), ]*4)
+
+    panel.border_width = ['bottom', 10.0]
+    npt.assert_equal(panel.border_width[3], 10.0)
+
+    with npt.assert_raises(ValueError):
+        panel.border_width = ['invalid_label', 10.0]
+
+    panel.border_color = ['bottom', (0.4, 0.5, 0.6)]
+    npt.assert_equal(panel.border_color[3], (0.4, 0.5, 0.6))
+
+    new_size = (400, 400)
+    panel.resize(new_size)
+    npt.assert_equal(panel.borders['bottom'].width, 400.0)
     # Assign the counter callback to every possible event.
     event_counter = EventCounter()
     event_counter.monitor(button_test)
