+++ fury/utils.py
@@ -752,7 +752,7 @@ def numpy_to_vtk_matrix(array):
 
 
 def get_bounding_box_sizes(actor):
-    """Get the bounding box sizes of an actor"""
+    """Get the bounding box sizes of an actor."""
     X1, X2, Y1, Y2, Z1, Z2 = actor.GetBounds()
     return (X2-X1, Y2-Y1, Z2-Z1)
 
