### 2021-07-13 #3384 @chrishavlin
Apologies for the slow reply! I just updated my checklist and I'm happy to recommend publishing the work. A very useful contribution, I look forward to using FURY myself in future work.
