### 2021-07-14 #3384 @danielskatz
@Garyfallidis - the next step is for me to proofread the paper, but I'm actually on vacation this week, so I will get it done by early next week at the latest.
