### 2021-07-16 #3384 @danielskatz
@Garyfallidis - in the paper, I don't understand the sentence in the Rendering Engine paragraph that starts with "Furthermore,"  Is this talking about your software? Why "only"?
