### 2021-07-16 #3384 @danielskatz
I think "Mayavi however FURY" should be "Mayavi. FURY, however,"
