### 2021-07-16 #3384 @danielskatz
In the references, you probably need to protect some casing with {}s - the Harris reference should probably say "CUDA" and "SIGGARAPH"   There may be others like this, though I don't see any immediately.
