### 2021-07-20 #3384 @Garyfallidis
Dear @danielskatz your changes have been addressed in this merged pull request https://github.com/fury-gl/fury/pull/464
