### 2021-07-20 #3384 @danielskatz
I think two more small changes are needed - see https://github.com/fury-gl/fury/pull/466
