### 2021-07-20 #3384 @Garyfallidis
Done and merged. Apologies we missed that.
