### 2021-07-21 #3384 @Garyfallidis
Sounds good. Working on the different steps.
