### 2021-08-04 #3384 @Garyfallidis
Hello @danielskatz, the release is tagged as shown here https://github.com/fury-gl/fury/tree/v0.7.1
The version of the release is 0.7.1
Here is the Zenodo link with the DOI https://zenodo.org/record/5160945
The metadata were updated as suggested with title, author names and ORCIDS.
The DOI is ``10.5281/zenodo.5160945``.
Let us know if you need anything else.
